% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxwhisker.R
\name{addBoxWhisker}
\alias{addBoxWhisker}
\title{addBoxWhisker}
\usage{
addBoxWhisker(data, metaData, dataMapping, plotConfiguration, plotObject)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{list of information on \code{data}}

\item{dataMapping}{\code{BoxWhiskerDataMapping} object}

\item{plotConfiguration}{\code{BoxWhiskerPlotConfiguration} object}

\item{plotObject}{a \code{ggplot} object}
}
\value{
A \code{ggplot} object
}
\description{
Add layer of boxes and whiskers  to a \code{ggplot} object
}
