% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom-plots.R
\name{addErrorbar}
\alias{addErrorbar}
\title{addErrorbar}
\usage{
addErrorbar(
  data = NULL,
  metaData = NULL,
  x = NULL,
  ymin = NULL,
  ymax = NULL,
  caption = NULL,
  color = NULL,
  size = NULL,
  linetype = NULL,
  includeCap = FALSE,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{data.frame containing the errorbar endpoints to be plotted}

\item{metaData}{list of information on \code{data} such as \code{dimension} and \code{unit} of their variables}

\item{x}{Mapping for x values.
If \code{data} is NULL or not input, \code{x} numeric values will be used as is for the plot.}

\item{ymin}{Mapping for ymin values.
If \code{data} is NULL or not input, \code{ymin} numeric values will be used as is for the plot.}

\item{ymax}{Mapping for ymin values.
If \code{data} is NULL or not input, \code{ymax} numeric values will be used as is for the plot.}

\item{caption}{vector of character strings defining the legend captions.
This parameter is optional: default value is NULL.}

\item{color}{vector of character strings defining the color of the scatter points.
This parameter is optional: default value \code{NULL} will choose colors according to the current theme.}

\item{size}{vector of numerical defining the sizes of the scatter points.
This parameter is optional: default value \code{NULL} will choose sizes according to the current theme.}

\item{linetype}{vector of character strings defining the linetype linking the scatter points.
This parameter is optional: default value \code{NULL} won't provide lines.}

\item{includeCap}{logical setting if error bars include caps at their ends.}

\item{dataMapping}{\code{RangeDataMapping} class or subclass
mapping x, ymin, ymax and aesthetic variables to the variable names of \code{data}.}

\item{plotConfiguration}{\code{PlotConfiguration} object defining the label and background properties of the plot.}

\item{plotObject}{\code{ggplot} graphical object to which the line layer is added
This parameter is optional: the \code{tlf} library will initialize an empty plot if the parameter is NULL or not provided}
}
\value{
A \code{ggplot} graphical object
}
\description{
Add a errorbar plot layer to a \code{ggplot} graphical object.
Use optional argument \code{caption} to set legend caption.
Since \code{ggplot} manage aesthetic properties across all layers,
aesthetic properties defined in \code{plotConfiguration} will apply across all layers.
If caption is the same as a previous scatter plot layer, the legend will merge their caption and aesthetic properties
}
