% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom-plots.R
\name{addLine}
\alias{addLine}
\title{addLine}
\usage{
addLine(
  data = NULL,
  metaData = NULL,
  x = NULL,
  y = NULL,
  caption = NULL,
  color = NULL,
  shape = NULL,
  size = NULL,
  linetype = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{data.frame containing the line endpoints to be plotted}

\item{metaData}{list of information on \code{data} such as \code{dimension} and \code{unit} of their variables}

\item{x}{Mapping for x values.
If \code{data} is NULL or not input, \code{x} numeric values will be used as is for the plot.
In case there is no \code{y} mapping, \code{x} are the x-values of vertical lines.}

\item{y}{Mapping for y values.
If \code{data} is NULL or not input, \code{y} numeric values will be used as is for the plot.
In case there is no \code{x} mapping, \code{y} are the y-values of horizontal lines.}

\item{caption}{vector of character strings defining the legend captions.
This parameter is optional: default value is NULL.}

\item{color}{vector of character strings defining the color of the scatter points.
This parameter is optional: default value \code{NULL} will choose colors according to the current theme.}

\item{shape}{vector of character strings or numerical defining the shapes of the scatter points.
This parameter is optional: default value \code{NULL} will choose shapes according to the current theme.}

\item{size}{vector of numerical defining the sizes of the scatter points.
This parameter is optional: default value \code{NULL} will choose sizes according to the current theme.}

\item{linetype}{vector of character strings defining the linetype linking the scatter points.
This parameter is optional: default value \code{NULL} won't provide lines.}

\item{dataMapping}{\code{XYGDataMapping} class or subclass
mapping x, y and aesthetic variables to the variable names of \code{data}.}

\item{plotConfiguration}{\code{PlotConfiguration} object defining the label and background properties of the plot.}

\item{plotObject}{\code{ggplot} graphical object to which the line layer is added
This parameter is optional: the \code{tlf} library will initialize an empty plot if the parameter is NULL or not provided}
}
\value{
A \code{ggplot} graphical object
}
\description{
Add a line layer to a \code{ggplot} graphical object.
Use optional argument \code{caption} to set legend caption.
Since \code{ggplot} manage aesthetic properties across all layers,
aesthetic properties defined in \code{plotConfiguration} will apply across all layers.
}
\examples{
\dontrun{
# Add vertical line at x = 2
addLine(x = 2)

# Add vertical line at x = 2 to a previous plot
p <- initializePlot() +
  ggplot2::geom_point(
    data = data.frame(x = c(1, 2, 3), y = c(5, 6, 4)),
    mapping = ggplot2::aes_string(x = "x", y = "y")
  )
p <- addLine(x = 2, plotObject = p)

# Add horizontal line at y = 5 to a previous plot
p <- addLine(y = 5, plotObject = p)
}
# Add a custom line
time <- seq(0, 30, 0.01)
customLineData <- data.frame(x = time, y = cos(time))

p <- addLine(
  data = customLineData,
  dataMapping = XYGDataMapping$new(x = "x", y = "y")
)

# Or for simple cases a smart mapping will get directly x and y from data
p <- addLine(data = customLineData)

# Add a line with caption
p <- addLine(data = customLineData, caption = "My line")
}
