% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-obs-vs-pred.R
\name{addObsVsPredLines}
\alias{addObsVsPredLines}
\title{addObsVsPredLines}
\usage{
addObsVsPredLines(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{data.frame containing the data to be used for the plot}

\item{metaData}{list of lists}

\item{dataMapping}{\code{ObsVsPredDataMapping} object}

\item{plotConfiguration}{\code{ObsVsPredConfiguration} object}

\item{plotObject}{\code{ggplot} graphical object}
}
\value{
A \code{ggplot} graphical object
}
\description{
Add layers of identity line and smoother
}
