% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxwhisker.R
\name{addOutliers}
\alias{addOutliers}
\title{addOutliers}
\usage{
addOutliers(data, metaData, dataMapping, plotConfiguration, plotObject)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{list of information on \code{data}}

\item{dataMapping}{\code{BoxWhiskerDataMapping} object}

\item{plotConfiguration}{\code{BoxWhiskerPlotConfiguration} object}

\item{plotObject}{a \code{ggplot} object}
}
\value{
A \code{ggplot} object
}
\description{
Add a outlier points layer to a \code{ggplot} object
}
