% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom-plots.R
\name{addScatter}
\alias{addScatter}
\title{addScatter}
\usage{
addScatter(
  data = NULL,
  metaData = NULL,
  x = NULL,
  y = NULL,
  caption = NULL,
  color = NULL,
  shape = NULL,
  size = NULL,
  linetype = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{x}{Numeric values to plot along the \code{x} axis. Only used instead of \code{data} if \code{data} is \code{NULL}.}

\item{y}{Numeric values to plot along the \code{y} axis. Only used instead of \code{data} if \code{data} is \code{NULL}.}

\item{caption}{Optional character values defining the legend captions of the plot.}

\item{color}{Optional character values defining the colors of the plot layer.
See \code{grDevices::colors()} to get names of colors}

\item{shape}{Optional character values defining the shapes/symbols of the plot layer.
See enum \code{Shapes} to get names of shapes.}

\item{size}{Optional numeric values defining the size of the plot layer.}

\item{linetype}{Optional character values defining the linetype of the plot layer.
See enum \code{Linetypes} to get names of linetype.}

\item{dataMapping}{A \code{XYGDataMapping} object mapping \code{x}, \code{y} and aesthetic groups to their variable names of \code{data}.}

\item{plotConfiguration}{An optional \code{PlotConfiguration} object defining labels, grid, background and watermark.}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layer}
}
\value{
A \code{ggplot} object
}
\description{
Add a scatter plot layer to a \code{ggplot} object
}
\examples{
# Add scatter using x and y
addScatter(x = c(1, 2, 1, 2, 3), y = c(5, 0, 2, 3, 4))

# Add scatter using a data.frame
time <- seq(0, 30, 0.1)
scatterData <- data.frame(x = time, y = cos(time))

addScatter(
  data = scatterData,
  dataMapping = XYGDataMapping$new(x = "x", y = "y")
)

# Or for simple cases a smart mapping will get directly x and y from data
addScatter(data = scatterData)

# Add a scatter with caption
addScatter(data = scatterData, caption = "My scatter plot")

# Add a scatter with specific properties
addScatter(
data = scatterData, 
color = "blue", shape = "diamond", size = 2, caption = "My data"
)

# Add a scatter with specific properties
p <- addScatter(
data = scatterData, 
color = "blue", shape = "diamond", size = 2, caption = "My data"
)
addScatter(
x = c(0, 1), y = c(1, 0), 
color = "red", shape = "circle", size = 3, 
plotObject = p
)

}
\references{
For examples, see:
\url{https://www.open-systems-pharmacology.org/TLF-Library/articles/atom-plots.html}
}
\seealso{
Other atom plots: 
\code{\link{addErrorbar}()},
\code{\link{addLine}()},
\code{\link{addRibbon}()},
\code{\link{initializePlot}()}
}
\concept{atom plots}
