% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom-plots.R
\name{addScatterPlot}
\alias{addScatterPlot}
\title{addScatterPlot}
\usage{
addScatterPlot(data, metaData = NULL, dataMapping = NULL,
  plotConfiguration = NULL, plotObject = NULL)
}
\arguments{
\item{data}{data.frame containing the data to be used for the plot}

\item{metaData}{list of lists
containing complementary information to data (e.g. their unit and dimension).
This parameter is optional.}

\item{dataMapping}{\code{XYGDataMapping} class or subclass mapping x and y variables to \code{data} variable names
This parameter is optional: the \code{tlf} library provides a smart mapping if only \code{data} is provided}

\item{plotConfiguration}{\code{PlotConfiguration} class or subclass defining labels, grid, background and watermark
This parameter is optional: the \code{tlf} library provides a default configuration according to the current theme}

\item{plotObject}{\code{ggplot} graphical object to which the scatter plot layer is added
This parameter is optional: the \code{tlf} library will initialize an empty plot if the parameter is NULL or not provided}
}
\value{
A \code{ggplot} graphical object
}
\description{
Add a scatter plot layer to a \code{ggplot} graphical object
}
\examples{
data <- data.frame(x = sample(seq(0, 1, 0.001), size = 100), 
y = sample(seq(0, 10, 0.001), size = 100))
addScatterPlot(data)
}
