% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-background.R
\name{addWatermark}
\alias{addWatermark}
\title{addWatermark}
\usage{
addWatermark(
  plotObject,
  watermark,
  color = NULL,
  size = NULL,
  angle = NULL,
  alpha = NULL
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{watermark}{A character value or a \code{Label} object}

\item{color}{Color of the watermark.}

\item{size}{Size of the watermark.}

\item{angle}{Angle of the watermark (in degree).}

\item{alpha}{Numeric value between 0 and 1 corresponding to transparency of the watermark
The closer to 0, the more transparent the watermark is.
The closer to 1, the more opaque the watermark is.}
}
\value{
A \code{ggplot} object
}
\description{
Creates a \code{ggplot} grob based on the label text and its font properties.
Then,  adds the grob to the \code{ggplot} object as a new layer using \code{ggplot2::annotation_custom}.
}
\examples{
# Add a watermark to an empty plot
p <- ggplot2::ggplot()
addWatermark(p, "watermark")

# Watermark with font properties
watermarkLabel <- Label$new(text = "watermark", color = "blue")
addWatermark(p, watermarkLabel)

# Horizontal watermark
addWatermark(p, watermarkLabel, angle = 0)

# Watermark totally opaque
addWatermark(p, watermarkLabel, alpha = 1)

# As multiple layers of watermark:
p2 <- addWatermark(p, watermarkLabel, alpha = 1)
addWatermark(p2, "other watermark", color = "red", angle = 90)

}
