% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\docType{data}
\name{AestheticProperties}
\alias{AestheticProperties}
\title{AestheticProperties}
\format{
An object of class \code{list} of length 6.
}
\usage{
AestheticProperties
}
\description{
Enum of aesthetic property names of \code{ggplot2}
}
\keyword{datasets}
