% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\docType{data}
\name{AestheticSelectionKeys}
\alias{AestheticSelectionKeys}
\title{AestheticSelectionKeys}
\format{
An object of class \code{list} of length 4.
}
\usage{
AestheticSelectionKeys
}
\description{
List of some \code{ggplot2} shapes
}
\keyword{datasets}
