% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-input.R
\name{AggregationInput}
\alias{AggregationInput}
\title{AggregationInput}
\description{
R6 class to be used to construct inputs to the AggregationSummary class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aggregationFunction}}{list of functions to use for aggregation}

\item{\code{aggregationFunctionName}}{vector of function names
that will be used as variable name of the aggregation}

\item{\code{aggregationUnit}}{unit of aggregation output}

\item{\code{aggregationDimension}}{dimension of aggregation output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AggregationInput-new}{\code{AggregationInput$new()}}
\item \href{#method-AggregationInput-clone}{\code{AggregationInput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AggregationInput-new"></a>}}
\if{latex}{\out{\hypertarget{method-AggregationInput-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{AggregationInput} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggregationInput$new(
  aggregationFunction = NULL,
  aggregationFunctionName = NULL,
  aggregationUnit = NULL,
  aggregationDimension = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aggregationFunction}}{list of functions to use for aggregation}

\item{\code{aggregationFunctionName}}{vector of function names
that will be used as variable name of the aggregation}

\item{\code{aggregationUnit}}{unit of aggregation output}

\item{\code{aggregationDimension}}{dimension of aggregation output}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{AggregationInput} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AggregationInput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AggregationInput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggregationInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
