% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-summary.R
\docType{class}
\name{AggregationSummary}
\alias{AggregationSummary}
\title{AggregationSummary}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AggregationSummary
}
\description{
#Splits the dataframe data into subsets defined by unique combinations of elements in the columns xColumnNames and groupingColumnNames.
}
\keyword{datasets}
