% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-enums.R
\docType{data}
\name{Alignments}
\alias{Alignments}
\title{Alignments}
\format{
An object of class \code{list} of length 3.
}
\usage{
Alignments
}
\description{
List of all available alignments/justifications for fonts
}
\seealso{
Other enum helpers: 
\code{\link{AestheticProperties}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{Linetypes}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{VerticalJustification}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
