% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-label.R
\name{asLabel}
\alias{asLabel}
\title{asLabel}
\usage{
asLabel(text = "", font = NULL)
}
\arguments{
\item{text}{A character value}

\item{font}{A \code{Font} object defining the font properties of the Label}
}
\value{
A \code{Label} object
}
\description{
Set a character string into a \code{Label} object associating font properties to input
If text is already a \code{Label} object, \code{asLabel} can be used to update its font properties
}
\examples{
title <- "Title of Plot"
title <- asLabel(title)
}
