% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{asThemeAestheticSelections}
\alias{asThemeAestheticSelections}
\title{asThemeAestheticSelections}
\usage{
asThemeAestheticSelections(themeSelectionObject)
}
\arguments{
\item{themeSelectionObject}{A \code{ThemeAestheticSelections} or a list that include values for selecting aesthetic properties}
}
\description{
Convert a list into a \code{ThemeAestheticSelections} object
}
\keyword{internal}
