% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-background.R
\docType{class}
\name{BackgroundConfiguration}
\alias{BackgroundConfiguration}
\title{BackgroundConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BackgroundConfiguration
}
\description{
Generic Background Configuration
}
\keyword{datasets}
