% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-background.R
\name{BackgroundElement}
\alias{BackgroundElement}
\title{BackgroundElement}
\description{
R6 class defining the properties of background elements
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fill}}{character defining the color filling of the background element}

\item{\code{color}}{character defining the color of the background element frame/line}

\item{\code{size}}{numeric defining the size of the background element frame/line}

\item{\code{linetype}}{character defining the size of the background element frame/line}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BackgroundElement$new()}}
\item \href{#method-createPlotElement}{\code{BackgroundElement$createPlotElement()}}
\item \href{#method-clone}{\code{BackgroundElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{BackgroundElement} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackgroundElement$new(fill = NULL, color = NULL, size = NULL, linetype = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fill}}{character color filling of the background element}

\item{\code{color}}{character color of the frame of the background element}

\item{\code{size}}{character size of the frame of the background element}

\item{\code{linetype}}{character linetype of the frame of the background element}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BackgroundElement} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createPlotElement"></a>}}
\if{latex}{\out{\hypertarget{method-createPlotElement}{}}}
\subsection{Method \code{createPlotElement()}}{
Create a \code{ggplot2::element_rect}  directly usable by \code{ggplot2::theme}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackgroundElement$createPlotElement(
  fill = NULL,
  color = NULL,
  size = NULL,
  linetype = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fill}}{character color filling of the background element}

\item{\code{color}}{character color of the frame of the background element}

\item{\code{size}}{character size of the frame of the background element}

\item{\code{linetype}}{character linetype of the frame of the background element}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{element_rect} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackgroundElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
