% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-background.R
\name{BackgroundElementConfiguration}
\alias{BackgroundElementConfiguration}
\title{BackgroundElementConfiguration}
\description{
R6 class defining the configuration of background elements
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fill}}{character color filling of the background element}

\item{\code{color}}{character color of the frame of the background element}

\item{\code{size}}{character size of the frame of the background element}

\item{\code{linetype}}{character linetype of the frame of the background element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BackgroundElementConfiguration$new()}}
\item \href{#method-print}{\code{BackgroundElementConfiguration$print()}}
\item \href{#method-clone}{\code{BackgroundElementConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{BackgroundElementConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackgroundElementConfiguration$new(
  fill = NULL,
  color = NULL,
  size = NULL,
  linetype = NULL,
  theme = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fill}}{character color filling of the background element}

\item{\code{color}}{character color of the frame of the background element}

\item{\code{size}}{character size of the frame of the background element}

\item{\code{linetype}}{character linetype of the frame of the background element}

\item{\code{theme}}{R6 class \code{Theme} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BackgroundElementConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print background element properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackgroundElementConfiguration$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Background element properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackgroundElementConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
