% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{data}
\name{bigTheme}
\alias{bigTheme}
\title{bigTheme}
\format{An object of class \code{Theme} (inherits from \code{ThemeFont}, \code{R6}) of length 14.}
\usage{
bigTheme
}
\description{
Big theme for plot configuration
}
\keyword{datasets}
