% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-boxwhisker.R
\docType{class}
\name{BoxWhiskerPlotConfiguration}
\alias{BoxWhiskerPlotConfiguration}
\title{BoxWhiskerPlotConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BoxWhiskerPlotConfiguration
}
\description{
Plot Configuration for Box Whisker Plots
}
\keyword{datasets}
