% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{data}
\name{bwTheme}
\alias{bwTheme}
\title{bwTheme}
\format{An object of class \code{Theme} (inherits from \code{ThemeFont}, \code{R6}) of length 16.}
\usage{
bwTheme
}
\description{
Black and White theme for plot configuration
}
\keyword{datasets}
