% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\docType{data}
\name{CaptionProperties}
\alias{CaptionProperties}
\title{CaptionProperties}
\format{
An object of class \code{list} of length 9.
}
\usage{
CaptionProperties
}
\description{
List of all Caption Properties
}
\keyword{internal}
