% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\name{createPlotScale}
\alias{createPlotScale}
\title{createPlotScale}
\usage{
createPlotScale(scale)
}
\arguments{
\item{scale}{character defining the name of the scale}
}
\value{
name of the \code{ggplot2} scale
}
\description{
Translate scale into a value directly usable by \code{ggplot2}
to give more flexibilty in the next functions
}
