% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-time-profile-datamapping.R
\name{CumulativeTimeProfileDataMapping}
\alias{CumulativeTimeProfileDataMapping}
\title{CumulativeTimeProfileDataMapping}
\description{
R6 class for mapping \code{x}, \code{y}, \code{GroupMapping} variables to \code{data}
}
\seealso{
Other DataMapping classes: 
\code{\link{BoxWhiskerDataMapping}},
\code{\link{DDIRatioDataMapping}},
\code{\link{GroupMapping}},
\code{\link{Grouping}},
\code{\link{HistogramDataMapping}},
\code{\link{ObsVsPredDataMapping}},
\code{\link{ObservedDataMapping}},
\code{\link{PKRatioDataMapping}},
\code{\link{QQDataMapping}},
\code{\link{RangeDataMapping}},
\code{\link{ResVsPredDataMapping}},
\code{\link{ResVsTimeDataMapping}},
\code{\link{TimeProfileDataMapping}},
\code{\link{TornadoDataMapping}},
\code{\link{XYDataMapping}},
\code{\link{XYGDataMapping}}
}
\concept{DataMapping classes}
\section{Super classes}{
\code{tlf::XYDataMapping} -> \code{tlf::XYGDataMapping} -> \code{CumulativeTimeProfileDataMapping}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CumulativeTimeProfileDataMapping-clone}{\code{CumulativeTimeProfileDataMapping$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tlf" data-topic="XYGDataMapping" data-id="checkMapData"><a href='../../tlf/html/XYGDataMapping.html#method-XYGDataMapping-checkMapData'><code>tlf::XYGDataMapping$checkMapData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="XYGDataMapping" data-id="initialize"><a href='../../tlf/html/XYGDataMapping.html#method-XYGDataMapping-initialize'><code>tlf::XYGDataMapping$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CumulativeTimeProfileDataMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CumulativeTimeProfileDataMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CumulativeTimeProfileDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
