% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-sub-classes.R
\name{CumulativeTimeProfilePlotConfiguration}
\alias{CumulativeTimeProfilePlotConfiguration}
\title{CumulativeTimeProfilePlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for cumulative time profile plots
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{AxisConfiguration}},
\code{\link{BackgroundConfiguration}},
\code{\link{BackgroundElement}},
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{ExportConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{LabelConfiguration}},
\code{\link{LegendConfiguration}},
\code{\link{LineElement}},
\code{\link{ObsVsPredPlotConfiguration}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{PlotGridConfiguration}},
\code{\link{QQPlotConfiguration}},
\code{\link{ResVsPredPlotConfiguration}},
\code{\link{ResVsTimePlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}},
\code{\link{XAxisConfiguration}},
\code{\link{YAxisConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{CumulativeTimeProfilePlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{defaultExpand}}{Default expand value when creating a \code{CumulativeTimeProfilePlotConfiguration} object}

\item{\code{colorPalette}}{color palette property from \code{ggplot2}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CumulativeTimeProfilePlotConfiguration-new}{\code{CumulativeTimeProfilePlotConfiguration$new()}}
\item \href{#method-CumulativeTimeProfilePlotConfiguration-clone}{\code{CumulativeTimeProfilePlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CumulativeTimeProfilePlotConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-CumulativeTimeProfilePlotConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CumulativeTimeProfilePlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CumulativeTimeProfilePlotConfiguration$new(colorPalette = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{colorPalette}}{color palette property from \code{ggplot2}}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{CumulativeTimeProfilePlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CumulativeTimeProfilePlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CumulativeTimeProfilePlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CumulativeTimeProfilePlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
