% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\name{dataMappingLabel}
\alias{dataMappingLabel}
\title{dataMappingLabel}
\usage{
dataMappingLabel(mapping = NULL, metaData = NULL)
}
\arguments{
\item{mapping}{mapping variable on label}

\item{metaData}{metaData containing unit of variable}
}
\value{
label character of variable with its unit
}
\description{
dataMappingLabel create an axis label with unit if available
}
