% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\docType{data}
\name{DDIComparisonTypes}
\alias{DDIComparisonTypes}
\title{DDIComparisonTypes}
\format{
An object of class \code{list} of length 2.
}
\usage{
DDIComparisonTypes
}
\description{
Options for comparison: residuals vs pred or obs vs pred
}
\keyword{datasets}
