% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-ddiratio.R
\docType{class}
\name{DDIRatioPlotConfiguration}
\alias{DDIRatioPlotConfiguration}
\title{DDIRatioPlotConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DDIRatioPlotConfiguration
}
\description{
Plot Configuration for DDIRatio
}
\keyword{datasets}
