% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\docType{data}
\name{DefaultDataMappingValues}
\alias{DefaultDataMappingValues}
\title{DefaultDataMappingValues}
\format{
An object of class \code{list} of length 6.
}
\usage{
DefaultDataMappingValues
}
\description{
List of default values used in dataMapping
}
\keyword{datasets}
