% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-molecule-plots.R
\name{.addLineLayer}
\alias{.addLineLayer}
\title{.addLineLayer}
\usage{
.addLineLayer(plotObject, type, value, position)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{type}{one of "horizontal", "vertical" or "diagonal"
Note that for "diagonal", geom_abline is used.
\code{value} of intercept is taken as is for linear scale but corresponds to the log of \code{value} for log scale.
For instance, intercept = c(-1, 0, 1) with log scale actually means that the line will go through c(0.1, 1, 10)
because c(-1, 0, 1) = log10(c(0.1, 1, 10)).}

\item{value}{value of xintercept or yintercept}

\item{position}{line position for aesthetic properties}
}
\value{
A \code{ggplot} object
}
\description{
Add line layer of a molecule plot
}
\keyword{internal}
