% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-molecule-plots.R
\name{.addScatterLayer}
\alias{.addScatterLayer}
\title{.addScatterLayer}
\usage{
.addScatterLayer(plotObject, data, mapLabels)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{data}{A data.frame with labels mapped to properties and obtained from a \code{DataMapping} object}

\item{mapLabels}{List of mapped label names passed to \code{ggplot2::aes_string}}
}
\value{
A \code{ggplot} object
}
\description{
Add scatter points layer of a molecule plot
}
\keyword{internal}
