% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-molecule-plots.R
\name{.applyColorPalette}
\alias{.applyColorPalette}
\title{.applyColorPalette}
\usage{
.applyColorPalette(plotObject, colorPalette = NULL)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{colorPalette}{A color palette name.
See enum \code{ColorPalettes} to get available color palettes.}
}
\value{
A \code{ggplot} object
}
\description{
Apply a color palette to a ggplot object
}
\keyword{internal}
