% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\name{.asLegendCaptionSubset}
\alias{.asLegendCaptionSubset}
\title{.asLegendCaptionSubset}
\usage{
.asLegendCaptionSubset(labels, unit = NULL)
}
\arguments{
\item{unit}{A character added as unit to label}
}
\description{
Creates default legend captions subset
}
\keyword{internal}
