% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{.checkIsBetween}
\alias{.checkIsBetween}
\title{Check if values are within \code{left}-\code{right} range(s)}
\usage{
.checkIsBetween(x, left, right, strict = FALSE)
}
\arguments{
\item{x}{Numeric values to assess}

\item{left}{Numeric value(s) used as lower bound}

\item{right}{Numeric value(s) used as upper bound}

\item{strict}{Logical value defining if \code{x} is strictly between \code{left} and \code{right}.
Default value is \code{FALSE}.}
}
\description{
Check if values are within \code{left}-\code{right} range(s)
}
\keyword{internal}
