% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{.checkIsInAvailableAngles}
\alias{.checkIsInAvailableAngles}
\title{Check if an input angle is within available angles}
\usage{
.checkIsInAvailableAngles(angle, availableAngles = c(0, 90, 180, 270))
}
\arguments{
\item{angle}{an input angle provided by user}

\item{availableAngles}{available angle for usage og \code{ggtext}}
}
\description{
Check if an input angle is within available angles
}
\keyword{internal}
