% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{.checkPlotFontFamily}
\alias{.checkPlotFontFamily}
\title{.checkPlotFontFamily}
\usage{
.checkPlotFontFamily(fontFamily)
}
\arguments{
\item{fontFamily}{character defining the family of font}
}
\value{
Name of font family if available in Windows font database
\code{NULL} otherwise
}
\description{
Check if font family is available in Windows font database.
Use function \code{grDevices::windowsFonts()} to get the list of font families available.
}
\keyword{internal}
