% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{.createPlotLegendPosition}
\alias{.createPlotLegendPosition}
\title{.createPlotLegendPosition}
\usage{
.createPlotLegendPosition(position)
}
\arguments{
\item{position}{Position of legend as defined in enum \code{LegendPositions}}
}
\value{
A list with fields \code{xPosition}, \code{xJustification}, \code{yPosition}, \code{yJustification}
}
\description{
Create legend position directly usable by \code{ggplot2}
}
\keyword{internal}
