% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\name{.createPlotScale}
\alias{.createPlotScale}
\title{.createPlotScale}
\usage{
.createPlotScale(scale)
}
\arguments{
\item{scale}{character defining the name of the scale}
}
\value{
A value available in enum \code{Scaling}
}
\description{
Translate scale into one of the values available in the enum \code{Scaling}.
}
\keyword{internal}
