% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\name{.createPlotTicks}
\alias{.createPlotTicks}
\title{.createPlotTicks}
\usage{
.createPlotTicks(ticks)
}
\arguments{
\item{ticks}{character, numeric or function defining the ticks}
}
\value{
name of the \code{ggplot2} scale
}
\description{
Translate ticks and ticklabels into a value directly usable by \code{ggplot2}
to give more flexibilty in the next functions
}
\keyword{internal}
