% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\name{.dataMappingLabel}
\alias{.dataMappingLabel}
\title{.dataMappingLabel}
\usage{
.dataMappingLabel(mapping = NULL, metaData = NULL)
}
\arguments{
\item{mapping}{Variable name in \code{data} and \code{metaData} to get label from}

\item{metaData}{List of information about \code{data} including \code{dimension} and \code{unit}}
}
\value{
Label as \code{dimention [unit]} from the mapping of the variable name
}
\description{
Get label with unit based on mapping
}
\keyword{internal}
