% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\name{.getAblineValues}
\alias{.getAblineValues}
\title{.getAblineValues}
\usage{
.getAblineValues(values, scale)
}
\arguments{
\item{values}{Numeric values to be transformed if scale is log}

\item{scale}{Scale of axis. Use enum \code{Scaling} to access names of scales.}
}
\value{
Numeric values
}
\description{
Apply log10 transformation to lines because plot is log scaled by default and geom_abline requires the log transformed values in input of intercept
Get list of values to provide to \code{lines} argument of \code{dataMapping} objects.
The \code{lines} are internally used as argument of \code{geom_hline} and \code{geom_abline} from \code{ggplot2}
}
\keyword{internal}
