% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\name{.getAesPropertyColumnNameFromLabels}
\alias{.getAesPropertyColumnNameFromLabels}
\title{.getAesPropertyColumnNameFromLabels}
\usage{
.getAesPropertyColumnNameFromLabels(mapLabels, propertyNames)
}
\arguments{
\item{mapLabels}{List of mapped label names passed to \code{ggplot2::aes}}

\item{propertyNames}{Names of aesthetic property (e.g. \code{"color"}, \code{"shape"}...)}
}
\value{
A list of variable names
}
\description{
Get the column names of the variables mapped to aesthetic properties
}
\keyword{internal}
