% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\name{.getAesPropertyLengthFromLabels}
\alias{.getAesPropertyLengthFromLabels}
\title{.getAesPropertyLengthFromLabels}
\usage{
.getAesPropertyLengthFromLabels(data, columnNames, propertyNames)
}
\arguments{
\item{data}{A data.frame with labels mapped to properties and obtained from a \code{DataMapping} object}

\item{columnNames}{List of mapped column names of \code{data} obtained}

\item{propertyNames}{Names of aesthetic property (e.g. \code{"color"}, \code{"shape"}...)}
}
\value{
A list of variable names
}
\description{
Get the names of the variables mapped to aesthetic properties
}
\keyword{internal}
