% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\name{.getAestheticValuesFromConfiguration}
\alias{.getAestheticValuesFromConfiguration}
\title{.getAestheticValuesFromConfiguration}
\usage{
.getAestheticValuesFromConfiguration(
  n = 1,
  position = 0,
  plotConfigurationProperty,
  propertyNames
)
}
\arguments{
\item{n}{integer defining size of returned aesthetic vector}

\item{position}{integer defining the current position in the aesthetic map}

\item{plotConfigurationProperty}{\code{PlotConfiguration} property name included in .
\code{"points"}, \code{"lines"}, \code{"ribbons"} or \code{"errorbars"}}

\item{propertyNames}{Names of aesthetic property (e.g. \code{"color"}, \code{"shape"}...)}

\item{plotObject}{A \code{ggplot} object}
}
\value{
A list of values for requested aesthetic property
}
\description{
Get list of values for requested aesthetic property
}
\keyword{internal}
