% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\name{.getColorNamesForFirstAesValues}
\alias{.getColorNamesForFirstAesValues}
\title{.getColorNamesForFirstAesValues}
\usage{
.getColorNamesForFirstAesValues(data, columnNames, propertyName)
}
\arguments{
\item{data}{A data.frame with labels mapped to properties and obtained from a \code{DataMapping} object}

\item{columnNames}{List of mapped column names of \code{data} obtained}

\item{propertyName}{Name of aesthetic property (e.g. \code{"shape"}...)}
}
\value{
Selected levels of \code{data[,columnNames$color]}
}
\description{
Get the first found value of map color for another map property
}
\keyword{internal}
