% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\name{.getNextAestheticValues}
\alias{.getNextAestheticValues}
\title{.getNextAestheticValues}
\usage{
.getNextAestheticValues(n, position = 0, map)
}
\arguments{
\item{n}{integer defining size of returned aesthetic vector}

\item{position}{integer defining at which position to look for aesthetic vector}

\item{map}{Aesthetic (e.g. color, shape, linetype) map from \code{Theme} object.}
}
\value{
Vector of aesthetics
}
\description{
Get the next aesthetic values (e.g color, shape, linetype) from an aesthetic map defined in \code{Theme}
}
\keyword{internal}
