% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\name{.getSameLimitsFromMapping}
\alias{.getSameLimitsFromMapping}
\title{.getSameLimitsFromMapping}
\usage{
.getSameLimitsFromMapping(data, dataMapping)
}
\arguments{
\item{data}{A data.frame with labels mapped to properties and obtained from a \code{DataMapping} object}

\item{dataMapping}{A \code{ObsVsPredDataMapping} object}
}
\description{
Get same axes limits from mapped data
}
\keyword{internal}
