% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enums.R
\name{.getTitlesFromFamilyTag}
\alias{.getTitlesFromFamilyTag}
\title{.getTitlesFromFamilyTag}
\usage{
.getTitlesFromFamilyTag(familyTag)
}
\arguments{
\item{familyTag}{Family tag used to document functions or R6 classes}
}
\value{
character array of titles
}
\description{
Get all title names from documented functions/R6 classes that have specific family tags.
The function aims at created automated and synchronized enums
}
\keyword{internal}
