% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{.isLogTicksIncludedInLimits}
\alias{.isLogTicksIncludedInLimits}
\title{Check that at least one log tick is included in limits}
\usage{
.isLogTicksIncludedInLimits(limits, scale)
}
\arguments{
\item{limits}{An array of numeric values}

\item{scale}{Name of log scale: \code{Scaling$log} for log10 scale, \code{Scaling$ln} for logarithmic scale}
}
\description{
Check that at least one log tick is included in limits
}
\keyword{internal}
