% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utilities.R
\name{.parseAddLineLayer}
\alias{.parseAddLineLayer}
\title{.parseAddLineLayer}
\usage{
.parseAddLineLayer(type, value, position)
}
\arguments{
\item{type}{one of "horizontal", "vertical" or "diagonal"
Note that for "diagonal", geom_abline is used.
\code{value} of intercept is taken as is for linear scale but corresponds to the log of \code{value} for log scale.
For instance, intercept = c(-1, 0, 1) with log scale actually means that the line will go through c(0.1, 1, 10)
because c(-1, 0, 1) = log10(c(0.1, 1, 10)).}

\item{value}{value of xintercept or yintercept}

\item{position}{line position for aesthetic properties}
}
\value{
An expression to \code{eval()}
}
\description{
Create an expression that adds scatter plot layer
TODO: create a vignette explaining how argument \code{lines} in dataMapping is related to this
}
\keyword{internal}
