% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utilities.R
\name{.parseUpdateAxes}
\alias{.parseUpdateAxes}
\title{.parseUpdateAxes}
\usage{
.parseUpdateAxes()
}
\value{
An expression to \code{eval()}
}
\description{
Create an expression that updates the plot axes
}
\keyword{internal}
