% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utilities.R
\name{.parseVariableFromObject}
\alias{.parseVariableFromObject}
\title{.parseVariableFromObject}
\usage{
.parseVariableFromObject(objectName, variableName, keepIfNull = FALSE)
}
\arguments{
\item{objectName}{Name of the object whose field is updated}

\item{variableName}{Name of the variable and field of \code{objectName}}

\item{keepIfNull}{logical \code{variableName <- objectName$variableName \%||\% variableName}}
}
\value{
An expression to \code{eval()}
}
\description{
Create an expression of type \code{variableName <- objectName$variableName}
}
\keyword{internal}
