% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-label.R
\name{.sanitizeLabel}
\alias{.sanitizeLabel}
\title{Sanitize Label Text}
\usage{
.sanitizeLabel(text)
}
\arguments{
\item{text}{a character string}
}
\value{
a sanitized character string
}
\description{
ggtext does not allow certain characters that can be converted to html
tags but that are not supported. This function removes this forbidden
characters.
}
\examples{
.sanitizeLabel("`code`")
}
