% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot.R
\name{.selectFontFamily}
\alias{.selectFontFamily}
\title{.selectFontFamily}
\usage{
.selectFontFamily(fontfamily = "sans")
}
\arguments{
\item{fontfamily}{default font family}
}
\description{
Select appropriate font family based on font and \code{showtext} package availability
}
\keyword{internal}
