% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utilities.R
\name{.setDataMapping}
\alias{.setDataMapping}
\title{.setDataMapping}
\usage{
.setDataMapping(dataMapping, DataMappingClass, data = NULL)
}
\arguments{
\item{dataMapping}{A \code{DataMappingClass} object}

\item{DataMappingClass}{Required class for \code{dataMapping}}

\item{data}{A data.frame potentially used for smart mapping}
}
\value{
A \code{DataMapping} object
}
\description{
Set \code{DataMapping} object internally using \code{tlf} default if \code{dataMapping} is not provided
}
\keyword{internal}
