% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utilities.R
\name{.setPlotConfiguration}
\alias{.setPlotConfiguration}
\title{.setPlotConfiguration}
\usage{
.setPlotConfiguration(
  plotConfiguration,
  PlotConfigurationClass,
  data = NULL,
  metaData = NULL,
  dataMapping = NULL
)
}
\arguments{
\item{plotConfiguration}{A \code{PlotConfigurationClass} object}

\item{PlotConfigurationClass}{Required class for \code{plotConfiguration}}

\item{data}{A data.frame potentially used for smart plot configuration}

\item{metaData}{A list of meta data potentially used for smart plot configuration}

\item{dataMapping}{A \code{DataMapping} object potentially used for smart plot configuration}
}
\value{
A \code{PlotConfiguration} object
}
\description{
Set \code{PlotConfiguration} object internally using \code{tlf} default if \code{plotConfiguration} is not provided
}
\keyword{internal}
