% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utilities.R
\name{.setPlotObject}
\alias{.setPlotObject}
\title{.setPlotObject}
\usage{
.setPlotObject(plotObject, plotConfiguration = NULL)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{plotConfiguration}{A \code{PlotConfiguration} object}
}
\value{
A \code{ggplot} object
}
\description{
Set a \code{ggplot} object associated with its \code{plotConfiguration}
}
\keyword{internal}
