% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\name{.smartMapping}
\alias{.smartMapping}
\title{.smartMapping}
\usage{
.smartMapping(data)
}
\arguments{
\item{data}{data.frame on which the smart mapping is used}
}
\value{
A list of usual mappings that can be guessed from data
}
\description{
Check data size and variable names,
Get the mapping if variable names correspond to usual aesthetic or mapping names.
Else return NULL for the mapping property.

If data has only one variable, it will be mapped as \code{x}.
If data has only 2 variables, they will be respectively mapped as \code{x} and \code{y}.

Recognized names for the mapping are \code{x}, \code{y}, \code{ymin}, \code{ymax},
\code{lower}, \code{middle}, \code{upper}, \code{color}, \code{shape}, \code{linetype},
\code{size} and \code{fill}
}
\examples{
\dontrun{
# Get variable names x and y directly from data
data <- data.frame(x = c(1, 2, 3), y = c(6, 5, 4), z = c(7, 8, 9))
mapping <- .smartMapping(data)

# If data has aesthetic propoerties
data <- data.frame(x = c(1, 2, 3), y = c(6, 5, 4), color = c("blue", "red", "blue"))
mapping <- .smartMapping(data)
}
}
