% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\name{.updateAesProperties}
\alias{.updateAesProperties}
\title{.updateAesProperties}
\usage{
.updateAesProperties(
  plotObject,
  plotConfigurationProperty,
  propertyNames,
  data,
  mapLabels
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{plotConfigurationProperty}{\code{PlotConfiguration} property name included in .
\code{"points"}, \code{"lines"}, \code{"ribbons"} or \code{"errorbars"}}

\item{propertyNames}{Names of aesthetic property (e.g. \code{"color"}, \code{"shape"}...)}

\item{data}{A data.frame with labels mapped to properties and obtained from a \code{DataMapping} object}

\item{mapLabels}{List of mapped label names passed to \code{ggplot2::aes_string}}
}
\value{
A \code{ggplot} object
}
\description{
Updates the aesthetic properties of \code{plotObject}
}
\keyword{internal}
