% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utilities.R
\name{.updateSameAxes}
\alias{.updateSameAxes}
\title{.updateSameAxes}
\usage{
.updateSameAxes(plotObject, data, dataMapping)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{data}{A data.frame}

\item{dataMapping}{A \code{DataMapping} object}
}
\value{
A \code{ggplot} object
}
\description{
Updates plot configuration axes to get same limits
}
\keyword{internal}
