% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utilities.R
\name{.updateSymmetricAxes}
\alias{.updateSymmetricAxes}
\title{.updateSymmetricAxes}
\usage{
.updateSymmetricAxes(plotObject, data, dataMapping)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{data}{A data.frame}

\item{dataMapping}{A \code{DataMapping} object}
}
\value{
A \code{ggplot} object
}
\description{
Updates plot configuration axes to get symmetric limits
}
\keyword{internal}
