% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{.validateEitherOrNullInput}
\alias{.validateEitherOrNullInput}
\title{Check that 2 conflicting inputs can't be provided at the same time}
\usage{
.validateEitherOrNullInput(eitherInput, orInput)
}
\arguments{
\item{eitherInput}{Input or list of inputs}

\item{orInput}{Input or list of inputs}
}
\description{
Check that 2 conflicting inputs can't be provided at the same time
}
\keyword{internal}
