% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{.validateIsStrictlyPositive}
\alias{.validateIsStrictlyPositive}
\title{Validate if values are strictly positive}
\usage{
.validateIsStrictlyPositive(values, nullAllowed = FALSE)
}
\arguments{
\item{values}{numeric values to be validated}

\item{nullAllowed}{logical allowing \code{NULL} as input}
}
\description{
Validate if values are strictly positive
}
\keyword{internal}
