% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enums.R
\docType{data}
\name{.ViridisPalettes}
\alias{.ViridisPalettes}
\title{.ViridisPalettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
.ViridisPalettes
}
\description{
Enum of viridis color palettes
}
\keyword{internal}
