% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-export.R
\name{ExportConfiguration}
\alias{ExportConfiguration}
\title{ExportConfiguration}
\description{
R6 class defining properties for saving a \code{ggplot} object
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{AxisConfiguration}},
\code{\link{BackgroundConfiguration}},
\code{\link{BackgroundElement}},
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{CumulativeTimeProfilePlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{LabelConfiguration}},
\code{\link{LegendConfiguration}},
\code{\link{LineElement}},
\code{\link{ObsVsPredPlotConfiguration}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{PlotGridConfiguration}},
\code{\link{QQPlotConfiguration}},
\code{\link{ResVsPredPlotConfiguration}},
\code{\link{ResVsTimePlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}},
\code{\link{XAxisConfiguration}},
\code{\link{YAxisConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character defining the name of the file to be saved (without extension)}

\item{\code{path}}{Path of the directory to save plot to: path and filename are
combined to create the fully qualified file name. Defaults to the working directory.}

\item{\code{format}}{character defining the format of the file to be saved}

\item{\code{width}}{numeric values defining the width in \code{units} of the plot dimensions after saving}

\item{\code{height}}{numeric values defining the height in \code{units} of the plot dimensions after saving}

\item{\code{units}}{character defining the unit of the saving dimension}

\item{\code{dpi}}{(dots per inch) numeric value defining plot resolution}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExportConfiguration-new}{\code{ExportConfiguration$new()}}
\item \href{#method-ExportConfiguration-print}{\code{ExportConfiguration$print()}}
\item \href{#method-ExportConfiguration-getFileName}{\code{ExportConfiguration$getFileName()}}
\item \href{#method-ExportConfiguration-savePlot}{\code{ExportConfiguration$savePlot()}}
\item \href{#method-ExportConfiguration-convertPixels}{\code{ExportConfiguration$convertPixels()}}
\item \href{#method-ExportConfiguration-clone}{\code{ExportConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExportConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ExportConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportConfiguration$new(
  path = NULL,
  name = NULL,
  format = NULL,
  width = NULL,
  height = NULL,
  units = NULL,
  dpi = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path of the directory to save plot to: path and filename are
combined to create the fully qualified file name. Defaults to the working directory.}

\item{\code{name}}{character defining the name of the file to be saved (without extension)}

\item{\code{format}}{character defining the format of the file to be saved.}

\item{\code{width}}{numeric values defining the width in \code{units} of the plot dimensions after saving}

\item{\code{height}}{numeric values defining the height in \code{units} of the plot dimensions after saving}

\item{\code{units}}{character defining the unit of the saving dimension}

\item{\code{dpi}}{numeric value defining plot resolution (dots per inch)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ExportConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportConfiguration-print"></a>}}
\if{latex}{\out{\hypertarget{method-ExportConfiguration-print}{}}}
\subsection{Method \code{print()}}{
Print properties of export configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportConfiguration$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Export configuration properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportConfiguration-getFileName"></a>}}
\if{latex}{\out{\hypertarget{method-ExportConfiguration-getFileName}{}}}
\subsection{Method \code{getFileName()}}{
Print the default exported file name from the export configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportConfiguration$getFileName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Default file name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportConfiguration-savePlot"></a>}}
\if{latex}{\out{\hypertarget{method-ExportConfiguration-savePlot}{}}}
\subsection{Method \code{savePlot()}}{
Save/Export a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportConfiguration$savePlot(plotObject, fileName = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{A \code{ggplot} object}

\item{\code{fileName}}{character file name of the exported plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file name of the exported plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportConfiguration-convertPixels"></a>}}
\if{latex}{\out{\hypertarget{method-ExportConfiguration-convertPixels}{}}}
\subsection{Method \code{convertPixels()}}{
If unit is in pixels, convert all export dimensions to inches to keep compatibility with older versions of ggplot2
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportConfiguration$convertPixels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExportConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExportConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
