% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-export.R
\docType{data}
\name{ExportFormats}
\alias{ExportFormats}
\title{ExportFormats}
\format{
An object of class \code{list} of length 10.
}
\usage{
ExportFormats
}
\description{
List of all available formats to export a ggplot object
}
\seealso{
Other enum helpers: 
\code{\link{AestheticProperties}},
\code{\link{Alignments}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{Linetypes}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{VerticalJustification}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
