% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-export.R
\name{exportPlotConfigurationCode}
\alias{exportPlotConfigurationCode}
\title{exportPlotConfigurationCode}
\usage{
exportPlotConfigurationCode(plotConfiguration, name = "plotConfiguration")
}
\arguments{
\item{plotConfiguration}{A \code{PlotConfiguration} object}

\item{name}{Name of \code{PlotConfiguration} object in the created code}
}
\value{
R code to recreate the plot configuration as character
}
\description{
Export a plot configuration as R code
}
