% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\docType{class}
\name{Font}
\alias{Font}
\title{Font}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Font
}
\description{
Font class defining properties of a text (size, color, font)
}
\keyword{datasets}
