% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\docType{data}
\name{FontFaces}
\alias{FontFaces}
\title{FontFaces}
\format{
An object of class \code{list} of length 4.
}
\usage{
FontFaces
}
\description{
List of all available font faces
}
\seealso{
Other enum helpers: 
\code{\link{AestheticProperties}},
\code{\link{Alignments}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{LegendPositions}},
\code{\link{Linetypes}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
