% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot.R
\docType{data}
\name{GeomTLFPoint}
\alias{GeomTLFPoint}
\title{GeomTLFPoint}
\format{
An object of class \code{GeomTLFPoint} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomTLFPoint
}
\description{
Define a Geom using \code{ggplot2::ggproto()} and based on GeomPoint.
The Geom internally uses \code{textGrob} instead of \code{pointsGrob} so that fonts leverage for drawing shapes.
The \code{grid} and \code{scales} packages are supposed to be required by \code{ggplot2}.
So there should not be any issue as installing \code{ggplot2} should install those 2 packages.
}
\keyword{datasets}
