% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\name{getAesStringMapping}
\alias{getAesStringMapping}
\title{getAesStringMapping}
\usage{
getAesStringMapping(dataMapping)
}
\arguments{
\item{dataMapping}{DataMapping class or subclass object}
}
\value{
A list of mappings that can be used as is by \code{aes_string}.
}
\description{
Get the \code{dataMapping} elements and convert them into
character string usable by ggplot mapping function \code{aes_string}.
The conversion fixes any issue of special characters by wrapping the string by \code{``}
if not already used.

\code{dataMapping} unmapped aesthetic elements (e.g. \code{color}) are associated to
the dummy aesthetic variable \code{"defaultAes"} which allows further modification of the plot aesthetics.
}
\examples{
\dontrun{
dataMapping <- XYGDataMapping$new(x = "Time [h]", y = "Concentration [mol/L]", color = "Dose")
mappingLabels <- getAesStringMapping(dataMapping)
}
}
