% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\name{getAestheticValues}
\alias{getAestheticValues}
\title{getAestheticValues}
\usage{
getAestheticValues(n, selectionKey = NA, position = 0, aesthetic = "color")
}
\arguments{
\item{n}{integer defining size of returned aesthetic vector}

\item{selectionKey}{value of aesthetic to be returned or key function from enum \code{AestheticSelectionKeys}}

\item{position}{integer defining the current position in the aesthetic map}

\item{aesthetic}{name of aesthetic property as defined in enum \code{AestheticProperties}}
}
\value{
Vector of aesthetics
}
\description{
Get aesthetic values (e.g color, shape, linetype) based on a selected strategy
}
