% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxwhisker-get-measure.R
\name{getBoxWhiskerMeasure}
\alias{getBoxWhiskerMeasure}
\title{getBoxWhiskerMeasure}
\usage{
getBoxWhiskerMeasure(
  data,
  dataMapping = NULL,
  y = NULL,
  group = NULL,
  quantiles = c(0.05, 0.25, 0.5, 0.75, 0.95)
)
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{dataMapping}{A \code{BoxWhiskerDataMapping} object mapping \code{x}, \code{y} and aesthetic groups to their variable names of \code{data}.}

\item{y}{Name of \code{y} variable in \code{data}.}

\item{group}{Name of grouping variable in \code{data}.}

\item{quantiles}{Numeric values between 0 and 1 defining the quantiles to summarize}
}
\value{
A data.frame of summary statistics
}
\description{
Get a summary table of Box Whisker percentiles
}
\examples{
# Get box-and-whisker plots of log-normal distributed data
boxData <- data.frame(x = c(rep("A", 500), rep("B",500)), y = rlnorm(1000))

getBoxWhiskerMeasure(data = boxData, dataMapping = BoxWhiskerDataMapping$new(x = "x", y = "y"))

}
