% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\name{getDefaultCaptions}
\alias{getDefaultCaptions}
\title{getDefaultCaptions}
\usage{
getDefaultCaptions(data, metaData, variableList = colnames(data),
  sep = "-")
}
\arguments{
\item{data}{input data.frame with variables to group by}

\item{metaData}{input data.frame with variables to group by}

\item{variableList}{groups as factor levels}

\item{sep}{characters separating variables in caption}
}
\value{
groupingVariable
}
\description{
getDefaultCaptions create a new column that groups the grouping variable
}
