% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-histogram.R
\name{getDistributionFit}
\alias{getDistributionFit}
\title{getDistributionFit}
\usage{
getDistributionFit(data, dataMapping)
}
\arguments{
\item{data}{data.frame containing the data to be used for the plot}

\item{dataMapping}{A \code{HistogramDataMapping} object
The object defines the distribution to be fitted and the option \code{stack}.
If the bars are stacked, the fit will account for the final histogram}
}
\value{
A data.frame with \code{x}, \code{y} and \code{legendLabels}
}
\description{
Get a data.frame from the fit of a distribution provided in \code{dataMapping}
If \code{normal} distribution is selected, its mean is plotted
If \code{logNormal} distribution is selected, its mode is plotted
}
\keyword{internal}
