% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-histogram.R
\name{getDistributionMed}
\alias{getDistributionMed}
\title{getDistributionMed}
\usage{
getDistributionMed(data, dataMapping)
}
\arguments{
\item{data}{data.frame containing the data to be used for the plot}

\item{dataMapping}{A \code{HistogramDataMapping} object
The object defines the distribution to be fitted and the option \code{stack}.
If the bars are stacked, the fit will account for the final histogram}
}
\value{
Numeric values for vertical lines
}
\description{
Get an array of values from the fit of a distribution provided in \code{dataMapping}
If \code{normal} distribution is selected, its mean is plotted
If \code{logNormal} distribution is selected, its mode is plotted
}
\keyword{internal}
