% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-molecule-plots.R
\name{getDualAxisPlot}
\alias{getDualAxisPlot}
\title{getDualAxisPlot}
\usage{
getDualAxisPlot(leftPlotObject, rightPlotObject)
}
\arguments{
\item{leftPlotObject}{A \code{ggplot} object with left y-axis}

\item{rightPlotObject}{A \code{ggplot} object with right y-axis}
}
\value{
A \code{ggplot} object with dual y-axis
}
\description{
Check if dual Y Axis is needed
}
