% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGrouping.R
\name{getGrouping}
\alias{getGrouping}
\title{Get Data Grouping}
\usage{
getGrouping(data, grouping)
}
\arguments{
\item{data}{input data.frame with variable to group}

\item{grouping}{vector of variable names to group by}
}
\value{
data with groups as factor
}
\description{
Function to group data and create groups according to variable
So far, the grouping is performed only along one variable
}
