% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ddiratio.R
\name{getGuestValues}
\alias{getGuestValues}
\title{getGuestValues}
\usage{
getGuestValues(x, delta = 1, residualsVsObserved = FALSE)
}
\arguments{
\item{x}{Numeric values input of Guest function}

\item{delta}{Numeric value parameter of Guest function}

\item{residualsVsObserved}{Logical value defining
if limits are calculated as residuals vs observed, instead of predicted vs observed.}
}
\value{
A data.frame with \code{x}, \code{ymin} and \code{ymax} defining Guest et al. limits
}
\description{
Get a data.frame with Guest et al. ratio limits with:
\itemize{
\item \code{ymax} = \code{x}.\code{limit}
\item \code{ymin} = \code{x}/\code{limit}
\item \code{limit} = (\code{delta}+2(\code{x}-1))/\code{x}
}
}
\examples{
# Get predicted vs observed Guest et al. limits
getGuestValues(x = 10^seq(-2, 2, 0.2))

# Get residuals vs observed Guest et al. limits
getGuestValues(x = 10^seq(-2, 2, 0.2), residualsVsObserved = TRUE)

}
\references{
\url{https://dmd.aspetjournals.org/content/39/2/170}
}
