% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ddiratio.R
\name{getGuestValuesFromDataMapping}
\alias{getGuestValuesFromDataMapping}
\title{getGuestValuesFromDataMapping}
\usage{
getGuestValuesFromDataMapping(data, dataMapping)
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{dataMapping}{A \code{DDIRatioDataMapping} object mapping \code{x}, \code{y} and aesthetic groups to their variable names of \code{data}.}
}
\value{
A data.frame with \code{x}, \code{ymin} and \code{ymax} defining Guest et al. limits
}
\description{
Get a data.frame with Guest et al. ratio limits from \code{data} and its \code{DDIRatioDataMapping}
}
\examples{
# Get the data.frame of Guest et al. limits
ddiData <- data.frame(x = c(1, 2, 1, 2, 3), y = c(5, 0.2, 2, 3, 4))

getGuestValuesFromDataMapping(
  data = ddiData,
  dataMapping = DDIRatioDataMapping$new(x = "x", y = "y")
)

}
