% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-label.R
\name{getLabelWithUnit}
\alias{getLabelWithUnit}
\title{getLabelWithUnit}
\usage{
getLabelWithUnit(label, unit = NULL)
}
\arguments{
\item{label}{text of axis label}

\item{unit}{Character value corresponding to unit of \code{label}}
}
\value{
\code{label [unit]} or \code{label} depending if \code{unit} is \code{NULL} or \code{""}
}
\description{
Get label with its unit within square brackets when available
}
\examples{
getLabelWithUnit("Time", "min")

getLabelWithUnit("Label without unit")

}
