% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R, R/utilities-legend.R
\name{getLegendCaption}
\alias{getLegendCaption}
\title{getLegendCaption}
\usage{
getLegendCaption(plotObject)

getLegendCaption(plotObject)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{captionData}{data used for legend caption}

\item{which}{caption position in the legend}
}
\value{
legend captions

A data.frame corresponding to legend caption
}
\description{
getLegendCaption return the legend captions or a subset using which

Get the legend caption
}
