% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{getLegendPosition}
\alias{getLegendPosition}
\title{getLegendPosition}
\usage{
getLegendPosition(plotObject)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}
}
\value{
Position of legend as defined in enum \code{LegendPositions}
}
\description{
Get the legend position
}
\seealso{
LegendPositions
}
