% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\name{getLinesFromFoldDistance}
\alias{getLinesFromFoldDistance}
\title{getLinesFromFoldDistance}
\usage{
getLinesFromFoldDistance(foldDistance)
}
\arguments{
\item{foldDistance}{Numeric values}
}
\value{
A list of numeric values
}
\description{
Get list of values to provide to \code{lines} argument of \code{dataMapping} objects.
The \code{lines} are internally used as argument of \code{geom_hline} and \code{geom_abline} from \code{ggplot2}
}
\examples{

# Get lines for identity and 2-fold distance
getLinesFromFoldDistance(c(1, 2))

}
