% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{getLnTickLabels}
\alias{getLnTickLabels}
\title{getLnTickLabels}
\usage{
getLnTickLabels(ticks)
}
\arguments{
\item{ticks}{numeric values of the ticks}
}
\value{
Expressions to use in \code{ticklabels} input parameter of \code{setXAxis} and \code{setYAxis} functions
}
\description{
Get ticklabels expressions for ln scale plots
}
\examples{
ticks <- exp(c(1, 5, 10, 50, 100, 500))
getLnTickLabels(ticks)
}
