% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{getPercentileTickLabels}
\alias{getPercentileTickLabels}
\title{getPercentileTickLabels}
\usage{
getPercentileTickLabels(ticks)
}
\arguments{
\item{ticks}{numeric values of the ticks}
}
\value{
Expressions to use in \code{ticklabels} input parameter of \code{setXAxis} and \code{setYAxis} functions
}
\description{
Get ticklabels expressions for percentiles of normal distribution scale plots
}
\examples{
ticks <- rnorm(5)
getPercentileTickLabels(ticks)

# Get percentile of normal distribution
ticks <- qnorm(seq(1, 9) / 10)
getPercentileTickLabels(ticks)

}
