% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkratio-get-measure.R
\name{getPKRatioMeasure}
\alias{getPKRatioMeasure}
\title{getPKRatioMeasure}
\usage{
getPKRatioMeasure(data, dataMapping = NULL, ratioLimits = c(1.5, 2))
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{dataMapping}{A \code{PKRatioDataMapping} object mapping \code{x}, \code{y} and aesthetic groups to their variable names of \code{data}.}

\item{ratioLimits}{Numeric positive values limits}
}
\value{
A data.frame of summary of PK Ratio within specific limits
}
\description{
Get a summary table of PK Ratio within specific limits
}
\examples{
# Get summary of usual PK Ratio limits
pkData <- data.frame(x = c(1, 2, 1, 2, 3), y = c(5, 0.2, 2, 3, 4))

getPKRatioMeasure(data = pkData, dataMapping = PKRatioDataMapping$new(x = "x", y = "y"))

# Get summary of other PK Ratio limits
getPKRatioMeasure(
data = pkData, 
dataMapping = PKRatioDataMapping$new(x = "x", y = "y"),
ratioLimits = seq(1.5, 5, 0.5)
)

}
