% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getSameLimits}
\alias{getSameLimits}
\title{getSameLimits}
\usage{
getSameLimits(...)
}
\arguments{
\item{...}{numeric values}
}
\value{
An array of 2 values
}
\description{
Get same limits from multiple sets of values
}
\examples{
getSameLimits(seq(-3, 8), seq(-12, 4))

}
