% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\name{getSpecificAestheticValues}
\alias{getSpecificAestheticValues}
\title{getSpecificAestheticValues}
\usage{
getSpecificAestheticValues(n, values)
}
\arguments{
\item{n}{integer defining size of returned aesthetic vector}

\item{values}{values of the aesthetic to be returned}
}
\value{
Vector of aesthetics
}
\description{
Get specific aesthetic values (e.g color, shape, linetype)
}
