% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getSymmetricLimits}
\alias{getSymmetricLimits}
\title{getSymmetricLimits}
\usage{
getSymmetricLimits(values)
}
\arguments{
\item{values}{numeric values}
}
\value{
An array of 2 symmetric values equally distant from 0
}
\description{
Get symmetric limits from a set of values
}
\examples{
getSymmetricLimits(seq(-3, 8))

}
