% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{getThemePropertyFor}
\alias{getThemePropertyFor}
\title{getThemePropertyFor}
\usage{
getThemePropertyFor(plotConfiguration, propertyName = NULL)
}
\arguments{
\item{plotConfiguration}{\code{PlotConfigurations} object}

\item{propertyName}{Name of property. One \code{lines}, \code{points}, \code{ribbons} or \code{errorbars}}
}
\description{
Clone a theme plot configuration property so the property is not affected in current theme
}
\keyword{internal}
