% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tlf.R
\name{getTLFSettings}
\alias{getTLFSettings}
\title{getTLFSettings}
\usage{
getTLFSettings(settingName)
}
\arguments{
\item{settingName}{setting name as defined in enum \code{tlfSettingNames}}
}
\description{
Get Names of the default/global settings stored in tlfEnv.
Can be used with \code{getTLFSettings()}
}
