% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{\\%||\\%}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{argument 1}

\item{rhs}{argument 2}
}
\value{
lhs if lhs is not null, rhs otherwise
}
\description{
Check if lhs argument is not null, output lhs if not null,
output rhs otherwise
}
\details{
Shortkey checking if argument 1 is not null,
output argument 1 if not null, or output argument 2 otherwise
}
