% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-groupings.R
\docType{class}
\name{Groupings}
\alias{Groupings}
\title{Groupings}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Groupings
}
\description{
Abstract class for Groupings
}
\keyword{datasets}
