% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-groupmapping.R
\docType{class}
\name{GroupMapping}
\alias{GroupMapping}
\title{GroupMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupMapping
}
\description{
Abstract class for GroupMapping
}
\section{Fields}{

\describe{
\item{\code{color}}{Grouping class mapping variables to colors}

\item{\code{fill}}{Grouping class mapping variables to color fillings}

\item{\code{linetype}}{Grouping class mapping variables to line types}

\item{\code{shape}}{Grouping class mapping variables to shapes}

\item{\code{size}}{Grouping class mapping variables to sizes}
}}

\keyword{datasets}
