% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-histogram.R
\docType{class}
\name{HistogramDataMapping}
\alias{HistogramDataMapping}
\title{HistogramDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HistogramDataMapping
}
\description{
Class for Histogram Group Mapping
}
\keyword{datasets}
