% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-histogram.R
\docType{class}
\name{HistogramPlotConfiguration}
\alias{HistogramPlotConfiguration}
\title{HistogramPlotConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HistogramPlotConfiguration
}
\description{
Plot Configuration for Histograms
}
\keyword{datasets}
