% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-grid.R
\docType{data}
\name{HorizontalJustification}
\alias{HorizontalJustification}
\title{HorizontalJustification}
\format{
An object of class \code{list} of length 3.
}
\usage{
HorizontalJustification
}
\description{
List of all available horizontal justifications for plot annotation text.
}
\seealso{
Other enum helpers: 
\code{\link{AestheticProperties}},
\code{\link{Alignments}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{LegendPositions}},
\code{\link{Linetypes}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{VerticalJustification}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
