% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifEqual}
\alias{ifEqual}
\title{ifEqual}
\usage{
ifEqual(x, y, outputIfEqual, outputIfNotEqual = NULL)
}
\arguments{
\item{x}{argument 1}

\item{y}{argument 2}

\item{outputIfEqual}{argument 3}

\item{outputIfNotEqual}{argument 4}
}
\value{
outputIfEqual if x=y, outputIfNotEqual otherwise
}
\description{
Check if x=y, if so output outputIfEqual,
otherwise, output outputIfNotEqual
}
\details{
Shortkey checking if arguments 1 and 2 are equal,
output argument 3 if equal, or output argument 4 otherwise
}
