% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifIncluded}
\alias{ifIncluded}
\title{ifIncluded}
\usage{
ifIncluded(x, y, outputIfIncluded, outputIfNotIncluded = NULL)
}
\arguments{
\item{x}{argument 1}

\item{y}{argument 2}

\item{outputIfIncluded}{argument 3}

\item{outputIfNotIncluded}{argument 4}
}
\value{
outputIfIncluded if x=y, outputIfNotIncluded otherwise
}
\description{
Check if x is in y, if so output outputIfIncluded,
otherwise, output outputIfNotIncluded
}
\details{
Shortkey checking if arguments 1 is included in 2,
output argument 3 if included, or output argument 4 otherwise
}
\keyword{internal}
