% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifnotnull}
\alias{ifnotnull}
\title{ifnotnull}
\usage{
ifnotnull(inputToCheck, outputIfNotNull, outputIfNull = NULL)
}
\arguments{
\item{inputToCheck}{argument 1}

\item{outputIfNotNull}{argument 2}

\item{outputIfNull}{argument 3}
}
\value{
outputIfNotNull if inputToCheck is not null, outputIfNull otherwise
}
\description{
Check if inputToCheck is not null, if so output outputIfNotNull,
otherwise, output outputIfNull
}
\details{
Shortkey checking if argument 1 is not null,
output the argument 2 if not null, or output argument 3 otherwise
}
