% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom-plots.R
\name{initializePlot}
\alias{initializePlot}
\title{initializePlot}
\usage{
initializePlot(plotConfiguration = NULL)
}
\arguments{
\item{plotConfiguration}{An optional \code{PlotConfiguration} object defining labels, grid, background and watermark}
}
\value{
A \code{ggplot} graphical object
}
\description{
Initialize a \code{ggplot} object and set its labels, grid, background and watermark
}
\examples{
# Initialize an empty plot
p <- initializePlot()

# Implement a customized configuration using PlotConfiguration
config <- PlotConfiguration$new(title = "My Plot", xlabel = "x variable", ylabel = "y variable")
p <- initializePlot(config)

}
\references{
For examples, see:
\url{https://www.open-systems-pharmacology.org/TLF-Library/articles/atom-plots.html}
}
\seealso{
Other atom plots: 
\code{\link{addErrorbar}()},
\code{\link{addLine}()},
\code{\link{addRibbon}()},
\code{\link{addScatter}()}
}
\concept{atom plots}
