% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isBetween}
\alias{isBetween}
\title{isBetween}
\usage{
isBetween(x, left, right, strict = FALSE)
}
\arguments{
\item{x}{Numeric values to assess}

\item{left}{Numeric value(s) used as lower bound}

\item{right}{Numeric value(s) used as upper bound}

\item{strict}{Logical value defining if \code{x} is strictly between \code{left} and \code{right}.
Default value is \code{FALSE}.}
}
\value{
Logical values
}
\description{
Assess if \code{x} is between \code{left} and \code{right} bounds.
Shortcut for \code{x >= left & x <= right} if \code{strict=FALSE} (default).
Shortcut for \code{x > left & x < right} if \code{strict=TRUE}.
}
\examples{
isBetween(1:12, 7, 9)

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]

isBetween(x, cos(x) + 1, cos(x) - 1)
}
