% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{isOfLength}
\alias{isOfLength}
\title{Check if the provided object has nbElements elements}
\usage{
isOfLength(object, nbElements)
}
\arguments{
\item{object}{An object or a list of objects}

\item{nbElements}{number of elements that are supposed in object}
}
\value{
TRUE if the object or all objects inside the list have nbElements.
Only the first level of the given list is considered.
}
\description{
Check if the provided object has nbElements elements
}
\keyword{internal}
