% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\docType{class}
\name{Label}
\alias{Label}
\title{Label}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Label
}
\description{
Label class associate font to a text
}
\keyword{datasets}
