% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\docType{data}
\name{LegendPositions}
\alias{LegendPositions}
\title{LegendPositions}
\format{
An object of class \code{list} of length 17.
}
\usage{
LegendPositions
}
\description{
List of all available legend positions
}
\keyword{internal}
