% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\docType{data}
\name{LegendTypes}
\alias{LegendTypes}
\title{LegendTypes}
\format{An object of class \code{list} of length 4.}
\usage{
LegendTypes
}
\description{
List of all available legend types
}
\keyword{datasets}
