% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-background.R
\name{LineElement}
\alias{LineElement}
\title{LineElement}
\description{
R6 class defining the properties of background line elements
}
\section{Super class}{
\code{\link[tlf:BackgroundElement]{tlf::BackgroundElement}} -> \code{LineElement}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-createPlotElement}{\code{LineElement$createPlotElement()}}
\item \href{#method-clone}{\code{LineElement$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="BackgroundElement" data-id="initialize">}\href{../../tlf/html/BackgroundElement.html#method-initialize}{\code{tlf::BackgroundElement$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createPlotElement"></a>}}
\if{latex}{\out{\hypertarget{method-createPlotElement}{}}}
\subsection{Method \code{createPlotElement()}}{
Create a \code{ggplot2::element_line} directly usable by \code{ggplot2::theme}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LineElement$createPlotElement(color = NULL, size = NULL, linetype = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color}}{character color of the frame of the background element}

\item{\code{size}}{character size of the frame of the background element}

\item{\code{linetype}}{character linetype of the frame of the background element}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{element_line} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LineElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
