% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enums.R
\docType{data}
\name{Linetypes}
\alias{Linetypes}
\title{Linetypes}
\format{
An object of class \code{list} of length 7.
}
\usage{
Linetypes
}
\description{
Enum of \code{ggplot2} linetypes
}
\examples{
# Plot available linetypes in enum Linetypes
addLine(
 # Leverage smart mapping by providing 
 # x, y and linetype as data.frame variable names
data = data.frame(
x = rep(1:2, each = length(Linetypes)),
y = rep(length(Linetypes):1, 2),
linetype = rep(factor(
names(tlf::Linetypes), 
levels = names(tlf::Linetypes)
), 2)
),
linetype = unlist(Linetypes),
size = 1,
color = "black"
)

}
\seealso{
Other enum helpers: 
\code{\link{AestheticFields}},
\code{\link{AestheticProperties}},
\code{\link{AestheticSelectionKeys}},
\code{\link{Alignments}},
\code{\link{AtomPlots}},
\code{\link{ColorMaps}},
\code{\link{DataMappings}},
\code{\link{DefaultDataMappingValues}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{LegendTypes}},
\code{\link{MoleculePlots}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{PlotConfigurations}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{TickLabelTransforms}},
\code{\link{VerticalJustification}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
