% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{loadThemeFromJson}
\alias{loadThemeFromJson}
\title{loadThemeFromJson}
\usage{
loadThemeFromJson(jsonFile)
}
\arguments{
\item{jsonFile}{path of json file}
}
\value{
A \code{Theme} object
}
\description{
Load theme object from json file.
A template of a json theme is available at system.file(package= "tlf", "theme-maker","theme-template.json")
}
