% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tlf.R
\name{loadTLFSettings}
\alias{loadTLFSettings}
\title{loadTLFSettings}
\usage{
loadTLFSettings(file)
}
\arguments{
\item{file}{\code{.RData} file containing the settings}
}
\description{
Load TLF global settings from a file
}
