% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-input.R
\name{mean-1.96sd}
\alias{mean-1.96sd}
\title{mean-1.96sd}
\usage{
`mean-1.96sd`(x)
}
\arguments{
\item{x}{Numeric values}
}
\value{
Numeric value corresponding to \code{mean(x)-1.96*sd(x)}
}
\description{
Calculate \verb{mean-1.96SD}
}
\examples{
# Calculate mean-1.96SD
`mean-1.96sd`(rnorm(1000))
 
}
\seealso{
Other stat functions: 
\code{\link{Percentile0\%}()},
\code{\link{Percentile100\%}()},
\code{\link{Percentile10\%}()},
\code{\link{Percentile15\%}()},
\code{\link{Percentile1\%}()},
\code{\link{Percentile2.5\%}()},
\code{\link{Percentile20\%}()},
\code{\link{Percentile25\%-1.5IQR}()},
\code{\link{Percentile25\%}()},
\code{\link{Percentile50\%}()},
\code{\link{Percentile5\%}()},
\code{\link{Percentile75\%+1.5IQR}()},
\code{\link{Percentile75\%}()},
\code{\link{Percentile80\%}()},
\code{\link{Percentile85\%}()},
\code{\link{Percentile90\%}()},
\code{\link{Percentile95\%}()},
\code{\link{Percentile97.5\%}()},
\code{\link{Percentile99\%}()},
\code{\link{mean+1.96sd}()},
\code{\link{mean+sd}()},
\code{\link{mean-sd}()},
\code{\link{median+1.5IQR}()},
\code{\link{median+IQR}()},
\code{\link{median-1.5IQR}()},
\code{\link{median-IQR}()}
}
\concept{stat functions}
