% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{mergeLegend}
\alias{mergeLegend}
\title{mergeLegend}
\usage{
mergeLegend(plotObject, newLabels, aestheticSelections)
}
\arguments{
\item{plotObject}{ggplot object}

\item{newLabels}{labels of caption to merge}

\item{aestheticSelections}{\code{ThemeAestheticSelections} object}
}
\description{
merge legend caption with existing legend caption
}
\keyword{internal}
