% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-helper.R
\name{metaDataHelper}
\alias{metaDataHelper}
\title{metaDataHelper}
\usage{
metaDataHelper(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
metaData with default OSP Suite units
}
\description{
Create a metaData object with information about units and dimensions of data
}
