% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{parseAddScatterLayer}
\alias{parseAddScatterLayer}
\title{parseAddScatterLayer}
\usage{
parseAddScatterLayer()
}
\value{
An expression to \code{eval()}
}
\description{
Create an expression that adds scatter plot layer
}
