% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utilities.R
\name{parseAddUncertaintyLayer}
\alias{parseAddUncertaintyLayer}
\title{parseAddUncertaintyLayer}
\usage{
parseAddUncertaintyLayer()
}
\value{
An expression to \code{eval()}
}
\description{
Create an expression that adds errorbars if uncertainty is included in dataMapping
}
\keyword{internal}
