% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{parseCheckPlotInputs}
\alias{parseCheckPlotInputs}
\title{parseCheckPlotInputs}
\usage{
parseCheckPlotInputs(plotType)
}
\arguments{
\item{plotType}{Type of plot (e.g. "PKRatio" for plotPKRatio)}
}
\value{
An expression to \code{eval()}
}
\description{
Create an expression that checks usual plot inputs
}
