% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{parseGetMapDataAndLabels}
\alias{parseGetMapDataAndLabels}
\title{parseGetMapDataAndLabels}
\usage{
parseGetMapDataAndLabels()
}
\value{
An expression to \code{eval()}
}
\description{
Create an expression that gets \code{mapData} and \code{mapLabels}
used by all the tlf plot functions
}
