% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utilities.R
\name{parseUpdateAestheticProperty}
\alias{parseUpdateAestheticProperty}
\title{parseUpdateAestheticProperty}
\usage{
parseUpdateAestheticProperty(aestheticProperty, plotConfigurationProperty)
}
\arguments{
\item{aestheticProperty}{Name of aesthetic property as defined in \code{AestheticProperties}}

\item{plotConfigurationProperty}{Name of PlotConfiguration property as defined in \code{AestheticProperties}}
}
\value{
An expression to \code{eval()}
}
\description{
Create an expression that updates the aesthetic properties based on
the information of \code{PlotConfiguration}
}
