% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-sub-classes.R
\name{PieChartPlotConfiguration}
\alias{PieChartPlotConfiguration}
\title{PieChartPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for pie charts
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{AxisConfiguration}},
\code{\link{BackgroundConfiguration}},
\code{\link{BackgroundElement}},
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{CumulativeTimeProfilePlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{ExportConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{LabelConfiguration}},
\code{\link{LegendConfiguration}},
\code{\link{LineElement}},
\code{\link{ObsVsPredPlotConfiguration}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{PlotGridConfiguration}},
\code{\link{QQPlotConfiguration}},
\code{\link{ResVsPredPlotConfiguration}},
\code{\link{ResVsTimePlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}},
\code{\link{XAxisConfiguration}},
\code{\link{YAxisConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{PieChartPlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{colorPalette}}{color palette property from \code{ggplot2}}

\item{\code{chartFont}}{\code{Font} object defining properties of text within pie chart}

\item{\code{start}}{Offset of starting point from 12 o'clock in radians.
Offset is applied clockwise or anticlockwise depending on value of direction}

\item{\code{clockwiseDirection}}{logical defining if values are displayed in clockwise order}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PieChartPlotConfiguration-new}{\code{PieChartPlotConfiguration$new()}}
\item \href{#method-PieChartPlotConfiguration-clone}{\code{PieChartPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PieChartPlotConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-PieChartPlotConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{PieChartPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PieChartPlotConfiguration$new(
  colorPalette = NULL,
  chartFont = NULL,
  start = 0,
  clockwiseDirection = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{colorPalette}}{color palette property from \code{ggplot2}}

\item{\code{chartFont}}{\code{Font} object defining properties of text within pie chart}

\item{\code{start}}{Offset of starting point from 12 o'clock in radians.
Offset is applied clockwise or anticlockwise depending on value of direction}

\item{\code{clockwiseDirection}}{logical defining if values are displayed in clockwise order}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PieChartPlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PieChartPlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PieChartPlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PieChartPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
