% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.R
\docType{data}
\name{pkRatioData}
\alias{pkRatioData}
\title{Test dataset of PK Ratios for 40 subjects.}
\format{A data frame with 40 rows and 11 variables:
\describe{
\item{IndividualID}
\item{Population}
\item{Gender}
\item{Age}
\item{Compound}
\item{Dose}
\item{Organ}
\item{Compartment}
\item{Simulated}
\item{Observed}
\item{Ratio}
}}
\usage{
pkRatioData
}
\description{
Test dataset of PK Ratios for 40 subjects.
}
\keyword{datasets}
