% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.R
\docType{data}
\name{pkRatioMetaData}
\alias{pkRatioMetaData}
\title{Test metaData for pkRatioData}
\format{A list of lists for the 11 variables of pkRatioData
Each variable includes:
\describe{
\item{unit}
\item{dimension}
}}
\usage{
pkRatioMetaData
}
\description{
Test metaData for pkRatioData
}
\keyword{datasets}
