% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxwhisker.R
\name{plotBoxWhisker}
\alias{plotBoxWhisker}
\title{plotBoxWhisker}
\usage{
plotBoxWhisker(
  data,
  metaData = NULL,
  outliers = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{A data.frame (or list of data.frames? TO BE DISCUSSED)
containing the data to be used for the plot}

\item{metaData}{A list of lists (structure TO BE DISCUSSED)
containing complementary information to data (e.g. unit)}

\item{outliers}{logical defining if outliers should be included in boxplot}

\item{dataMapping}{\code{BoxWhiskerDataMapping} object
mapping of x, y axes + mapping of colorGrouping, sizeGrouping, shapeGrouping}

\item{plotConfiguration}{A \code{BoxWhiskerConfiguration} object
Plot Configuration defining title, subtitle, xlabel, ylabel watermark, and legend}

\item{plotObject}{ggplot object, if null creates new plot, if not add time profile layers to ggplot}
}
\value{
a ggplot graphical object
}
\description{
Producing box-and-whisker plots
}
\references{
For examples, see:
\url{https://www.open-systems-pharmacology.org/TLF-Library/articles/box-whisker-vignette.html}
}
