% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\docType{class}
\name{PlotConfiguration}
\alias{PlotConfiguration}
\title{PlotConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PlotConfiguration
}
\description{
Class for Plot Configuration
}
\section{Fields}{

\describe{
\item{\code{legend}}{R6 class defining legendConfiguration}

\item{\code{xAxis}}{R6 class defining xAxisConfiguration}

\item{\code{yAxis}}{R6 class defining yAxisConfiguration}

\item{\code{background}}{R6 class defining backgroundConfiguration}

\item{\code{theme}}{R6 class defining theme aesthtic properties}

\item{\code{filename}}{Name of the saved plot}
}}

\section{Methods}{

\describe{
\item{new(title = NULL, subtitle = NULL, xlabel = NULL, ylabel = NULL, watermark = NULL, legendTitles = NULL, filename = "TestPlot.png",
legend = NULL, xAxis = NULL, yAxis = NULL, background = NULL, data = NULL, metaData = NULL, dataMapping = NULL, theme = tlfEnv$currentTheme, ...)}{
Initialize PlotConfiguration.}
\item{setPlotProperties(plotObject)}{Apply properties of plot labels.}
\item{setPlotBackground(plotObject)}{Apply background properties to plot.}
\item{savePlot(plotObject)}{Save ggplot as file.}
}
}

\keyword{datasets}
