% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ddiratio.R
\name{plotDDIRatio}
\alias{plotDDIRatio}
\title{plotDDIRatio}
\usage{
plotDDIRatio(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL,
  residualsVsObserved = NULL
)
}
\arguments{
\item{data}{data.frame containing the data to be used for the plot}

\item{metaData}{list of lists
containing complementary information to data (e.g. their unit and dimension).
This parameter is optional.}

\item{dataMapping}{A \code{DDIRatioDataMapping} object}

\item{plotConfiguration}{A \code{DDIRatioPlotConfiguration} object}

\item{plotObject}{\code{ggplot} graphical object to which the PK Ratio plot layer is added
This parameter is optional: the \code{tlf} library will initialize an empty plot if the parameter is NULL or not provided}

\item{residualsVsObserved}{if \code{TRUE}, lines of constant residuals are drawn horizontally, as for a residuals vs observations
DDI ratio plot.  Otherwise, these lines are drawn diagonally, as for a predictions vs observations DDI ratio plot.}
}
\value{
A \code{ggplot} graphical object
}
\description{
Add DDI Ratio plot layers to a \code{ggplot} graphical object.
Inclcuding Guest et al. limits, DDI Ratio limits as diagonal lines and
DDI Ratios as a scatter plot.
}
