% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-grid.R
\name{plotGrid}
\alias{plotGrid}
\title{Create a plot grid}
\usage{
plotGrid(plotGridConfiguration)
}
\arguments{
\item{plotGridConfiguration}{A \code{PlotGridConfiguration} object, which
is an \code{R6} class object that defines properties of a plot grid (like number
of rows, columns, labels, etc.).}
}
\description{
Create a plot grid using the \code{patchwork::wrap_plots()} function. The required
arguments are supplied through the \code{PlotGridConfiguration} object.
}
\examples{

library(ggplot2)
library(tlf)

# only `{tlf}` ---------------------

# plots to be arranged in a grid
set.seed(123)
ls_plots <- list(
  plotHistogram(x = rnorm(100)),
  plotHistogram(x = rnorm(100, mean = 3)),
  plotHistogram(x = rnorm(100, mean = 10))
)

# create an instance of plot configuration class
plotGridObj <- PlotGridConfiguration$new(plotList = ls_plots)

# specify further customizations for the plot grid
plotGridObj$title <- "my combined plot"
plotGridObj$subtitle <- "something clever"
plotGridObj$caption <- "my sources"
plotGridObj$nColumns <- 2L
plotGridObj$tagLevels <- "A"
plotGridObj$tagPrefix <- "Plot ("
plotGridObj$tagSuffix <- ")"
plotGridObj$tagColor <- "blue"
plotGridObj$tagSize <- 15
plotGridObj$tagAngle <- 45
plotGridObj$tagPosition <- TagPositions$top
plotGridObj$titleHorizontalJustification <- HorizontalJustification$middle
plotGridObj$subtitleHorizontalJustification <- HorizontalJustification$middle

# plot the grid
plotGrid(plotGridObj)

#  `{tlf}` and `{ggplot2}` ---------------------

# `{tlf}` plot
set.seed(123)
p1 <- plotBoxWhisker(mtcars,
  dataMapping = BoxWhiskerDataMapping$new(x = "am", y = "wt"), outliers = FALSE
)

# custom `{ggplot2}` plot
set.seed(123)
p2 <- ggplot(mtcars, aes(wt, mpg)) +
  geom_point()

# create an instance of plot configuration class
plotGridObj2 <- PlotGridConfiguration$new(list(p1, p2))

# specify further customizations for the plot grid
plotGridObj2$nColumns <- 1L
plotGridObj2$tagLevels <- "i"

# plot the grid
plotGrid(plotGridObj2)

}
\references{
For more, see: \url{https://patchwork.data-imaginist.com/articles/patchwork.html}
}
\seealso{
Other molecule plots: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotCumulativeTimeProfile}()},
\code{\link{plotDDIRatio}()},
\code{\link{plotHistogram}()},
\code{\link{plotObsVsPred}()},
\code{\link{plotObservedTimeProfile}()},
\code{\link{plotPKRatio}()},
\code{\link{plotPieChart}()},
\code{\link{plotQQ}()},
\code{\link{plotResVsPred}()},
\code{\link{plotResVsTime}()},
\code{\link{plotSimulatedTimeProfile}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotTornado}()}
}
\concept{molecule plots}
