% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-histogram.R
\name{plotHistogram}
\alias{plotHistogram}
\title{plotHistogram}
\usage{
plotHistogram(
  data,
  metaData = NULL,
  dataMapping = NULL,
  bins = NULL,
  stack = NULL,
  fitNormalDist = NULL,
  fitDensity = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{data.frames containing the data to be used for the plot}

\item{metaData}{list of lists (structure TO BE DISCUSSED)
containing complementary information to data (e.g. unit)}

\item{dataMapping}{R6 class HistogramDataMapping
mapping of which data to use for histogram}

\item{bins}{(optional)}

\item{plotConfiguration}{R6 class HistogramPlotConfiguration
Plot Configuration defining title, subtitle, xlabel, ylabel watermark, and legend}

\item{plotObject}{ggplot object, if null creates new plot, if not add time profile layers to ggplot}

\item{binWidth}{(optional)}

\item{verticalLineFunctions}{(optional)}

\item{verticalLineFunctionNames}{(optional)}
}
\value{
a ggplot graphical object
}
\description{
plotHistogram(data, metaData, dataMapping, plotConfiguration)
}
