% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-obs-vs-pred.R
\name{plotObsVsPred}
\alias{plotObsVsPred}
\title{plotObsVsPred}
\usage{
plotObsVsPred(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  foldDistance = NULL,
  smoother = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{dataMapping}{A \code{ObsVsPredDataMapping} object mapping \code{x}, \code{y} and aesthetic groups to their variable names of \code{data}.}

\item{plotConfiguration}{An optional \code{ObsVsPredConfiguration} object defining labels, grid, background and watermark.}

\item{foldDistance}{Numeric values of fold distance lines to display in log plots.
This argument is internally translated into \code{lines} field of \code{dataMapping}.
\strong{Caution}: this argument is meant for log scaled plots and since fold distance is a ratio it is expected positive.
In particular, line of identity corresponds to a \code{foldDistance} of \code{1}.}

\item{smoother}{Optional name of smoother function:
\itemize{
\item \code{"loess"} for loess regression
\item \code{"lm"} for linear regression
}}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layer}
}
\value{
A \code{ggplot} object
}
\description{
Producing observed vs predicted plots
}
\examples{
# Produce Obs vs Pred plot
obsVsPredData <- data.frame(x = c(1, 2, 1, 2, 3), y = c(5, 0.2, 2, 3, 4))

plotObsVsPred(data = obsVsPredData, dataMapping = ObsVsPredDataMapping$new(x = "x", y = "y"))

# Produce Obs vs Pred plot with linear regression
plotObsVsPred(
  data = obsVsPredData,
  dataMapping = ObsVsPredDataMapping$new(x = "x", y = "y"),
  smoother = "lm"
)

# Produce Obs vs Pred plot with user-defined fold distance lines
plotObsVsPred(
  data = obsVsPredData,
  dataMapping = ObsVsPredDataMapping$new(x = "x", y = "y"),
  plotConfiguration = ObsVsPredPlotConfiguration$new(
    xScale = Scaling$log, xLimits = c(0.05, 50),
    yScale = Scaling$log, yLimits = c(0.05, 50)
  ),
  foldDistance = c(1, 10)
)

}
\seealso{
Other molecule plots: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotDDIRatio}()},
\code{\link{plotHistogram}()},
\code{\link{plotPKRatio}()},
\code{\link{plotResVsPred}()},
\code{\link{plotResVsTime}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotTornado}()}
}
\concept{molecule plots}
