% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-piechart.R
\name{plotPieChart}
\alias{plotPieChart}
\title{plotPieChart}
\usage{
plotPieChart(
  data = NULL,
  metaData = NULL,
  dataMapping = NULL,
  colorPalette = NULL,
  start = NULL,
  clockwiseDirection = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{dataMapping}{A \code{PieChartDataMapping} object mapping \code{x} and \code{fill} aesthetic groups to their variable names of \code{data}.
Values mapped to \code{y} variable will be displayed as text within the pie chart}

\item{colorPalette}{color palette property from \code{ggplot2}}

\item{start}{Offset of starting point from 12 o'clock in radians.
Offset is applied clockwise or anticlockwise depending on value of direction}

\item{clockwiseDirection}{logical defining if values are displayed in clockwise order}

\item{plotConfiguration}{An optional \code{PieChartPlotConfiguration} object defining labels, grid, background and watermark.}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layer}
}
\value{
A \code{ggplot} object
}
\description{
Producing a Pie Chart
}
\examples{
# Data for the pie chart
values <- runif(5)
data <- data.frame(
values = values, 
text = paste0(round(100*values/sum(values)), "\%"),
legend = letters[1:5]
)

# Plot pie chart with its legend 
plotPieChart(
data = data,
dataMapping = PieChartDataMapping$new(x = "values", fill = "legend")
)

# Plot pie chart with text within pie
plotPieChart(
data = data,
dataMapping = PieChartDataMapping$new(x = "values", y = "text", fill = "legend")
)

# Reverse direction of pie chart
plotPieChart(
data = data,
dataMapping = PieChartDataMapping$new(x = "values", y = "text", fill = "legend"),
clockwiseDirection = FALSE
)

# Start first slice of pie at 90 degrees
plotPieChart(
data = data,
dataMapping = PieChartDataMapping$new(x = "values", y = "text", fill = "legend"),
start = pi/2
)

# Leverages ggplot color palettes
plotPieChart(
data = data,
dataMapping = PieChartDataMapping$new(x = "values", y = "text", fill = "legend"),
colorPalette = ColorPalettes$Set1
)

}
\seealso{
Other molecule plots: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotCumulativeTimeProfile}()},
\code{\link{plotDDIRatio}()},
\code{\link{plotGrid}()},
\code{\link{plotHistogram}()},
\code{\link{plotObsVsPred}()},
\code{\link{plotObservedTimeProfile}()},
\code{\link{plotPKRatio}()},
\code{\link{plotQQ}()},
\code{\link{plotResVsPred}()},
\code{\link{plotResVsTime}()},
\code{\link{plotSimulatedTimeProfile}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotTornado}()}
}
\concept{molecule plots}
