% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pkratio.R
\name{plotPKRatio}
\alias{plotPKRatio}
\title{Create a PK-Ratio plot}
\usage{
plotPKRatio(data, metaData, dataMapping = NULL,
  plotConfiguration = NULL)
}
\arguments{
\item{data}{TODO}

\item{metaData}{TODO}

\item{dataMapping}{TODO}

\item{plotConfiguration}{TODO}
}
\value{
a ggplot graphical object
}
\description{
Create a PK-Ratio plot
}
