% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-qq.R
\name{plotQQ}
\alias{plotQQ}
\title{plotQQ}
\usage{
plotQQ(
  data = NULL,
  metaData = NULL,
  y = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{y}{Numeric values to plot along the \code{y} axis. Only used instead of \code{data} if \code{data} is \code{NULL}.}

\item{dataMapping}{A \code{QQDataMapping} object mapping \code{y} and aesthetic groups to their variable names of \code{data}.}

\item{plotConfiguration}{An optional \code{QQPlotConfiguration} object defining labels, grid, background and watermark.}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layer}
}
\value{
A \code{ggplot} object
}
\description{
Producing Histograms
}
\examples{
# Produce QQ plot of normally distributed data
plotQQ(y = rnorm(100))

# Produce QQ plot of normally distributed data split by group
qqData <- data.frame(
  residuals = c(rnorm(100), rnorm(100)),
  groups = c(rep("Group A", 100), rep("Group B", 100))
)
plotQQ(
  data = qqData,
  dataMapping = QQDataMapping$new(y = "residuals", group = "groups")
)

}
\seealso{
Other molecule plots: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotCumulativeTimeProfile}()},
\code{\link{plotDDIRatio}()},
\code{\link{plotGrid}()},
\code{\link{plotHistogram}()},
\code{\link{plotObsVsPred}()},
\code{\link{plotObservedTimeProfile}()},
\code{\link{plotPKRatio}()},
\code{\link{plotResVsPred}()},
\code{\link{plotResVsTime}()},
\code{\link{plotSimulatedTimeProfile}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotTornado}()}
}
\concept{molecule plots}
