% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-tornado.R
\name{plotTornado}
\alias{plotTornado}
\title{plotTornado}
\usage{
plotTornado(
  data = NULL,
  metaData = NULL,
  x = NULL,
  y = NULL,
  sorted = NULL,
  colorPalette = NULL,
  bar = TRUE,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{data.frame containing the data to be used for the plot}

\item{metaData}{list of lists
containing complementary information to data (e.g. their unit and dimension).
This parameter is optional.}

\item{x}{Mapping for x (values of tornado plot)
If \code{data} is NULL or not input, \code{x} numeric values will be used as is for the plot.}

\item{y}{Mapping for y (labels of tornado plot)
If \code{data} is NULL or not input, \code{y} character values will be used as is for the plot.}

\item{sorted}{logical indicating if values should be sorted if \code{dataMapping} is not input.
By default, values are sorted by their absolute values}

\item{colorPalette}{Define a \code{ggplot2} colorPalette (e.g \code{colorPalette}="Spectral")}

\item{bar}{logical setting tornado as bar plot}

\item{dataMapping}{\code{TornadoDataMapping} class or subclass mapping x and y variables to \code{data} variable names.
\code{dataMapping} provides also the values of the PK Ratio limits plotted as horizontal lines.
This parameter is optional: the \code{tlf} library provides a smart mapping if only \code{data} is provided
and default values of the PK Ratio limits.}

\item{plotConfiguration}{\code{TornadoPlotConfiguration} class or subclass defining labels, grid, background and watermark
This parameter is optional: the \code{tlf} library provides a default configuration according to the current theme}

\item{plotObject}{\code{ggplot} graphical object to which the tornado plot layer is added
This parameter is optional: the \code{tlf} library will initialize an empty plot if the parameter is NULL or not provided}
}
\value{
A \code{ggplot} graphical object
}
\description{
Add tornado plot layers to a \code{ggplot} graphical object.
Tornado limits are plotted as vertical lines.
Tornado values are plotted as horizontal bars or point according to option \code{bar}
}
