% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-sub-classes.R
\name{QQPlotConfiguration}
\alias{QQPlotConfiguration}
\title{QQPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for Quantile-Quantile plots
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{CumulativeTimeProfilePlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{ObsVsPredPlotConfiguration}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{tlf::PlotConfiguration} -> \code{QQPlotConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QQPlotConfiguration-new}{\code{QQPlotConfiguration$new()}}
\item \href{#method-QQPlotConfiguration-clone}{\code{QQPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QQPlotConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-QQPlotConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{QQPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QQPlotConfiguration$new(xlabel = "Standard Normal Quantiles", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xlabel}}{QQ-plot default display is "Standard Normal Quantiles"}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{QQPlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QQPlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QQPlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QQPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
