% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-xy.R
\docType{class}
\name{RangeDataMapping}
\alias{RangeDataMapping}
\title{RangeDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RangeDataMapping
}
\description{
Abstract class for X Ymin Ymax Mapping
}
\section{Fields}{

\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{ymin}}{Name of ymin variable to map}

\item{\code{ymax}}{Name of ymax variable to map}

\item{\code{data}}{Dataset after mapping}
}}

\section{Methods}{

\describe{
\item{new(x, ymin, ymax)}{Initialize RangeDataMapping.}
\item{checkMapData(data, metaData = NULL)}{Check data mapping is correct. Create output data.frame with map data only.}
}
}

\keyword{datasets}
