% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tlf.R
\name{resetTLFSettingsToDefault}
\alias{resetTLFSettingsToDefault}
\title{resetTLFSettingsToDefault}
\usage{
resetTLFSettingsToDefault()
}
\description{
Reset the global settings stored in \code{tlfEnv} to default values defined by the package.
}
