% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs-vs-pred-datamapping.R
\name{ResVsPredDataMapping}
\alias{ResVsPredDataMapping}
\title{ResVsPredDataMapping}
\description{
Class for mapping variables in residuals vs predictions/time plot
}
\seealso{
Other DataMapping classes: 
\code{\link{BoxWhiskerDataMapping}},
\code{\link{DDIRatioDataMapping}},
\code{\link{GroupMapping}},
\code{\link{Grouping}},
\code{\link{HistogramDataMapping}},
\code{\link{ObsVsPredDataMapping}},
\code{\link{ObservedDataMapping}},
\code{\link{PKRatioDataMapping}},
\code{\link{RangeDataMapping}},
\code{\link{TimeProfileDataMapping}},
\code{\link{TornadoDataMapping}},
\code{\link{XYDataMapping}},
\code{\link{XYGDataMapping}}
}
\concept{DataMapping classes}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{\link[tlf:PKRatioDataMapping]{tlf::PKRatioDataMapping}} -> \code{\link[tlf:ObsVsPredDataMapping]{tlf::ObsVsPredDataMapping}} -> \code{ResVsPredDataMapping}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{ResVsPredDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="PKRatioDataMapping" data-id="checkMapData">}\href{../../tlf/html/PKRatioDataMapping.html#method-checkMapData}{\code{tlf::PKRatioDataMapping$checkMapData()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="ObsVsPredDataMapping" data-id="initialize">}\href{../../tlf/html/ObsVsPredDataMapping.html#method-initialize}{\code{tlf::ObsVsPredDataMapping$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResVsPredDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
