% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{runPlotMaker}
\alias{runPlotMaker}
\title{runPlotMaker}
\usage{
runPlotMaker()
}
\description{
Run shiny app that allows to easily make plots from user interface.
}
\references{
For tutorial, see:
\url{https://www.open-systems-pharmacology.org/TLF-Library/articles/plot-maker.html}
}
\seealso{
Other shiny apps: 
\code{\link{runThemeMaker}()}
}
\concept{shiny apps}
