% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{runThemeMaker}
\alias{runThemeMaker}
\title{runThemeMaker}
\usage{
runThemeMaker()
}
\description{
Run shiny app that allows easy setting of Theme objects.
Theme objects drive default properties of plots
}
\references{
For tutorial, see:
\link{https://www.open-systems-pharmacology.org/TLF-Library/articles/theme-maker.html}
}
\seealso{
Other shiny apps: 
\code{\link{runPlotMaker}()}
}
\concept{shiny apps}
