% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-save.R
\docType{class}
\name{SaveConfiguration}
\alias{SaveConfiguration}
\title{saveConfiguration}
\usage{
SaveConfiguration
}
\description{
Configuration to save a gglot
}
\section{Fields}{

\describe{
\item{\code{filename}}{character}

\item{\code{width}}{numeric}

\item{\code{height}}{numeric}

\item{\code{units}}{character}
}}

\section{Methods}{

\describe{
\item{new()}{Initialize SaveConfiguration}
\item{print()}{Show SaveConfiguration properties}
\item{savePlot(plotObject)}{Save ggplot object into filename file}
}
}

\keyword{datasets}
