% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{saveThemeToJson}
\alias{saveThemeToJson}
\title{saveThemeToJson}
\usage{
saveThemeToJson(jsonFile, theme = NULL)
}
\arguments{
\item{jsonFile}{path of json file}

\item{theme}{\code{Theme} object path of json file}
}
\description{
Save theme object to a json file.
}
