% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tlf.R
\name{saveTLFSettings}
\alias{saveTLFSettings}
\title{saveTLFSettings
Save the current TLF global settings in a \code{.RData} file}
\usage{
saveTLFSettings(file)
}
\arguments{
\item{file}{\code{.RData} file containing the settings}
}
\description{
saveTLFSettings
Save the current TLF global settings in a \code{.RData} file
}
