% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\docType{data}
\name{Scaling}
\alias{Scaling}
\title{Scaling}
\format{
An object of class \code{list} of length 7.
}
\usage{
Scaling
}
\description{
Pre-defined transformation of axes
Not that built-in transformations from \code{ggplot2} includes more transformations
}
\examples{
# Continuous linear scale
Scaling$lin
# Continuous log10 scale
Scaling$log
# Discrete scale
Scaling$discrete
# Reverse continuous linear scale
Scaling$reverse
# Date scale
Scaling$date
}
\keyword{datasets}
