% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-background.R
\name{setBackground}
\alias{setBackground}
\title{setBackground}
\usage{
setBackground(
  plotObject,
  fill = NULL,
  color = NULL,
  linetype = NULL,
  size = NULL
)
}
\arguments{
\item{plotObject}{ggplot object to set}

\item{fill}{character color fill of the background}

\item{color}{character color of the background frame}

\item{linetype}{character linetype of the background frame}

\item{size}{numeric size of the background frame}
}
\value{
ggplot object
}
\description{
Set background properties on a ggplot object
}
