% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setCaptionColor}
\alias{setCaptionColor}
\title{setCaptionColor}
\usage{
setCaptionColor(plotObject, color, name = NULL)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{color}{colors of the data in plot and legend caption}

\item{name}{reference names in caption$name identifying the legend caption lines}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the colors of the data in plot and legend caption
}
