% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setCaptionLabels}
\alias{setCaptionLabels}
\title{setCaptionLabel}
\usage{
setCaptionLabels(plotObject, label, name = NULL)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{label}{new labels of the legend caption}

\item{name}{reference names in caption$name identifying the legend caption lines}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the legend caption labels
}
