% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setCaptionLinetype}
\alias{setCaptionLinetype}
\title{setCaptionLinetype}
\usage{
setCaptionLinetype(plotObject, linetype, name = NULL)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{linetype}{linetypes of the data in plot and legend caption}

\item{name}{reference names in caption$name identifying the legend caption lines}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the linetypes of the data in plot and legend caption
}
