% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setCaptionOrder}
\alias{setCaptionOrder}
\title{setCaptionOrder}
\usage{
setCaptionOrder(plotObject, order, name = NULL)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{order}{numeric order of the legend caption labels}

\item{name}{reference names in caption$name identifying the legend caption lines}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the order of the legend caption labels
}
