% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setCaptionShape}
\alias{setCaptionShape}
\title{setCaptionShape}
\usage{
setCaptionShape(plotObject, shape, name = NULL)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{shape}{shapes of the data in plot and legend caption}

\item{name}{reference names in caption$name identifying the legend caption lines}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the shapes of the data in plot and legend caption
}
