% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setCaptionSize}
\alias{setCaptionSize}
\title{setCaptionSize}
\usage{
setCaptionSize(plotObject, size, name = NULL)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{size}{sizes of the data in plot and legend caption}

\item{name}{reference names in caption$name identifying the legend caption lines}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the sizes of the data in plot and legend caption
}
