% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setCaptionVisibility}
\alias{setCaptionVisibility}
\title{setCaptionVisibility}
\usage{
setCaptionVisibility(plotObject, visibility, name = NULL)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{visibility}{logical visibility of the legend caption labels}

\item{name}{reference names in caption$name identifying the legend caption lines}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the visibility of the legend caption labels
}
