% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultAggregationBins}
\alias{setDefaultAggregationBins}
\title{setDefaultAggregationBins}
\usage{
setDefaultAggregationBins(bins = NULL)
}
\arguments{
\item{bins}{number of bins if value, edges if vector or binning function if function}
}
\description{
Set default aggregation bins of tlf environment
}
