% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tlf.R
\name{setDefaultAggregationFunctions}
\alias{setDefaultAggregationFunctions}
\title{setDefaultAggregationFunctions}
\usage{
setDefaultAggregationFunctions(y = NULL, ymin = NULL, ymax = NULL)
}
\arguments{
\item{y}{function or its name as median aggregation}

\item{ymin}{function or its name as min aggregation}

\item{ymax}{function or its name as max aggregation}
}
\description{
Set default aggregation functions of tlf environment
}
