% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tlf.R
\name{setDefaultAggregationLabels}
\alias{setDefaultAggregationLabels}
\title{setDefaultAggregationLabels}
\usage{
setDefaultAggregationLabels(y = NULL, range = NULL)
}
\arguments{
\item{y}{label for median aggregation}

\item{range}{label for range aggregation}
}
\description{
Set default aggregation labels of tlf environment
}
