% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultDefaultAlphaRatio}
\alias{setDefaultDefaultAlphaRatio}
\title{setDefaultAlphaRatio}
\usage{
setDefaultDefaultAlphaRatio(alphaRatio)
}
\arguments{
\item{alphaRatio}{alpha ratio to set as default. Must be between 0 and 1.}
}
\description{
Set the default alpha (transparency) difference ratio between
points above and bellow lloq
}
