% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultErrorbarCapExtent}
\alias{setDefaultErrorbarCapExtent}
\title{setDefaultErrorbarCapExtent}
\usage{
setDefaultErrorbarCapExtent(extent)
}
\arguments{
\item{extent}{A numeric defining the extent of the error bar caps}
}
\description{
Set default cap extent of error bars
Caution the value corresponds to the ratio of the mean spacing between plotted error bars.
For instance, an \code{extent} of \code{1} will fill the caps until the next error bar
}
