% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultErrorbarCapSize}
\alias{setDefaultErrorbarCapSize}
\title{setDefaultErrorbarCapSize}
\usage{
setDefaultErrorbarCapSize(size)
}
\arguments{
\item{size}{A numeric defining the size of the error bar caps in pts}
}
\description{
Set default cap size of error bars
}
