% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultErrorbarCapWidth}
\alias{setDefaultErrorbarCapWidth}
\title{setDefaultErrorbarCapWidth}
\usage{
setDefaultErrorbarCapWidth(width)
}
\arguments{
\item{width}{A numeric defining the width of the error bars}
}
\description{
Set default cap width of error bars
}
