% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultExportName}
\alias{setDefaultExportName}
\title{setDefaultExportName}
\usage{
setDefaultExportName(name)
}
\arguments{
\item{name}{base file name of the exported plots}
}
\description{
Set default save name property of tlf environment
}
