% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultExportParameters}
\alias{setDefaultExportParameters}
\title{setDefaultSaveParameters}
\usage{
setDefaultExportParameters(format, width = NULL, height = NULL, units = NULL)
}
\arguments{
\item{format}{file format of the exported plots}

\item{width}{plot width in \code{unit}}

\item{height}{plot height in \code{unit}}

\item{units}{units of \code{width} and \code{height}}
}
\description{
Set default save properties of tlf environment
}
