% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultLegendPosition}
\alias{setDefaultLegendPosition}
\title{setDefaultLegendPosition}
\usage{
setDefaultLegendPosition(position)
}
\arguments{
\item{position}{legend position.
Use Enum \code{LegendPositions} to assess available legend positions}
}
\description{
Set default legend position of tlf environment
}
