% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultLLOQLinetype}
\alias{setDefaultLLOQLinetype}
\title{setDefaultLLOQLinetype}
\usage{
setDefaultLLOQLinetype(linetype)
}
\arguments{
\item{linetype}{default linetype for lloq lines}
}
\description{
Set default cap linetype for lloq lines
}
