% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultLogTicks}
\alias{setDefaultLogTicks}
\title{setDefaultLogTicks}
\usage{
setDefaultLogTicks(ticks)
}
\arguments{
\item{ticks}{numeric values where ticks are placed.
Ensure that the values are positive (they are meant for log scale)}
}
\description{
Set default values for log ticks
}
