% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tlf.R
\name{setDefaultMaxCharacterWidth}
\alias{setDefaultMaxCharacterWidth}
\title{setDefaultMaxCharacterWidth}
\usage{
setDefaultMaxCharacterWidth(maxCharacterWidth)
}
\arguments{
\item{maxCharacterWidth}{the maximum number of characters per row}
}
\description{
Set the  maximum number of characters per row for axis ticks labels and legend.
Long strings will be wraped on spaces or non-word characters.
}
