% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-font.R
\name{setFontProperties}
\alias{setFontProperties}
\title{setFontProperties}
\usage{
setFontProperties(plotHandle, titleFont = NULL, subtitleFont = NULL,
  xAxisFont = NULL, yAxisFont = NULL, legendFont = NULL)
}
\arguments{
\item{plotHandle}{ggplot object}

\item{titleFont}{Font Class for title}

\item{subtitleFont}{Font Class for subtitle}

\item{xAxisFont}{Font Class for xaxis and ticks}

\item{yAxisFont}{Font Class for yaxis and ticks}

\item{legendFont}{Font Class for legend}
}
\value{
plotHandle ggplot object with updated fonts
}
\description{
setFontProperties set Font Properties on a ggplot object
}
\examples{
p <- ggplot2::ggplot() + ggplot2::labs(title = "Title")
newFont <- Font$new(color = "blue", size = 20)
p <- setFontProperties(plotHandle = p, titleFont = newFont)
}
