% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegend}
\alias{setLegend}
\title{setLegend}
\usage{
setLegend(plotObject, position, title = NULL, caption = NULL)
}
\arguments{
\item{plotObject}{Graphical object created from ggplot}

\item{position}{legend position.
Use enum \code{LegendPositions} to access the list of legend positions.}

\item{title}{title of legend}

\item{caption}{data.frame containing the caption properties of the legend}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set legend position, title and/or caption
}
