% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegendCaption}
\alias{setLegendCaption}
\title{setLegendCaption}
\usage{
setLegendCaption(plotObject, caption = NULL)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{caption}{data.frame containing the caption properties of the legend}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the legend caption
}
