% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-utils.R, R/utilities-legend.R
\name{setLegendPosition}
\alias{setLegendPosition}
\title{SetLegendPosition}
\usage{
setLegendPosition(plotHandle,
  legendPosition = legendPositions$outsideRight, Subplot.Index = NA)

setLegendPosition(plotHandle,
  legendPosition = legendPositions$outsideRight, Subplot.Index = NA)
}
\arguments{
\item{plotHandle}{Graphical object created from ggplot}

\item{legendPosition}{element of legendPositions list}

\item{Subplot.Index}{(optional) OPTION NOT IMPLEMENTED YET}

\item{Location}{Inside/Oustide/None}

\item{X.Location}{NA/Left/Center/Right}

\item{Y.Location}{NA/Top/Center/Bottom}

\item{plotHandle}{Graphical object created from ggplot}

\item{Subplot.Index}{(optional) OPTION NOT IMPLEMENTED YET}
}
\value{
Current plot with Legend Layer

Current plot with Legend Layer
}
\description{
SetLegendPosition set a Legend Position in a plot (case insensitive)
Default legend setting is within plot on the top right corner

SetLegendPosition set a Legend Position as defined by legendPosition key
Default legend setting is within plot on the top right corner
}
