% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setMaxCharacterWidth}
\alias{setMaxCharacterWidth}
\title{setMaxCharacterWidth}
\usage{
setMaxCharacterWidth(maxCharacterWidth)
}
\arguments{
\item{maxCharacterWidth}{the maximum number of characters per row}
}
\description{
Set the  maximum number of characters per row for axis ticks labels and legend.
Long strings will be wraped on spaces or non-word characters.
}
