% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-export.R
\name{setPlotExportDimensions}
\alias{setPlotExportDimensions}
\title{setPlotExportDimensions}
\usage{
setPlotExportDimensions(
  plotObject,
  width = NULL,
  height = NULL,
  units = NULL,
  dpi = NULL
)
}
\arguments{
\item{plotObject}{Graphical object created from ggplot}

\item{width}{plot width in \code{unit}}

\item{height}{plot height in \code{unit}}

\item{units}{units of \code{width} and \code{height}}

\item{dpi}{numeric value defining plot resolution (dots per inch)}
}
\value{
ggplot object with updated labels
}
\description{
Set plot export properties
}
