% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-export.R
\name{setPlotExportFormat}
\alias{setPlotExportFormat}
\title{setPlotExportFormat}
\usage{
setPlotExportFormat(plotObject, format = NULL)
}
\arguments{
\item{plotObject}{Graphical object created from ggplot}

\item{format}{file format of the exported plot}
}
\value{
ggplot object with updated labels
}
\description{
Set plot export properties
}
