% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-export.R
\name{setPlotExportSize}
\alias{setPlotExportSize}
\title{setPlotExportSize}
\usage{
setPlotExportSize(plotObject, width = NULL, height = NULL, units = NULL)
}
\arguments{
\item{plotObject}{Graphical object created from ggplot}

\item{width}{plot width in \code{unit}}

\item{height}{plot height in \code{unit}}

\item{units}{units of \code{width} and \code{height}}
}
\value{
ggplot object with updated labels
}
\description{
Set plot export properties
}
