% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-background.R
\name{setWatermark}
\alias{setWatermark}
\title{setWatermark}
\usage{
setWatermark(
  plotObject,
  watermark = NULL,
  color = NULL,
  size = NULL,
  angle = NULL,
  alpha = NULL
)
}
\arguments{
\item{plotObject}{ggplot object to which the watermark is set}

\item{watermark}{Character or \code{Label} object corresponding to the watermark text
(and its font properties if \code{Label})}

\item{color}{Color of the watermark label.}

\item{size}{Size of the watermark label.}

\item{angle}{Angle of the watermark label (in degree).}

\item{alpha}{Transparency of the watermark label.
Alpha is a numeric between 0 and 1: 0 label is totally transparent, 1 label is totally opaque.}
}
\value{
ggplot object to which the watermark is added.
}
\description{
setWatermark
}
