% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{setXAxis}
\alias{setXAxis}
\title{setXAxis}
\usage{
setXAxis(
  plotObject,
  scale = NULL,
  limits = NULL,
  ticks = NULL,
  ticklabels = NULL,
  font = NULL
)
}
\arguments{
\item{plotObject}{ggplot object to set X-axis properties}

\item{scale}{Scale of X-axis. Use enum \code{Scaling} to access names of scales.}

\item{limits}{X-axis limits}

\item{ticks}{X-axis ticks}

\item{ticklabels}{X-axis ticklabels}

\item{font}{\code{Font} object  defining font of ticklabels}
}
\value{
ggplot object with updated X-axis
}
\description{
Set X-axis properties on a ggplot object
}
\examples{
myPlot <- addLine(x = c(1, 2, 3), y = c(10, 50, 100))

# Set x-axis in log scale
setXAxis(myPlot, scale = Scaling$log)

# Set x-axis ticklabels to Greek letters
setXAxis(myPlot, ticks = c(1, 2, 3), ticklabels = parse(text = c("alpha", "beta", "gamma")))

# Set x-axis limits
setXAxis(myPlot, limits = c(1, 2.5))

# Set x-axis fonts
setXAxis(myPlot, font = Font$new(color = "blue", size = 14))
}
