% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-background.R
\name{setXGrid}
\alias{setXGrid}
\title{setXGrid}
\usage{
setXGrid(plotObject, color = NULL, linetype = NULL, size = NULL)
}
\arguments{
\item{plotObject}{ggplot object to set}

\item{color}{character color of the grid}

\item{linetype}{character linetype of the grid. Use "blank" to remove grid.}

\item{size}{numeric size of the grid lines}
}
\value{
ggplot object
}
\description{
Set x-grid properties on a ggplot object
}
