% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{setY2Axis}
\alias{setY2Axis}
\title{setY2Axis}
\usage{
setY2Axis(
  plotObject,
  scale = NULL,
  limits = NULL,
  ticks = NULL,
  ticklabels = NULL,
  minorTicks = NULL,
  font = NULL,
  expand = NULL
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object to set X-axis properties}

\item{scale}{Scale of axis. Use enum \code{Scaling} to access names of scales.}

\item{limits}{Optional numeric values of axis limits}

\item{ticks}{Optional values or function for axis ticks}

\item{ticklabels}{Optional values or function for axis ticklabels}

\item{minorTicks}{Optional values or function for axis minor ticks}

\item{font}{A \code{Font} object  defining font of ticklabels}

\item{expand}{Logical defining if data is expanded until axis}
}
\value{
A \code{ggplot} object
}
\description{
Set right Y-axis properties of a \code{ggplot} object
}
