% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{setYAxis}
\alias{setYAxis}
\title{setYAxis}
\usage{
setYAxis(
  plotObject,
  scale = NULL,
  limits = NULL,
  ticks = NULL,
  ticklabels = NULL
)
}
\arguments{
\item{plotObject}{ggplot object to set Y-axis properties}

\item{scale}{Scale of Y-axis. Use enum \code{Scaling} to access names of scales.}

\item{limits}{Y-axis limits}

\item{ticks}{Y-axis ticks}

\item{ticklabels}{Y-axis ticklabels}
}
\value{
ggplot object with updated Y-axis
}
\description{
Set Y-axis properties on a ggplot object
}
