% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enums.R
\docType{data}
\name{Shapes}
\alias{Shapes}
\title{Shapes}
\format{
An object of class \code{list} of length 60.
}
\usage{
Shapes
}
\description{
List of some \code{ggplot2} shapes.
All the shapes of this enum correspond to one character value
due to compatibility issues when providing shape names
as described in the articles below.
https://stackoverflow.com/questions/67859447/error-when-combining-unicode-shape-with-legal-ggplot2-shapes
https://ggplot2.tidyverse.org/articles/ggplot2-specs.html#point

Depending on the packages used for displaying/exporting figures,
some of the shapes may not appear.
}
\examples{
# Plot some available shapes in enum Shapes
addScatter(
 # Leverage smart mapping by providing 
 # x, y and shape as data.frame variable names
data = data.frame(
x = rep(1, 12),
y = 12:1,
shape = factor(
names(Shapes)[1:12], 
levels = names(Shapes)[1:12]
)),
shape = unlist(Shapes[1:12]),
size = 3,
color = "black"
)

}
\seealso{
Other enum helpers: 
\code{\link{AestheticFields}},
\code{\link{AestheticProperties}},
\code{\link{AestheticSelectionKeys}},
\code{\link{Alignments}},
\code{\link{AtomPlots}},
\code{\link{ColorMaps}},
\code{\link{DataMappings}},
\code{\link{DefaultDataMappingValues}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{LegendTypes}},
\code{\link{Linetypes}},
\code{\link{MoleculePlots}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{PlotConfigurations}},
\code{\link{Scaling}},
\code{\link{TagPositions}},
\code{\link{TickLabelTransforms}},
\code{\link{VerticalJustification}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
