% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\docType{data}
\name{TagPositions}
\alias{TagPositions}
\title{TagPositions}
\format{
An object of class \code{list} of length 8.
}
\usage{
TagPositions
}
\description{
List of all available tag positions in a plot grid.
}
\seealso{
Other enum helpers: 
\code{\link{AestheticProperties}},
\code{\link{Alignments}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{Linetypes}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{VerticalJustification}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
