% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{Theme}
\alias{Theme}
\title{Theme}
\description{
R6 class defining theme properties
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fonts}}{\code{ThemeFont} object}

\item{\code{background}}{\code{ThemeBackground} object}

\item{\code{aestheticMaps}}{\code{ThemeAestheticMaps} object}

\item{\code{plotConfigurations}}{\code{ThemePlotConfiguration} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Theme-new}{\code{Theme$new()}}
\item \href{#method-Theme-save}{\code{Theme$save()}}
\item \href{#method-Theme-clone}{\code{Theme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Theme-new"></a>}}
\if{latex}{\out{\hypertarget{method-Theme-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Theme} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$new(
  fonts = NULL,
  background = NULL,
  aestheticMaps = NULL,
  plotConfigurations = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fonts}}{\code{ThemeFont} object}

\item{\code{background}}{\code{ThemeBackground} object}

\item{\code{aestheticMaps}}{\code{ThemeAestheticMaps} object}

\item{\code{plotConfigurations}}{\code{ThemePlotConfiguration} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Theme} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Theme-save"></a>}}
\if{latex}{\out{\hypertarget{method-Theme-save}{}}}
\subsection{Method \code{save()}}{
Save \code{Theme} as a json file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$save(jsonFile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{jsonFile}}{name of json file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Theme-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Theme-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
