% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ThemeAesProperties}
\alias{ThemeAesProperties}
\title{ThemeAesProperties}
\description{
R6 class defining theme aesthetic properties plots
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{color}}{character vector of color properties}

\item{\code{shape}}{character vector of shape properties}

\item{\code{size}}{numeric vector of size properties}

\item{\code{fill}}{character vector of fill properties}

\item{\code{linetype}}{character vector of linetype properties}

\item{\code{alpha}}{numeric vector of alpha properties}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ThemeAesProperties$new()}}
\item \href{#method-clone}{\code{ThemeAesProperties$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{ThemeAesProperties} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeAesProperties$new(
  aesProperties = tlfEnvThemesProperties$default$aesProperties,
  color = NULL,
  shape = NULL,
  size = NULL,
  fill = NULL,
  linetype = NULL,
  alpha = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aesProperties}}{list of aesthetic properties}

\item{\code{color}}{character vector of color properties}

\item{\code{shape}}{character vector of shape properties}

\item{\code{size}}{numeric vector of size properties}

\item{\code{fill}}{character vector of fill properties}

\item{\code{linetype}}{character vector of linetype properties}

\item{\code{alpha}}{numeric vector of alpha properties}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ThemeAesProperties} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeAesProperties$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
