% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ThemeAestheticMaps}
\alias{ThemeAestheticMaps}
\title{ThemeAestheticMaps}
\description{
R6 class defining theme aesthetic maps
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{color}}{color map as character or numeric vector}

\item{\code{fill}}{fill map as character or numeric vector}

\item{\code{size}}{size map as numeric vector}

\item{\code{shape}}{shape map as numeric vector}

\item{\code{linetype}}{linetype as character vector}

\item{\code{alpha}}{map as numeric vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ThemeAestheticMaps$new()}}
\item \href{#method-toJson}{\code{ThemeAestheticMaps$toJson()}}
\item \href{#method-clone}{\code{ThemeAestheticMaps$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ThemeAestheticMaps} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeAestheticMaps$new(
  color = NULL,
  fill = NULL,
  shape = NULL,
  size = NULL,
  linetype = NULL,
  alpha = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color}}{color map as list, character or numeric vector}

\item{\code{fill}}{fill map as list, character or numeric vector}

\item{\code{shape}}{shape map as list, character or numeric vector}

\item{\code{size}}{size  map as list, character or numeric vector}

\item{\code{linetype}}{linetype map as list, character or numeric vector}

\item{\code{alpha}}{alpha map as list, character or numeric vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ThemeAestheticMaps} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJson"></a>}}
\if{latex}{\out{\hypertarget{method-toJson}{}}}
\subsection{Method \code{toJson()}}{
Translate object into a json list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeAestheticMaps$toJson()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list that can be saved into a json file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeAestheticMaps$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
