% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ThemeAestheticSelections}
\alias{ThemeAestheticSelections}
\title{ThemeAestheticSelections}
\description{
R6 class defining how plot configurations will use aesthetic maps
}
\section{Super class}{
\code{\link[tlf:ThemeAestheticMaps]{tlf::ThemeAestheticMaps}} -> \code{ThemeAestheticSelections}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ThemeAestheticSelections-new}{\code{ThemeAestheticSelections$new()}}
\item \href{#method-ThemeAestheticSelections-toJson}{\code{ThemeAestheticSelections$toJson()}}
\item \href{#method-ThemeAestheticSelections-clone}{\code{ThemeAestheticSelections$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeAestheticSelections-new"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeAestheticSelections-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ThemeAestheticSelections} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeAestheticSelections$new(
  color = NULL,
  fill = NULL,
  shape = NULL,
  size = NULL,
  linetype = NULL,
  alpha = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color}}{selection key or values for choice of color}

\item{\code{fill}}{selection key or values for choice of fill}

\item{\code{shape}}{selection key or values for choice of shape}

\item{\code{size}}{selection key or values for choice of size}

\item{\code{linetype}}{selection key or values for choice of linetype}

\item{\code{alpha}}{selection key or values for choice of alpha}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ThemeAestheticSelections} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeAestheticSelections-toJson"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeAestheticSelections-toJson}{}}}
\subsection{Method \code{toJson()}}{
Translate object into a json list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeAestheticSelections$toJson()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list that can be saved into a json file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeAestheticSelections-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeAestheticSelections-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeAestheticSelections$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
