% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ThemeBackground}
\alias{ThemeBackground}
\title{ThemeBackground}
\description{
R6 class defining theme background properties
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{watermark}}{character defining content of watermark}

\item{\code{legendPosition}}{character defining where legend should usually be placed}

\item{\code{legendTitle}}{character defining the content of legend title}

\item{\code{plot}}{\code{BackgroundElement} object for plot area properties (outside of panel)}

\item{\code{panel}}{\code{BackgroundElement} object for plot area properties (inside of panel)}

\item{\code{xAxis}}{\code{BackgroundElement} object for x axis properties}

\item{\code{yAxis}}{\code{BackgroundElement} object for y axis properties}

\item{\code{xGrid}}{\code{BackgroundElement} object for x grid properties}

\item{\code{yGrid}}{\code{BackgroundElement} object for y grid properties}

\item{\code{legend}}{\code{BackgroundElement} object for legend area properties}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ThemeBackground-new}{\code{ThemeBackground$new()}}
\item \href{#method-ThemeBackground-toJson}{\code{ThemeBackground$toJson()}}
\item \href{#method-ThemeBackground-clone}{\code{ThemeBackground$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeBackground-new"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeBackground-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ThemeBackground} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeBackground$new(
  watermark = NULL,
  legendPosition = NULL,
  legendTitle = NULL,
  plot = NULL,
  panel = NULL,
  xAxis = NULL,
  yAxis = NULL,
  xGrid = NULL,
  yGrid = NULL,
  legend = NULL,
  baseFill = "white",
  baseColor = "black",
  baseSize = 0.5,
  baseLinetype = "solid"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{watermark}}{character defining content of watermark}

\item{\code{legendPosition}}{character defining where legend should usually be placed}

\item{\code{legendTitle}}{character defining the content of legend title}

\item{\code{plot}}{\code{BackgroundElement} object or list for plot area properties (outside of panel)}

\item{\code{panel}}{\code{BackgroundElement} object or list for plot area properties (inside of panel)}

\item{\code{xAxis}}{\code{BackgroundElement} object or list for x axis properties}

\item{\code{yAxis}}{\code{BackgroundElement} object or list for y axis properties}

\item{\code{xGrid}}{\code{BackgroundElement} object or list for x grid properties}

\item{\code{yGrid}}{\code{BackgroundElement} object or list for y grid properties}

\item{\code{legend}}{\code{BackgroundElement} object or list for legend area properties}

\item{\code{baseFill}}{name of base color fill of undefined background elements. Default is "white".}

\item{\code{baseColor}}{name of base color of undefined background elements. Default is "black".}

\item{\code{baseSize}}{name of base size of undefined background elements. Default is 0.5.}

\item{\code{baseLinetype}}{name of base size of undefined background elements. Default is "solid".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ThemeBackground} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeBackground-toJson"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeBackground-toJson}{}}}
\subsection{Method \code{toJson()}}{
Translate object into a json list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeBackground$toJson()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list that can be saved into a json file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemeBackground-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ThemeBackground-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeBackground$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
