% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ThemeFont}
\alias{ThemeFont}
\title{ThemeFont}
\description{
R6 class defining theme fonts
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{titleFont}}{R6 class \code{Font} object}

\item{\code{subtitleFont}}{R6 class \code{Font} object}

\item{\code{xlabelFont}}{R6 class \code{Font} object}

\item{\code{ylabelFont}}{R6 class \code{Font} object}

\item{\code{watermarkFont}}{R6 class \code{Font} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ThemeFont$new()}}
\item \href{#method-clone}{\code{ThemeFont$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{ThemeFont} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeFont$new(
  labelColors = tlfEnvThemesProperties$default$labelColors,
  labelBaseSize = 14
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labelColors}}{list of colors for each label}

\item{\code{labelBaseSize}}{numeric value for theme base size}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ThemeFont} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeFont$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
