% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ThemePlotConfigurations}
\alias{ThemePlotConfigurations}
\title{ThemePlotConfigurations}
\description{
R6 class defining theme of plot configuration objects
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{addScatter}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addScatter()}}

\item{\code{addLine}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addLine()}}

\item{\code{addRibbon}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addRibbon()}}

\item{\code{addErrorbar}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addErrorbar()}}

\item{\code{plotPKRatio}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotPKRatio()}}

\item{\code{plotBoxWhisker}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotBoxWhisker()}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ThemePlotConfigurations$new()}}
\item \href{#method-toJson}{\code{ThemePlotConfigurations$toJson()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ThemePlotConfigurations} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemePlotConfigurations$new(
  addScatter = NULL,
  addLine = NULL,
  addRibbon = NULL,
  addErrorbar = NULL,
  plotPKRatio = NULL,
  plotBoxWhisker = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{addScatter}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addScatter()}}

\item{\code{addLine}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addLine()}}

\item{\code{addRibbon}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addRibbon()}}

\item{\code{addErrorbar}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addErrorbar()}}

\item{\code{plotPKRatio}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotPKRatio()}}

\item{\code{plotBoxWhisker}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotBoxWhisker()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ThemeAestheticMaps} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJson"></a>}}
\if{latex}{\out{\hypertarget{method-toJson}{}}}
\subsection{Method \code{toJson()}}{
Translate object into a json list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemePlotConfigurations$toJson()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list that can be saved into a json file
}
}
}
