% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.R
\docType{data}
\name{timeProfileDataFrame}
\alias{timeProfileDataFrame}
\title{Test dataset for time profiles with 3 subjects.}
\format{
A data frame with 12 rows and 10 variables:
\describe{
\item{Time}{}
\item{IndividualID}{}
\item{Population}{}
\item{Gender}{}
\item{Age}{}
\item{Compound}{}
\item{Dose}{}
\item{Organ}{}
\item{Compartment}{}
\item{Simulated}{}
}
}
\usage{
timeProfileDataFrame
}
\description{
Test dataset for time profiles with 3 subjects.
}
\keyword{datasets}
