% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\docType{data}
\name{tlfEnv}
\alias{tlfEnv}
\title{tlfEnv}
\format{
An object of class \code{environment} of length 15.
}
\usage{
tlfEnv
}
\description{
Environment that holds various global variables and settings for the tlf library.
It is not exported and should not be directly manipulated by other packages.
}
\section{Fields}{

\describe{
\item{\code{packageName}}{Name of the package.
This will be used to retrieve information on the package at run time.}

\item{\code{defaultLegendPosition}}{Default legend position for plots. Use \code{setDefaultLegendPosition()} to change this value.}

\item{\code{defaultTagPosition}}{Default tag position of subplots.}

\item{\code{defaultHorizontalJustification}}{Default horizontal justification for plot labels.}

\item{\code{defaultVerticalJustificationDefault}}{horizontal vertical for plot labels.}

\item{\code{defaultExportParameters}}{Default properties for exporting/saving plots. Use \code{setDefaultExportParameters()} to change its values.}

\item{\code{defaultAggregation}}{Default aggregation binning, functions and labels for aggregating data.
Use \code{setDefaultAggregationBins()}, \code{setDefaultAggregationFunctions()}
and \code{setDefaultAggregationLabels()} to change its values.}

\item{\code{logTicks}}{Default tick values when using log ticks. Use \code{setDefaultLogTicks()} to change its values.}

\item{\code{lnTicks}}{Default tick values when using ln ticks}

\item{\code{logMinorTicks}}{Default minor tick values when using log ticks}

\item{\code{defaultErrorbarCapSize}}{Default cap size displayed for errorbars. Use \code{setDefaultErrorbarCapSize()} to change its values.
Currently \code{defaultErrorbarCapSize=0}, which hides the caps.}

\item{\code{defaultLLOQLinetype}}{LLOQ lines are of dotted type in the default settings. Use \code{setDefaultLLOQLinetype()} to change its values.}

\item{\code{DefaultAlphaRatio}}{Points bellow LLOQ have a transparency level at \verb{.618 * base alpha level}.
The \code{.618} value (1/golden ratio) ensures that both levels are always visible and
distinguishable even if the user set a low value for default alpha.
Use \code{setDefaultAlphaRatio()} to change its values.}

\item{\code{maxCharacterWidth}}{Maximum character width before wrap for axis ticks labels and legend
Use \code{setDefaultMaxCharacterWidth()} to change its values.}

\item{\code{currentTheme}}{Current \code{Theme} object used by plots providing their default plot configuration values.
Use \code{useTheme()} to change its values.}
}}

\keyword{internal}
