% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-input.R
\docType{data}
\name{tlfStatFunctions}
\alias{tlfStatFunctions}
\title{tlfStatFunctions}
\format{
An object of class \code{list} of length 31.
}
\usage{
tlfStatFunctions
}
\description{
Bank of predefined functions ready to use by Aggregation methods. Bank defined as Enum.
To access the function from its name, use match.fun: e.g. testFun <- match.fun("mean-1.96sd")
}
\seealso{
Other enum helpers: 
\code{\link{AestheticProperties}},
\code{\link{Alignments}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{LegendPositions}},
\code{\link{Linetypes}},
\code{\link{Scaling}},
\code{\link{Shapes}}
}
\concept{enum helpers}
\keyword{datasets}
