% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado-datamapping.R
\name{TornadoDataMapping}
\alias{TornadoDataMapping}
\title{TornadoDataMapping}
\description{
R6 class for mapping \code{values}, \code{labels} to \code{data}
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{TornadoDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lines}}{numeric vector of limits to plot}

\item{\code{sorted}}{logical indicating if values should be sorted}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TornadoDataMapping$new()}}
\item \href{#method-checkMapData}{\code{TornadoDataMapping$checkMapData()}}
\item \href{#method-clone}{\code{TornadoDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TornadoDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TornadoDataMapping$new(
  lines = DefaultDataMappingValues$tornado,
  sorted = NULL,
  x = NULL,
  y = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lines}}{numeric vector of limits to plot}

\item{\code{sorted}}{logical indicating if values should be sorted}

\item{\code{x}}{Variable including the values of tornado plot}

\item{\code{y}}{Variable including the labels of tornado plot}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TornadoDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkMapData"></a>}}
\if{latex}{\out{\hypertarget{method-checkMapData}{}}}
\subsection{Method \code{checkMapData()}}{
Check that \code{data} variables include map variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TornadoDataMapping$checkMapData(data, metaData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}

\item{\code{metaData}}{list containing information on \code{data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with map and \code{defaultAes} variables.
Dummy variable \code{defaultAes} is necessary to allow further modification of plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TornadoDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
