% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado-plotconfiguration.R
\name{TornadoPlotConfiguration}
\alias{TornadoPlotConfiguration}
\title{TornadoPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for tornado plots
}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{TornadoPlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bar}}{logical setting if tornado is uses a bar plot instead of regular points}

\item{\code{colorPalette}}{color palette property from \code{ggplot2}}

\item{\code{dodge}}{space between the bars/points}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TornadoPlotConfiguration$new()}}
\item \href{#method-clone}{\code{TornadoPlotConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TornadoPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TornadoPlotConfiguration$new(
  bar = TRUE,
  colorPalette = NULL,
  dodge = 0.5,
  lines = NULL,
  points = NULL,
  ribbons = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bar}}{logical setting if tornado is uses a bar plot instead of regular points}

\item{\code{colorPalette}}{color palette property from \code{ggplot2}}

\item{\code{dodge}}{space between the bars/points}

\item{\code{lines}}{\code{ThemeAestheticSelections} object defining properties for Tornado vertical lines}

\item{\code{points}}{\code{ThemeAestheticSelections} object defining properties for scatter points}

\item{\code{ribbons}}{\code{ThemeAestheticSelections} object defining properties for bars}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TornadoPlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TornadoPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
