% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-export.R
\name{updateExportDimensionsForLegend}
\alias{updateExportDimensionsForLegend}
\title{updateExportDimensionsForLegend}
\usage{
updateExportDimensionsForLegend(plotObject)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}
}
\value{
A \code{ggplot} object
}
\description{
Update plot dimensions based on size and position of legend
}
