% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\name{updateLegendCaption}
\alias{updateLegendCaption}
\title{updateLegendCaption}
\usage{
updateLegendCaption(captionData, newCaption = NULL, which = NULL)
}
\arguments{
\item{captionData}{data used for legend caption}

\item{newCaption}{new legend caption}

\item{which}{caption position in the legend}
}
\value{
updated caption data
}
\description{
updateLegendCaption return the updated data used for legend caption
}
