% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-timeprofile.R
\name{updateTimeProfileLegend}
\alias{updateTimeProfileLegend}
\title{updateTimeProfileLegend}
\usage{
updateTimeProfileLegend(plotObject, caption)
}
\arguments{
\item{plotObject}{A ggplot object}

\item{caption}{A data.frame as obtained from \code{getLegendCaption} to use for updating a plot legend.}
}
\value{
A \code{ggplot} object
}
\description{
Update time profile legend caption
}
