% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{useExcelTheme}
\alias{useExcelTheme}
\title{useExcelTheme}
\usage{
useExcelTheme()
}
\description{
Set default Excel theme to be used as the current default of the tlf environment
}
\examples{
useExcelTheme()
addScatter(x = seq(1,10), y = rnorm(10))
}
