% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{useHighChartTheme}
\alias{useHighChartTheme}
\title{useHighChartTheme}
\usage{
useHighChartTheme()
}
\description{
Set default HighChart theme to be used as the current default of the tlf environment
}
\examples{
useHighChartTheme()
addScatter(x = seq(1,10), y = rnorm(10))
}
