% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{useMatlabTheme}
\alias{useMatlabTheme}
\title{useMatlabTheme}
\usage{
useMatlabTheme()
}
\description{
Set default Matlab theme to be used as the current default of the tlf environment
}
\examples{
useMatlabTheme()
addScatter(x = seq(1, 10), y = rnorm(10))
}
