% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{useMinimalTheme}
\alias{useMinimalTheme}
\title{useMinimalTheme}
\usage{
useMinimalTheme()
}
\description{
Set default minimal theme to be used as the current default of the tlf environment
}
\examples{
useMinimalTheme()
addScatter(x = seq(1,10), y = rnorm(10))
}
