% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{useRETheme}
\alias{useRETheme}
\title{useRETheme}
\usage{
useRETheme()
}
\description{
Set default Reporting Engine theme to be used as the current default of the tlf environment
}
\examples{
useRETheme()
addScatter(x = seq(1,10), y = rnorm(10))
}
