% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{useTemplateTheme}
\alias{useTemplateTheme}
\title{useTemplateTheme}
\usage{
useTemplateTheme()
}
\description{
Set default theme to be used as the current default of the tlf environment
}
\examples{
useTemplateTheme()
addScatter(x = seq(1, 10), y = rnorm(10))
}
