% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-theme.R
\name{useTheme}
\alias{useTheme}
\title{useTheme}
\usage{
useTheme(theme)
}
\arguments{
\item{theme}{Theme to be used as default of the tlf environment}
}
\description{
set the theme to be used as the current default of the tlf environment
}
