% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enums.R
\docType{data}
\name{VerticalJustification}
\alias{VerticalJustification}
\title{VerticalJustification}
\format{
An object of class \code{list} of length 3.
}
\usage{
VerticalJustification
}
\description{
List of all available vertical justifications for plot annotation text.
}
\seealso{
Other enum helpers: 
\code{\link{AestheticFields}},
\code{\link{AestheticProperties}},
\code{\link{Alignments}},
\code{\link{Atoms}},
\code{\link{DataMappings}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{Linetypes}},
\code{\link{Molecules}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{PlotConfigurations}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{TickLabelTransforms}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
