% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\docType{class}
\name{XAxisConfiguration}
\alias{XAxisConfiguration}
\title{XAxisConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
XAxisConfiguration
}
\description{
Generic X axis Configuration
}
\keyword{datasets}
