% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{xAxisDefaultExpand}
\alias{xAxisDefaultExpand}
\title{xAxisDefaultExpand}
\usage{
xAxisDefaultExpand(plotConfiguration)
}
\arguments{
\item{plotConfiguration}{A \code{PlotConfiguration} object}
}
\value{
The default expand.
}
\description{
Return x-axis default expand from a plot configuration
}
\examples{
\dontrun{
# Regular plots use continuous linear scale for x-axis
plotConfiguration <- PlotConfiguration$new()
xAxisDefaultExpand(plotConfiguration)

# DDI plots use log scale for x-axis
ddiPlotConfiguration <- DDIRatioPlotConfiguration$new()
xAxisDefaultExpand(ddiPlotConfiguration)

# Boxplots use discrete scale for x-axis
boxPlotConfiguration <- BoxWhiskerPlotConfiguration$new()
xAxisDefaultExpand(boxPlotConfiguration)
}
}
\keyword{internal}
