% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{xAxisDefaultScale}
\alias{xAxisDefaultScale}
\title{xAxisDefaultScale}
\usage{
xAxisDefaultScale(plotConfiguration)
}
\arguments{
\item{plotConfiguration}{A \code{PlotConfiguration} object}
}
\value{
The default scale.
The enum \code{Scaling} provides a list of available scales.
}
\description{
Return x-axis default scale from a plot configuration
}
\examples{
\dontrun{
# Regular plots use continuous linear scale for x-axis
plotConfiguration <- PlotConfiguration$new()
xAxisDefaultScale(plotConfiguration)

# DDI plots use log scale for x-axis
ddiPlotConfiguration <- DDIRatioPlotConfiguration$new()
xAxisDefaultScale(ddiPlotConfiguration)

# Boxplots use discrete scale for x-axis
boxPlotConfiguration <- BoxWhiskerPlotConfiguration$new()
xAxisDefaultScale(boxPlotConfiguration)
}
}
\keyword{internal}
