% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-xrange.R
\docType{class}
\name{XRangeDataMapping}
\alias{XRangeDataMapping}
\title{XRangeDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
XRangeDataMapping
}
\description{
Abstract class for X Range Mapping
}
\keyword{datasets}
