% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-xy.R
\docType{class}
\name{XYDataMapping}
\alias{XYDataMapping}
\title{XYDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
XYDataMapping
}
\description{
Abstract class for X Y Mapping
}
\keyword{datasets}
