% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy-datamapping.R
\docType{class}
\name{XYEDataMapping}
\alias{XYEDataMapping}
\title{XYEDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
XYEDataMapping
}
\description{
Abstract class for X Y Error Mapping
}
\keyword{datasets}
