% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{yAxisDefaultExpand}
\alias{yAxisDefaultExpand}
\title{yAxisDefaultExpand}
\usage{
yAxisDefaultExpand(plotConfiguration)
}
\arguments{
\item{plotConfiguration}{A \code{PlotConfiguration} object}
}
\value{
The default expand.
}
\description{
Return x-axis default expand from a plot configuration
}
\examples{
\dontrun{
# Regular plots use continuous linear scale for x-axis
plotConfiguration <- PlotConfiguration$new()
yAxisDefaultExpand(plotConfiguration)

# DDI plots use log scale for x-axis
ddiPlotConfiguration <- DDIRatioPlotConfiguration$new()
yAxisDefaultExpand(ddiPlotConfiguration)

# Boxplots use discrete scale for x-axis
boxPlotConfiguration <- BoxWhiskerPlotConfiguration$new()
yAxisDefaultExpand(boxPlotConfiguration)
}
}
\keyword{internal}
