% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{yAxisDefaultScale}
\alias{yAxisDefaultScale}
\title{yAxisDefaultScale}
\usage{
yAxisDefaultScale(plotConfiguration)
}
\arguments{
\item{plotConfiguration}{A \code{PlotConfiguration} object}
}
\value{
The default scale.
The enum \code{Scaling} provides a list of available scales.
}
\description{
Return y-axis default scale from a plot configuration
}
\examples{
\dontrun{
# Regular plots use continuous linear scale for x-axis
plotConfiguration <- PlotConfiguration$new()
yAxisDefaultScale(plotConfiguration)

# DDI plots use log scale for y-axis
ddiPlotConfiguration <- DDIRatioPlotConfiguration$new()
yAxisDefaultScale(ddiPlotConfiguration)

# Tornado plots use discrete scale for y-axis
tornadoPlotConfiguration <- TornadoPlotConfiguration$new()
yAxisDefaultScale(tornadoPlotConfiguration)
}
}
\keyword{internal}
