% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom-plots.R
\name{addErrorbar}
\alias{addErrorbar}
\title{addErrorbar}
\usage{
addErrorbar(
  data = NULL,
  metaData = NULL,
  x = NULL,
  ymin = NULL,
  ymax = NULL,
  caption = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{data.frame containing the errorbar endpoints to be plotted}

\item{metaData}{list of information on \code{data} such as \code{dimension} and \code{unit} of their variables}

\item{x}{Mapping for x values.
If \code{data} is NULL or not input, \code{x} numeric values will be used as is for the plot.}

\item{ymin}{Mapping for ymin values.
If \code{data} is NULL or not input, \code{ymin} numeric values will be used as is for the plot.}

\item{ymax}{Mapping for ymin values.
If \code{data} is NULL or not input, \code{ymax} numeric values will be used as is for the plot.}

\item{caption}{vector of character strings defining the legend captions.
This parameter is optional: default value is NULL.}

\item{dataMapping}{\code{RangeDataMapping} class or subclass
mapping x, ymin, ymax and aesthetic variables to the variable names of \code{data}.}

\item{plotConfiguration}{\code{PlotConfiguration} object defining the label and background properties of the plot.}

\item{plotObject}{\code{ggplot} graphical object to which the line layer is added
This parameter is optional: the \code{tlf} library will initialize an empty plot if the parameter is NULL or not provided}
}
\value{
A \code{ggplot} graphical object
}
\description{
Add a errorbar plot layer to a \code{ggplot} graphical object.
Use optional argument \code{caption} to set legend caption.
Since \code{ggplot} manage aesthetic properties across all layers,
aesthetic properties defined in \code{plotConfiguration} will apply across all layers.
If caption is the same as a previous scatter plot layer, the legend will merge their caption and aesthetic properties
}
\examples{
# Add errorbar using x, ymin and ymax
p <- addErrorbar(x = c(1, 2, 1, 2, 3), ymax = c(5, 0, 2, 3, 4), ymin = c(2, -1, 1.5, 0, 0.3))

# Add a custom scatter
time <- seq(0, 10, 0.1)
customData <- data.frame(x = time, y = cos(time), ymin = (cos(time) - 0.2 * cos(time)), ymax = cos(time) + 0.2 * cos(time))

p <- addErrorbar(
  data = customData,
  dataMapping = RangeDataMapping$new(x = "x", ymin = "ymin", ymax = "ymax")
)

# Or for simple cases a smart mapping will get directly x, ymin and ymax from data
p <- addErrorbar(data = customData)

# Add an errorbar with caption
pe <- addErrorbar(data = customData, caption = "My plot with error bars")

# Add a scatter to the errorbar
pScatter <- addScatter(data = customData, caption = "My plot with error bars", plotObject = pe)

# Add a scatter to the errorbar
pLine <- addLine(data = customData, caption = "My plot with error bars", plotObject = pe)
}
