% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom-plots.R
\name{addLine}
\alias{addLine}
\title{addLine}
\usage{
addLine(data = NULL, x = NULL, y = NULL, dataMapping = NULL,
  color = "black", linetype = "solid", size = 1, plotObject = NULL)
}
\arguments{
\item{data}{data.frame containing the line endpoints to be plotted}

\item{x}{numeric value for vertical line}

\item{y}{numeric value for horizontal line}

\item{dataMapping}{\code{XYDataMapping} class or subclass
mapping x and y variables to the variable names of \code{data}.}

\item{color}{character string defining the color of the line.
This parameter is optional: default value is "black".}

\item{linetype}{character string defining the type of the line (plain, dash, dotted ...).
This parameter is optional: default value is "solid".}

\item{size}{numerical defining the width of the line.
This parameter is optional: default value is 1.}

\item{plotObject}{\code{ggplot} graphical object to which the line layer is added
This parameter is optional: the \code{tlf} library will initialize an empty plot if the parameter is NULL or not provided}
}
\value{
A \code{ggplot} graphical object
}
\description{
Add a line layer to a \code{ggplot} graphical object
}
\examples{
# Add vertical line at x = 2
addLine(x = 2)

# Add vertical line at x = 2 to a previous plot
p <- initializePlot() + 
ggplot2::geom_point(data = data.frame(x = c(1, 2, 3), y = c(5, 6, 4)),
 mapping = ggplot2::aes_string(x = "x", y = "y"))
p <- addLine(x = 2, plotObject = p)

# Add horizontal line at y = 5 to a previous plot
p <- addLine(y = 5, plotObject = p)

# Add a custom line
customLineData <- data.frame(x = c(0, 1, 2), y = c(0, 1, 1))
p <- addLine(data = customLineData, 
dataMapping = LineDataMapping$new(x = "x", y = "y"))

# Or for simple cases a smart mapping will get directly x and y from data
p <- addLine(data = customLineData)

# Add a line with custom aesthetic properties
p <- addLine(data = customLineData, color = "blue", linetype = "longdash", size = 0.5)
}
