% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom-plots.R
\name{addScatter}
\alias{addScatter}
\title{addScatter}
\usage{
addScatter(
  data = NULL,
  metaData = NULL,
  x = NULL,
  y = NULL,
  caption = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{data.frame containing the scatter points to be plotted}

\item{metaData}{list of information on \code{data} such as \code{dimension} and \code{unit} of their variables}

\item{x}{Mapping for x values.
If \code{data} is NULL or not input, \code{x} numeric values will be used as is for the plot.}

\item{y}{Mapping for y values.
If \code{data} is NULL or not input, \code{y} numeric values will be used as is for the plot.}

\item{caption}{vector of character strings defining the legend captions.
This parameter is optional: default value is NULL.}

\item{dataMapping}{\code{XYGDataMapping} class or subclass
mapping x, y and aesthetic variables to the variable names of \code{data}.}

\item{plotConfiguration}{\code{PlotConfiguration} object defining the label and background properties of the plot.}

\item{plotObject}{\code{ggplot} graphical object to which the line layer is added
This parameter is optional: the \code{tlf} library will initialize an empty plot if the parameter is NULL or not provided}
}
\value{
A \code{ggplot} graphical object
}
\description{
Add a scatter plot layer to a \code{ggplot} graphical object.
Use optional argument \code{caption} to set legend caption.
Since \code{ggplot} manage aesthetic properties across all layers,
aesthetic properties defined in \code{plotConfiguration} will apply across all layers.
}
\examples{
# Add scatter using x and y
p <- addScatter(x = c(1,2,1,2,3), y = c(5,0,2,3,4))

# Add a custom scatter
time <- seq(0, 30, 0.1)
customScatterData <- data.frame(x = time, y = cos(time))

p <- addScatter(
  data = customScatterData,
  dataMapping = XYGDataMapping$new(x = "x", y = "y")
)

# Or for simple cases a smart mapping will get directly x and y from data
p <- addScatter(data = customScatterData)

# Add a scatter with caption
p <- addScatter(data = customScatterData, caption = "My scatter plot")
}
