% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-background.R
\name{addWatermark}
\alias{addWatermark}
\title{addWatermark}
\usage{
addWatermark(
  plotObject,
  watermark,
  color = NULL,
  size = NULL,
  angle = NULL,
  alpha = NULL
)
}
\arguments{
\item{plotObject}{ggplot object to which the watermark is added}

\item{watermark}{Character or \code{Label} object corresponding to the watermark text
(and its font properties if \code{Label})}

\item{color}{Color of the watermark label.}

\item{size}{Size of the watermark label.}

\item{angle}{Angle of the watermark label (in degree).}

\item{alpha}{Transparency of the watermark label.
Alpha is a numeric between 0 and 1: 0 label is totally transparent, 1 label is totally opaque.}
}
\value{
\code{plotObject} ggplot object to which the watermark is added.
}
\description{
addWatermark creates a ggplot grob based on the label text and its font properties.
Then,  adds the grob to the ggplot object input \code{plotObject} as a new layer using \code{ggplot2::annotation_custom}.
}
\examples{
p <- ggplot2::ggplot()
addWatermark(p, "watermark")

# Watermark with font properties
watermarkLabel <- Label$new(text = "watermark", color = "blue")
addWatermark(p, watermarkLabel)

# Horizontal watermark
addWatermark(p, watermarkLabel, angle = 0)

# Watermark totally opaque
addWatermark(p, watermarkLabel, alpha = 1)

# Create a sun as background
for (angle in seq(0, 340, 20)) {
  p <- addWatermark(p,
    watermark = "            >",
    color = "yellow", angle = angle, alpha = 1
  )
}
}
