% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-background.R
\name{addWatermark}
\alias{addWatermark}
\title{addWatermark}
\usage{
addWatermark(plotObject, label, angle = 30, alpha = 0.4)
}
\arguments{
\item{plotObject}{ggplot object to which the watermark is added}

\item{label}{Character or Label class object corresponding to the watermark text
(and its font properties if Label)}

\item{angle}{Angle in degree from horizontal of the watermark label. Default angle is 30 degrees.}

\item{alpha}{Transparency of the watermark label.
Alpha is a numeric between 0 and 1: 0 label is totally transparent, 1 label is totally opaque.
Default alpha is 0.4.}
}
\value{
\code{plotObject} ggplot object to which the watermark is added.
}
\description{
addWatermark creates a ggplot grob based on the label text and its font properties.
Then,  adds the grob to the ggplot object input \code{plotObject} as a new layer using \code{ggplot2::annotation_custom}.
\code{angle} and \code{alpha} are optional input to customize the angle and transparency of the watermark text.
}
\examples{
p <- ggplot2::ggplot()
addWatermark(p, "watermark")

# Watermark with font properties
watermarkLabel <- Label$new(text = "watermark", color = "blue")
addWatermark(p, watermarkLabel)

# Horizontal watermark
addWatermark(p, watermarkLabel, angle = 0)

# Watermark totally opaque
addWatermark(p, watermarkLabel, alpha = 1)

# Create a sun as background
for (angle in seq(0, 340, 20)) {
  p <- addWatermark(p,
    label = Label$new(text = "            >", color = "yellow"),
    angle = angle, alpha = 1
  )
}
}
