% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enums.R
\docType{data}
\name{AestheticProperties}
\alias{AestheticProperties}
\title{AestheticProperties}
\format{
An object of class \code{list} of length 6.
}
\usage{
AestheticProperties
}
\description{
Enum of aesthetic property names of \code{ggplot2}
}
\seealso{
Other enum helpers: 
\code{\link{AestheticFields}},
\code{\link{AestheticSelectionKeys}},
\code{\link{Alignments}},
\code{\link{AtomPlots}},
\code{\link{ColorMaps}},
\code{\link{ColorPalettes}},
\code{\link{DataMappings}},
\code{\link{DefaultDataMappingValues}},
\code{\link{Directions}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{LegendTypes}},
\code{\link{Linetypes}},
\code{\link{MoleculePlots}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{PlotConfigurations}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{TickLabelTransforms}},
\code{\link{VerticalJustification}},
\code{\link{tlfSettingsNames}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
