% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\docType{data}
\name{AestheticProperties}
\alias{AestheticProperties}
\title{AestheticProperties}
\format{
An object of class \code{list} of length 6.
}
\usage{
AestheticProperties
}
\description{
Enum of aesthetic property names of \code{ggplot2}
}
\seealso{
Other enum helpers: 
\code{\link{Alignments}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{LegendPositions}},
\code{\link{Linetypes}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
