% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-input.R
\docType{class}
\name{AggregationInput}
\alias{AggregationInput}
\title{AggregationInput}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AggregationInput
}
\description{
#Class to be used to construct inputs to the AggregationSummary class
}
\keyword{datasets}
