% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-summary.R
\name{AggregationSummary}
\alias{AggregationSummary}
\title{AggregationSummary}
\description{
R6 class to split a data.frame data into subsets defined by unique combinations of elements
in the columns \code{xColumnNames} and \code{groupingColumnNames}.
Applies functions defined in \code{aggregationFunctionsVector} to column \code{yColumnNames.}
Returns a list of data.frame, one data.frame for each function listed in \code{aggregationFunctionsVector}.
Each data.frame in list element is named after the function's corresponding string in \code{aggregationFunctionNames}.
The summary statistic column name in each data.frame is the same as the name of the data.frame in the returned list.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{data.frame}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{xColumnNames}}{character names of grouping variables}

\item{\code{groupingColumnNames}}{character names of grouping variables}

\item{\code{yColumnNames}}{character names of dependent variables (that are grouped)}

\item{\code{aggregationInputsVector}}{list of R6 class \code{AggregationInput} objects}

\item{\code{aggregationFunctionsVector}}{list of functions to use for aggregation}

\item{\code{aggregationFunctionNames}}{vector of function names that will be used as variable name of the aggregation}

\item{\code{aggregationUnitsVector}}{character vector of units of aggregation output}

\item{\code{aggregationDimensionsVector}}{character vector of dimensions of aggregation output}

\item{\code{dfHelper}}{data.frame of aggregated values}

\item{\code{metaDataHelper}}{list of information on \code{dfHelper}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AggregationSummary$new()}}
\item \href{#method-applyAggregationFunctions}{\code{AggregationSummary$applyAggregationFunctions()}}
\item \href{#method-generateAggregatedValues}{\code{AggregationSummary$generateAggregatedValues()}}
\item \href{#method-clone}{\code{AggregationSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{AggregationSummary} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggregationSummary$new(
  data,
  metaData = NULL,
  xColumnNames = NULL,
  groupingColumnNames = NULL,
  yColumnNames = NULL,
  aggregationInputsVector = NULL,
  aggregationFunctionsVector = NULL,
  aggregationFunctionNames = NULL,
  aggregationUnitsVector = NULL,
  aggregationDimensionsVector = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{xColumnNames}}{character names of grouping variables}

\item{\code{groupingColumnNames}}{character names of grouping variables}

\item{\code{yColumnNames}}{character names of dependent variables (that are grouped)}

\item{\code{aggregationInputsVector}}{list of R6 class \code{AggregationInput} objects}

\item{\code{aggregationFunctionsVector}}{list of functions to use for aggregation}

\item{\code{aggregationFunctionNames}}{vector of function names that will be used as variable name of the aggregation}

\item{\code{aggregationUnitsVector}}{character vector of units of aggregation output}

\item{\code{aggregationDimensionsVector}}{character vector of dimensions of aggregation output}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{AggregationSummary} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-applyAggregationFunctions"></a>}}
\subsection{Method \code{applyAggregationFunctions()}}{
Apply aggregation functions on \code{x}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggregationSummary$applyAggregationFunctions(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list or vector of aggregated values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateAggregatedValues"></a>}}
\subsection{Method \code{generateAggregatedValues()}}{
Generate aggregated values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggregationSummary$generateAggregatedValues()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list or vector of aggregated values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggregationSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
