% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{asLabel}
\alias{asLabel}
\title{asLabel}
\usage{
asLabel(text)
}
\arguments{
\item{text}{text to set as label class}
}
\value{
Label
}
\description{
set a character type into a Label class
}
\examples{
title <- "Title of Plot"
title <- asLabel(title)
}
