% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-label.R
\name{asLabel}
\alias{asLabel}
\title{asLabel}
\usage{
asLabel(text = "", font = NULL)
}
\arguments{
\item{text}{text to be set as label class}

\item{font}{font properties of Label}
}
\value{
Label class object
}
\description{
Set a character string into a Label class associating font properties to input
If text is already a Label class, asLabel can be used to updated the font properties
}
\examples{
title <- "Title of Plot"
title <- asLabel(title)
}
