% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-utils.R
\name{asLabel}
\alias{asLabel}
\title{asLabel}
\usage{
asLabel(text = "", font = NULL)
}
\arguments{
\item{text}{text to be set as label class}

\item{font}{font properties of Label}
}
\value{
Label
}
\description{
Set a character class into a Label class
Associate font if input
If text is already Label class return it and update font if input
}
\examples{
title <- "Title of Plot"
title <- asLabel(title)
}
