% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\name{AxisConfiguration}
\alias{AxisConfiguration}
\title{AxisConfiguration}
\description{
R6 class defining the configuration of axis
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{limits}}{numeric vector of axis limits}

\item{\code{scale}}{character defining axis scale}

\item{\code{ticks}}{numeric vector or function defining where to position axis ticks}

\item{\code{ticklabels}}{character vector or function defining what to print on axis ticks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AxisConfiguration$new()}}
\item \href{#method-print}{\code{AxisConfiguration$print()}}
\item \href{#method-clone}{\code{AxisConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{AxisConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxisConfiguration$new(
  limits = NULL,
  scale = Scaling$lin,
  ticks = "default",
  ticklabels = "default"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limits}}{numeric vector of axis limits}

\item{\code{scale}}{character defining axis scale
Use enum \code{Scaling} to access predefined scales.}

\item{\code{ticks}}{numeric vector or function defining where to position axis ticks}

\item{\code{ticklabels}}{character vector or function defining what to print on axis ticks}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{AxisConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print axis properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxisConfiguration$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Axis properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxisConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
