<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\docType{class}
\name{AxisConfiguration}
\alias{AxisConfiguration}
\title{AxisConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AxisConfiguration
}
\description{
Generic axis Configuration
}
\keyword{datasets}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\docType{class}
\name{AxisConfiguration}
\alias{AxisConfiguration}
\title{AxisConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AxisConfiguration
}
\description{
Generic axis Configuration
}
\keyword{datasets}
>>>>>>> timeprofile-helper-grouping
