% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-background.R
\docType{class}
\name{BackgroundConfiguration}
\alias{BackgroundConfiguration}
\title{BackgroundConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BackgroundConfiguration
}
\description{
Class for Background Configuration
}
\section{Fields}{

\describe{
\item{\code{outerBackground}}{R6 class defining backgroundElement}

\item{\code{innerBackground}}{R6 class defining backgroundElement}

\item{\code{grid}}{R6 class defining backgroundElement}

\item{\code{watermark}}{R6 class defining watermark label}
}}

\section{Methods}{

\describe{
\item{new(outerBackground = NULL, innerBackground = NULL, grid = NULL, watermark = NULL, watermarkFont = NULL,
theme = tlfEnv$currentTheme)}{Initialize BackgroundConfiguration.}
\item{setBackground(plotObject)}{Apply properties to plot background.}
}
}

\keyword{datasets}
