% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-background.R
\name{BackgroundConfiguration}
\alias{BackgroundConfiguration}
\title{BackgroundConfiguration}
\description{
R6 class defining the configuration of background
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{outerBackground}}{R6 class \code{BackgroundElementConfiguration} object}

\item{\code{innerBackground}}{R6 class \code{BackgroundElementConfiguration} object}

\item{\code{grid}}{R6 class \code{BackgroundElementConfiguration} object}

\item{\code{watermark}}{R6 class \code{Label} object defining watermark background}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BackgroundConfiguration$new()}}
\item \href{#method-print}{\code{BackgroundConfiguration$print()}}
\item \href{#method-setBackground}{\code{BackgroundConfiguration$setBackground()}}
\item \href{#method-clone}{\code{BackgroundConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{BackgroundConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackgroundConfiguration$new(
  outerBackground = NULL,
  innerBackground = NULL,
  grid = NULL,
  watermark = NULL,
  watermarkFont = NULL,
  theme = tlfEnv$currentTheme
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outerBackground}}{R6 class \code{BackgroundElementConfiguration} object}

\item{\code{innerBackground}}{R6 class \code{BackgroundElementConfiguration} object}

\item{\code{grid}}{R6 class \code{BackgroundElementConfiguration} object}

\item{\code{watermark}}{R6 class \code{Label} object defining watermark background}

\item{\code{watermarkFont}}{R6 class \code{Font} object defining watermark font}

\item{\code{theme}}{R6 class \code{Theme} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BackgroundConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print background properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackgroundConfiguration$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Background properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setBackground"></a>}}
\if{latex}{\out{\hypertarget{method-setBackground}{}}}
\subsection{Method \code{setBackground()}}{
Set background properties of a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackgroundConfiguration$setBackground(plotObject)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{a \code{ggplot} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackgroundConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
