% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-boxwhisker.R
\docType{class}
\name{BoxWhiskerDataMapping}
\alias{BoxWhiskerDataMapping}
\title{BoxWhiskerDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BoxWhiskerDataMapping
}
\description{
Class for Box Whisker Mapping
}
\keyword{datasets}
