% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxwhisker-datamapping.R
\name{BoxWhiskerDataMapping}
\alias{BoxWhiskerDataMapping}
\title{BoxWhiskerDataMapping}
\description{
R6 class for mapping \code{y}, \code{GroupMapping}, \code{boxWhiskerLimits} and \code{outlierLimits} to \code{data}
}
\seealso{
Other DataMapping classes: 
\code{\link{DDIRatioDataMapping}},
\code{\link{GroupMapping}},
\code{\link{Grouping}},
\code{\link{HistogramDataMapping}},
\code{\link{ObsVsPredDataMapping}},
\code{\link{ObservedDataMapping}},
\code{\link{PKRatioDataMapping}},
\code{\link{QQDataMapping}},
\code{\link{RangeDataMapping}},
\code{\link{ResVsPredDataMapping}},
\code{\link{ResVsTimeDataMapping}},
\code{\link{TimeProfileDataMapping}},
\code{\link{TornadoDataMapping}},
\code{\link{XYDataMapping}},
\code{\link{XYGDataMapping}}
}
\concept{DataMapping classes}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{BoxWhiskerDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{outlierLimits}}{List of \code{minOutlierLimit} and \code{maxOutlierLimit} functions
outside which \code{data} is flagged as outlier}

\item{\code{boxWhiskerLimits}}{List of \code{ymin}, \code{lower}, \code{middle}, \code{upper} and \code{ymax} functions
calculated on \code{data} to obtain box whiskers}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BoxWhiskerDataMapping-new}{\code{BoxWhiskerDataMapping$new()}}
\item \href{#method-BoxWhiskerDataMapping-getBoxWhiskerLimits}{\code{BoxWhiskerDataMapping$getBoxWhiskerLimits()}}
\item \href{#method-BoxWhiskerDataMapping-getOutliers}{\code{BoxWhiskerDataMapping$getOutliers()}}
\item \href{#method-BoxWhiskerDataMapping-clone}{\code{BoxWhiskerDataMapping$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tlf" data-topic="XYGDataMapping" data-id="checkMapData"><a href='../../tlf/html/XYGDataMapping.html#method-XYGDataMapping-checkMapData'><code>tlf::XYGDataMapping$checkMapData()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoxWhiskerDataMapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-BoxWhiskerDataMapping-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{BoxWhiskerDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxWhiskerDataMapping$new(
  x = NULL,
  y,
  ymin = tlfStatFunctions$`Percentile5\%`,
  lower = tlfStatFunctions$`Percentile25\%`,
  middle = tlfStatFunctions$`Percentile50\%`,
  upper = tlfStatFunctions$`Percentile75\%`,
  ymax = tlfStatFunctions$`Percentile95\%`,
  minOutlierLimit = tlfStatFunctions$`Percentile25\%-1.5IQR`,
  maxOutlierLimit = tlfStatFunctions$`Percentile75\%+1.5IQR`,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Name of x variable to map
Default value is NULL in case of a unique box in the boxplot.}

\item{\code{y}}{Name of y variable to map}

\item{\code{ymin}}{Name of function used for calculating lower whisker.
Default value is \verb{Percentile5\%}.}

\item{\code{lower}}{Name of function used for calculating lower line of box
Default value is \verb{Percentile25\%}.}

\item{\code{middle}}{Name of function used for calculating middle line
Default value is \verb{Percentile55\%}.}

\item{\code{upper}}{Name of function used for calculating upper line of box
Default value is \verb{Percentile75\%}.}

\item{\code{ymax}}{Name of function used for calculating upper whisker
Default value is \verb{Percentile95\%}.}

\item{\code{minOutlierLimit}}{Name of function used for calculating lower outlier limit
Default value is \verb{Percentile25-1.5IQR\%}.}

\item{\code{maxOutlierLimit}}{Name of function used for calculating upper outlier limit
Default value is \verb{Percentile75+1.5IQR\%}.}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BoxWhiskerDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoxWhiskerDataMapping-getBoxWhiskerLimits"></a>}}
\if{latex}{\out{\hypertarget{method-BoxWhiskerDataMapping-getBoxWhiskerLimits}{}}}
\subsection{Method \code{getBoxWhiskerLimits()}}{
Get a data.frame with box-whisker limit by group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxWhiskerDataMapping$getBoxWhiskerLimits(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with \code{ymin}, \code{lower}, \code{middle}, \code{upper}, \code{ymax} variables.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoxWhiskerDataMapping-getOutliers"></a>}}
\if{latex}{\out{\hypertarget{method-BoxWhiskerDataMapping-getOutliers}{}}}
\subsection{Method \code{getOutliers()}}{
Get a data.frame flagging outliers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxWhiskerDataMapping$getOutliers(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with \code{minOutliers} and \code{maxOutliers} variables.
Values not flagged are \code{NA} in the outliers variables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoxWhiskerDataMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BoxWhiskerDataMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxWhiskerDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
