% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-boxwhisker.R
\docType{class}
\name{BoxWhiskerPlotConfiguration}
\alias{BoxWhiskerPlotConfiguration}
\title{BoxWhiskerPlotConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BoxWhiskerPlotConfiguration
}
\description{
Class for BoxWhisker Plot Configuration
}
\section{Fields}{

\describe{
\item{\code{legend}}{R6 class defining legendConfiguration}

\item{\code{xAxis}}{R6 class defining xAxisConfiguration}

\item{\code{yAxis}}{R6 class defining yAxisConfiguration}

\item{\code{background}}{R6 class defining backgroundConfiguration}

\item{\code{theme}}{R6 class defining theme aesthtic properties}

\item{\code{filename}}{Name of the saved plot}
}}

\section{Methods}{

\describe{
\item{new(title = "Box Whisker Plot", subtitle = paste("Date:", format(Sys.Date(), "%y-%m-%d")), ...)}{
Initialize BoxWhiskerPlotConfiguration}
\item{setPlotProperties(plotObject)}{Apply properties of plot labels.}
\item{setPlotBackground(plotObject)}{Apply background properties to plot.}
\item{savePlot(plotObject)}{Save ggplot as file.}
\item{addBoxWhisker(plotObject, data, metaData, dataMapping)}{Add Box and whiskers to plot.}
\item{addOutliers(plotObject, data, metaData, dataMapping)}{Add Outliers to plot.}
}
}

\keyword{datasets}
