% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxwhisker-plotconfiguration.R
\name{BoxWhiskerPlotConfiguration}
\alias{BoxWhiskerPlotConfiguration}
\title{BoxWhiskerPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for boxplots
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{ObsVsPredPlotConfiguration}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{QQPlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{BoxWhiskerPlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{defaultXScale}}{Default xAxis scale value when creating a \code{BoxWhiskerPlotConfiguration} object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{outliers}}{logical defining if outliers should be included in boxplot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BoxWhiskerPlotConfiguration-new}{\code{BoxWhiskerPlotConfiguration$new()}}
\item \href{#method-BoxWhiskerPlotConfiguration-clone}{\code{BoxWhiskerPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoxWhiskerPlotConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-BoxWhiskerPlotConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{BoxWhiskerPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxWhiskerPlotConfiguration$new(outliers = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outliers}}{logical defining if outliers should be included in boxplot}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BoxWhiskerPlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoxWhiskerPlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BoxWhiskerPlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxWhiskerPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
