% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxwhisker-plotconfiguration.R
\name{BoxWhiskerPlotConfiguration}
\alias{BoxWhiskerPlotConfiguration}
\title{BoxWhiskerPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object
}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{BoxWhiskerPlotConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoxWhiskerPlotConfiguration$new()}}
\item \href{#method-addBoxWhisker}{\code{BoxWhiskerPlotConfiguration$addBoxWhisker()}}
\item \href{#method-addOutliers}{\code{BoxWhiskerPlotConfiguration$addOutliers()}}
\item \href{#method-clone}{\code{BoxWhiskerPlotConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="PlotConfiguration" data-id="print">}\href{../../tlf/html/PlotConfiguration.html#method-print}{\code{tlf::PlotConfiguration$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{BoxWhiskerPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxWhiskerPlotConfiguration$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BoxWhiskerPlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addBoxWhisker"></a>}}
\subsection{Method \code{addBoxWhisker()}}{
Add a boxplot layer to a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxWhiskerPlotConfiguration$addBoxWhisker(
  plotObject,
  data,
  metaData,
  dataMapping
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{a \code{ggplot} object}

\item{\code{data}}{data.frame}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{dataMapping}}{R6 class \code{BoxWhiskerDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addOutliers"></a>}}
\subsection{Method \code{addOutliers()}}{
Add a outlier points layer to a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxWhiskerPlotConfiguration$addOutliers(
  plotObject,
  data,
  metaData,
  dataMapping
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{a \code{ggplot} object}

\item{\code{data}}{data.frame}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{dataMapping}}{R6 class \code{BoxWhiskerDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxWhiskerPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
