% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\docType{data}
\name{ColorMaps}
\alias{ColorMaps}
\title{ColorMaps}
\format{
An object of class \code{list} of length 5.
}
\usage{
ColorMaps
}
\description{
List with some color maps for \code{Theme} object
}
\keyword{datasets}
