% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enums.R
\docType{data}
\name{ColorMaps}
\alias{ColorMaps}
\title{ColorMaps}
\format{
An object of class \code{list} of length 6.
}
\usage{
ColorMaps
}
\description{
List with some color maps for \code{Theme} object.

The \code{ospDefault} color map is based on colors in \code{default_igv} qualitative
color palette from \code{{ggsci}} package.
}
\seealso{
Other enum helpers: 
\code{\link{AestheticFields}},
\code{\link{AestheticProperties}},
\code{\link{AestheticSelectionKeys}},
\code{\link{Alignments}},
\code{\link{AtomPlots}},
\code{\link{DataMappings}},
\code{\link{DefaultDataMappingValues}},
\code{\link{Directions}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{LegendTypes}},
\code{\link{Linetypes}},
\code{\link{MoleculePlots}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{PlotConfigurations}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{TickLabelTransforms}},
\code{\link{VerticalJustification}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
