% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark-utils.R
\name{createWatermarkGrob}
\alias{createWatermarkGrob}
\title{createWatermarkGrob}
\usage{
createWatermarkGrob(label, angle = 30)
}
\arguments{
\item{label}{Label class of Watermark}

\item{angle}{orientation of the watermark}
}
\value{
watermark Background as a grob for ggplot objects
}
\description{
createWatermarkGrob use a label to define a Background Watermark as a grob
}
