% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-watermark.R
\name{createWatermarkGrob}
\alias{createWatermarkGrob}
\title{createWatermarkGrob}
\usage{
createWatermarkGrob(label, angle = 30, alpha = 0.6)
}
\arguments{
\item{label}{Character or Label class object corresponding to the watermark text
(and its font properties if Label)}

\item{angle}{Angle in degree from horizontal of the watermark label. Default angle is 30 degrees.}

\item{alpha}{Transparency of the watermark label.
Alpha is a numeric between 0 and 1: 0 label is totally transparent, 1 label is totally opaque.
Default alpha is 0.6.}
}
\value{
Watermark background as a ggplot grob object
}
\description{
createWatermarkGrob creates a ggplot grob based on the label text and its font properties.
}
