% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-background.R
\name{createWatermarkGrob}
\alias{createWatermarkGrob}
\title{createWatermarkGrob}
\usage{
createWatermarkGrob(label, alpha = NULL)
}
\arguments{
\item{label}{A character value or \code{Label} object}

\item{alpha}{Numeric value between 0 and 1 corresponding to transparency of the watermark
The closer to 0, the more transparent the watermark is.
The closer to 1, the more opaque the watermark is.}
}
\value{
A \code{ggplot} grob
}
\description{
Creates a \code{ggplot} grob based on the label text and its font properties.
}
