% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-ddiratio.R
\docType{class}
\name{DDIRatioDataMapping}
\alias{DDIRatioDataMapping}
\title{DDIRatioDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DDIRatioDataMapping
}
\description{
Data Mapping for DDI Ratio plots
}
\keyword{datasets}
