% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiratio-datamapping.R
\name{DDIRatioDataMapping}
\alias{DDIRatioDataMapping}
\title{DDIRatioDataMapping}
\description{
R6 class for mapping \code{x}, \code{y}, \code{GroupMapping} and \code{ddiRatioLines} variables to \code{data}
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{\link[tlf:PKRatioDataMapping]{tlf::PKRatioDataMapping}} -> \code{DDIRatioDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{comparisonType}}{Options for comparison from enum \code{MappingComparisonTypes}}

\item{\code{minRange}}{Mininmum range for guest and ratio lines}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DDIRatioDataMapping$new()}}
\item \href{#method-getDDIRatioLines}{\code{DDIRatioDataMapping$getDDIRatioLines()}}
\item \href{#method-getGuestLines}{\code{DDIRatioDataMapping$getGuestLines()}}
\item \href{#method-clone}{\code{DDIRatioDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="PKRatioDataMapping" data-id="checkMapData">}\href{../../tlf/html/PKRatioDataMapping.html#method-checkMapData}{\code{tlf::PKRatioDataMapping$checkMapData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{DDIRatioDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$new(
  comparisonType = DDIComparisonTypes$obsVsPred,
  minRange = c(0.01, 100),
  lines = DefaultDataMappingValues$ddiRatio,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comparisonType}}{Options for comparison from enum \code{DDIComparisonTypes}}

\item{\code{minRange}}{Mininmum range for guest and ratio lines}

\item{\code{lines}}{list of ratio and guest limits to plot as horizontal lines}

\item{\code{...}}{parameters inherited from \code{PKRatioDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DDIRatioDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDDIRatioLines"></a>}}
\if{latex}{\out{\hypertarget{method-getDDIRatioLines}{}}}
\subsection{Method \code{getDDIRatioLines()}}{
Create a data.frame with DDI ratio limits
This data.frame is necessary in case if log-log plots as
\code{geom_abline} doesn't work properly in log scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$getDDIRatioLines()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame with DDI ratio limits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGuestLines"></a>}}
\if{latex}{\out{\hypertarget{method-getGuestLines}{}}}
\subsection{Method \code{getGuestLines()}}{
Get a data.frame with Guest et al ratio limits
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$getGuestLines()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame with Guest et al ratio limits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
