% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiratio-datamapping.R
\name{DDIRatioDataMapping}
\alias{DDIRatioDataMapping}
\title{DDIRatioDataMapping}
\description{
R6 class for mapping \code{x}, \code{y}, \code{GroupMapping} and \code{ddiRatioLines} variables to \code{data}
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{DDIRatioDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ddiRatioLines}}{numeric vector of ratio limits to plot}

\item{\code{deltaGuest}}{numeric value of Guest et al ratio limits}

\item{\code{range}}{2 elements vector of x limits}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DDIRatioDataMapping$new()}}
\item \href{#method-getDDIRatioLines}{\code{DDIRatioDataMapping$getDDIRatioLines()}}
\item \href{#method-getGuestLines}{\code{DDIRatioDataMapping$getGuestLines()}}
\item \href{#method-clone}{\code{DDIRatioDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="XYGDataMapping" data-id="checkMapData">}\href{../../tlf/html/XYGDataMapping.html#method-checkMapData}{\code{tlf::XYGDataMapping$checkMapData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{DDIRatioDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$new(
  ddiRatioValues = DefaultDataMappingValues$ddiRatio,
  range = c(0.01, 100),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ddiRatioValues}}{list of values for ratio and guest limits to plot}

\item{\code{range}}{2 elements vector of x limits}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DDIRatioDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDDIRatioLines"></a>}}
\subsection{Method \code{getDDIRatioLines()}}{
Create a data.frame with DDI ratio limits
This data.frame is necessary in case if log-log plots as
\code{geom_abline} doesn't work properly in log scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$getDDIRatioLines()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame with DDI ratio limits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGuestLines"></a>}}
\subsection{Method \code{getGuestLines()}}{
Get a data.frame with Guest et al ratio limits
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$getGuestLines()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame with Guest et al ratio limits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
