% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiratio-datamapping.R
\name{DDIRatioDataMapping}
\alias{DDIRatioDataMapping}
\title{DDIRatioDataMapping}
\description{
R6 class for mapping \code{x}, \code{y}, \code{GroupMapping} and \code{ddiRatioLines} variables to \code{data}
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{\link[tlf:PKRatioDataMapping]{tlf::PKRatioDataMapping}} -> \code{DDIRatioDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{minRange}}{Mininmum range for guest and ratio lines}

\item{\code{deltaGuest}}{Value of \code{delta} in Guest et al. equation}

\item{\code{residualsVsObserved}}{if true, returns a mapping for generating a residuals (predicted/observed) vs observed DDI ratio plot.  If false (default) returns a mapping for predicted vs observed DDI ratio plot.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DDIRatioDataMapping$new()}}
\item \href{#method-clone}{\code{DDIRatioDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="PKRatioDataMapping" data-id="checkMapData">}\href{../../tlf/html/PKRatioDataMapping.html#method-checkMapData}{\code{tlf::PKRatioDataMapping$checkMapData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{DDIRatioDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$new(
  deltaGuest = 1,
  minRange = c(0.01, 100),
  lines = DefaultDataMappingValues$ddiRatio,
  residualsVsObserved = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deltaGuest}}{Value of parameter \code{delta} in Guest et al. equation.
Default value is 1.}

\item{\code{minRange}}{Mininmum range for guest lines}

\item{\code{lines}}{list of ratio limits to plot as diagonal lines}

\item{\code{residualsVsObserved}}{if true, returns a mapping for generating a residuals (predicted/observed) vs observed DDI ratio plot.  If false (default) returns a mapping for predicted vs observed DDI ratio plot.}

\item{\code{...}}{parameters inherited from \code{PKRatioDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DDIRatioDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
