% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiratio-datamapping.R
\name{DDIRatioDataMapping}
\alias{DDIRatioDataMapping}
\title{DDIRatioDataMapping}
\description{
R6 class for mapping \code{x}, \code{y}, \code{GroupMapping} and DDI ratio \code{lines} variables to \code{data}
}
\seealso{
Other DataMapping classes: 
\code{\link{BoxWhiskerDataMapping}},
\code{\link{GroupMapping}},
\code{\link{Grouping}},
\code{\link{HistogramDataMapping}},
\code{\link{ObsVsPredDataMapping}},
\code{\link{ObservedDataMapping}},
\code{\link{PKRatioDataMapping}},
\code{\link{RangeDataMapping}},
\code{\link{ResVsPredDataMapping}},
\code{\link{ResVsTimeDataMapping}},
\code{\link{TimeProfileDataMapping}},
\code{\link{TornadoDataMapping}},
\code{\link{XYDataMapping}},
\code{\link{XYGDataMapping}}
}
\concept{DataMapping classes}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{\link[tlf:PKRatioDataMapping]{tlf::PKRatioDataMapping}} -> \code{DDIRatioDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{minRange}}{Mininmum range for guest and ratio lines}

\item{\code{deltaGuest}}{Value of \code{delta} in Guest et al. equation}

\item{\code{residualsVsObserved}}{if true, returns a mapping for generating a residuals (predicted/observed) vs observed DDI ratio plot.  If false (default) returns a mapping for predicted vs observed DDI ratio plot.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DDIRatioDataMapping-new}{\code{DDIRatioDataMapping$new()}}
\item \href{#method-DDIRatioDataMapping-clone}{\code{DDIRatioDataMapping$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tlf" data-topic="PKRatioDataMapping" data-id="checkMapData"><a href='../../tlf/html/PKRatioDataMapping.html#method-PKRatioDataMapping-checkMapData'><code>tlf::PKRatioDataMapping$checkMapData()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DDIRatioDataMapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-DDIRatioDataMapping-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{DDIRatioDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$new(
  deltaGuest = 1,
  minRange = c(0.01, 100),
  lines = DefaultDataMappingValues$ddiRatio,
  residualsVsObserved = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deltaGuest}}{Value of parameter \code{delta} in Guest et al. equation.
Default value is 1.}

\item{\code{minRange}}{Mininmum range for guest lines}

\item{\code{lines}}{List of ratio limits to display as diagonal/horizontal lines}

\item{\code{residualsVsObserved}}{if true, returns a mapping for generating a residuals (predicted/observed) vs observed DDI ratio plot.  If false (default) returns a mapping for predicted vs observed DDI ratio plot.}

\item{\code{...}}{parameters inherited from \code{PKRatioDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DDIRatioDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DDIRatioDataMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DDIRatioDataMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
