% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiratio-plotconfiguration.R
\name{DDIRatioPlotConfiguration}
\alias{DDIRatioPlotConfiguration}
\title{DDIRatioPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for DDI Ratio plots
}
\section{Super classes}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{\link[tlf:PKRatioPlotConfiguration]{tlf::PKRatioPlotConfiguration}} -> \code{DDIRatioPlotConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{DDIRatioPlotConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="PKRatioPlotConfiguration" data-id="initialize">}\href{../../tlf/html/PKRatioPlotConfiguration.html#method-initialize}{\code{tlf::PKRatioPlotConfiguration$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDIRatioPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
