% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{.removeInfiniteValues}
\alias{.removeInfiniteValues}
\title{.removeInfiniteValues}
\usage{
.removeInfiniteValues(x, range = c(0, 1))
}
\arguments{
\item{x}{numeric vector of values to manipulate}

\item{range}{numeric vector of length two giving desired output range}
}
\description{
Censor any values outside of range
}
\keyword{internal}
