% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-save.R
\name{ExportConfiguration}
\alias{ExportConfiguration}
\title{ExportConfiguration}
\description{
R6 class defining the save configuration of a \code{ggplot} object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{format}}{character defining the format of the file to be saved}

\item{\code{width}}{numeric values defining the width in \code{units} of the plot dimensions after saving}

\item{\code{height}}{numeric values defining the height in \code{units} of the plot dimensions after saving}

\item{\code{units}}{character defining the unit of the saving dimension}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExportConfiguration$new()}}
\item \href{#method-print}{\code{ExportConfiguration$print()}}
\item \href{#method-clone}{\code{ExportConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{ExportConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportConfiguration$new(
  format = tlfEnv$defaultExportParameters$format,
  width = tlfEnv$defaultExportParameters$width,
  height = tlfEnv$defaultExportParameters$height,
  units = tlfEnv$defaultExportParameters$units
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{character defining the format of the file to be saved.}

\item{\code{width}}{numeric values defining the width in \code{units} of the plot dimensions after saving}

\item{\code{height}}{numeric values defining the height in \code{units} of the plot dimensions after saving}

\item{\code{units}}{character defining the unit of the saving dimension}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ExportConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print file export properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportConfiguration$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
File export properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
