% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-export.R
\name{exportPlot}
\alias{exportPlot}
\title{exportPlot}
\usage{
exportPlot(plotObject, name = NULL)
}
\arguments{
\item{plotObject}{Graphical object created from ggplot}

\item{name}{name of exported file without extension}
}
\description{
Save a \code{ggplot} object according to its export properties
}
