% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-export.R
\name{exportPlotConfigurationCode}
\alias{exportPlotConfigurationCode}
\title{exportPlotConfigurationCode}
\usage{
exportPlotConfigurationCode(plotConfiguration)
}
\arguments{
\item{plotConfiguration}{A \code{PlotConfiguration} object}
}
\value{
R code to recreate the plot configuration as character
}
\description{
Export plot configuration as R code
}
