% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{Font}
\alias{Font}
\title{Font}
\description{
R6 class defining \code{size}, \code{color}, \code{fontFamily}, \code{fontFace} of font
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{size}}{numeric size of font}

\item{\code{color}}{character color of font}

\item{\code{fontFamily}}{character family of font}

\item{\code{fontFace}}{character face of font}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Font$new()}}
\item \href{#method-print}{\code{Font$print()}}
\item \href{#method-setFont}{\code{Font$setFont()}}
\item \href{#method-clone}{\code{Font$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Font} object.
Default font properties are defined directly in the object field,
so \code{NULL} input is allowed will lead to default properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Font$new(size = NULL, color = NULL, fontFamily = NULL, fontFace = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{numeric size of font}

\item{\code{color}}{character color of font}

\item{\code{fontFamily}}{character family of font}

\item{\code{fontFace}}{character face of font}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Font} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print \code{Font} properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Font$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setFont"></a>}}
\if{latex}{\out{\hypertarget{method-setFont}{}}}
\subsection{Method \code{setFont()}}{
Create an \code{element_text} for ggplot with \code{Font} properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Font$setFont()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An \code{element_text} for ggplot with \code{Font} properties.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Font$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
