% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{Font}
\alias{Font}
\title{Font}
\description{
R6 class defining font properties
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{size}}{numeric defining the size of font}

\item{\code{color}}{character defining the color of font}

\item{\code{fontFamily}}{character defining the family of font}

\item{\code{fontFace}}{character defining the face of font}

\item{\code{angle}}{numeric defining the angle of font}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Font$new()}}
\item \href{#method-createPlotFont}{\code{Font$createPlotFont()}}
\item \href{#method-clone}{\code{Font$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Font} object.
Default font properties are defined directly in the object field,
so \code{NULL} input is allowed will lead to default properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Font$new(
  size = NULL,
  color = NULL,
  fontFamily = NULL,
  fontFace = NULL,
  angle = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{numeric defining the size of font}

\item{\code{color}}{character defining the color of font}

\item{\code{fontFamily}}{character defining the family of font}

\item{\code{fontFace}}{character defining the face of font}

\item{\code{angle}}{numeric defining the angle of font}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Font} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createPlotFont"></a>}}
\if{latex}{\out{\hypertarget{method-createPlotFont}{}}}
\subsection{Method \code{createPlotFont()}}{
Create a \code{ggplot2::element_text} directly convertible by \code{ggplot2::theme}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Font$createPlotFont()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An \code{element_text} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Font$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
