% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\docType{class}
\name{Font}
\alias{Font}
\title{Font}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Font
}
\description{
Font properties for ggplot labels
}
\section{Fields}{

\describe{
\item{\code{size}}{numeric}

\item{\code{color}}{character}

\item{\code{fontFamily}}{character}

\item{\code{fontFace}}{character}
}}

\section{Methods}{

\describe{
\item{new(...)}{Initialize Font}
\item{print(...)}{Print Font properties}
\item{setFont()}{Create element_text for ggplot with Font properties}
}
}

\keyword{datasets}
