% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-boxwhisker-measure.R
\name{getBoxWhiskerMeasure}
\alias{getBoxWhiskerMeasure}
\title{getBoxWhiskerMeasure}
\usage{
getBoxWhiskerMeasure(data, dataMapping = NULL, y = NULL,
  group = NULL, quantiles = c(0.05, 0.25, 0.5, 0.75, 0.95))
}
\arguments{
\item{data}{data.frame (or list of data.frames? TO BE DISCUSSED)
containing the data to be used for the plot}

\item{dataMapping}{R6 class BoxWhiskerDataMapping
mapping of x, y and groupings}

\item{ratioLimits}{vector of numeric
containing the fold limits of PK ratios
As default limits are 1.5 and 2 folds}
}
\value{
a data.frame reporting the PK Ratio Qualification Measure
}
\description{
getPKRatioMeasure(data, dataMapping, ratioLimits) get the number of PK ratios
that are within specific limits
}
\details{
get summary table of Box Whisker plot
}
