% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkratio-datamapping.R
\name{getDefaultCaptions}
\alias{getDefaultCaptions}
\title{getDefaultCaptions}
\usage{
getDefaultCaptions(data, metaData, variableList = colnames(data))
}
\arguments{
\item{data}{input data.frame with variables to group by}

\item{metaData}{input data.frame with variables to group by}

\item{variableList}{groups as factor levels}
}
\value{
groupingVariable
}
\description{
createGroupingVariable create a new column that groups the grouping variable
}
