% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ddiratio.R
\name{getGuestValues}
\alias{getGuestValues}
\title{getGuestValues}
\usage{
getGuestValues(x, delta = 1, residualsVsObserved = FALSE)
}
\arguments{
\item{x}{input values of Guest function}

\item{delta}{parameter of Guest function}

\item{residualsVsObserved}{if true, returns a dataframe for a residuals (predicted/observed) vs observed DDI ratio plot}
}
\value{
A data.frame with x, ymin and ymax defining Guest et al. limits
}
\description{
Get a data.frame with Guest et al. ratio limits
}
