% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ddiratio.R
\name{getGuestValues}
\alias{getGuestValues}
\title{getGuestValues}
\usage{
getGuestValues(x, delta = 1)
}
\arguments{
\item{x}{input values of Guest function}

\item{delta}{parameter of Guest function}
}
\value{
A data.frame with Guest et al. ratio limits as ymin and ymax
}
\description{
Get a data.frame with Guest et al. ratio limits
}
