% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ddiratio.R
\name{getGuestValuesFromDataMapping}
\alias{getGuestValuesFromDataMapping}
\title{getGuestValuesFromDataMapping}
\usage{
getGuestValuesFromDataMapping(data, dataMapping)
}
\arguments{
\item{data}{data.frame containing the data to be used for the plot}

\item{dataMapping}{A \code{DDIRatioDataMapping} object}
}
\value{
A data.frame with x, ymin and ymax defining Guest et al. limits
}
\description{
Get a data.frame with Guest et al. ratio limits
}
