% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\name{getLabelWithUnit}
\alias{getLabelWithUnit}
\title{getLabelWithUnit}
\usage{
getLabelWithUnit(label, unit = NULL)
}
\arguments{
\item{label}{text of axis label}

\item{unit}{metaData containing unit of variable}
}
\value{
label character of variable with its unit when available
}
\description{
getLabelWithUnit paste unit to label when available
}
