% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{getLegendCaption}
\alias{getLegendCaption}
\title{getLegendCaption}
\usage{
getLegendCaption(plotObject)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}
}
\value{
A data.frame corresponding to legend caption
}
\description{
Get the legend caption
}
