% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\name{getLegendCaption}
\alias{getLegendCaption}
\title{getLegendCaption}
\usage{
getLegendCaption(captionData, which = seq(1,
  length(levels(captionData))))
}
\arguments{
\item{captionData}{data used for legend caption}

\item{which}{caption position in the legend}
}
\value{
legend captions
}
\description{
getLegendCaption return the legend captions or a subset using which
}
