% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPKRatioMeasure.R
\name{getPKRatioMeasure}
\alias{getPKRatioMeasure}
\title{getPKRatioMeasure}
\usage{
getPKRatioMeasure(data, dataMapping = NULL, ratioLimits = c(1.5, 2))
}
\arguments{
\item{data}{data.frame (or list of data.frames? TO BE DISCUSSED)
containing the data to be used for the plot}

\item{dataMapping}{R6 class PKRatioDataMapping
mapping of x, y axes + mapping of colorGrouping, sizeGrouping, shapeGrouping}

\item{metaData}{list of lists (structure TO BE DISCUSSED)
containing complementary information to data (e.g. unit)}
}
\value{
a data.frame reporting the PK Ratio Qualification Measure
}
\description{
getPKRatioMeasure(data, metaData, dataMapping)
}
\details{
get PK-Ratio qualification measure
}
