% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkratio-get-measure.R
\name{getPKRatioMeasure}
\alias{getPKRatioMeasure}
\title{getPKRatioMeasure}
\usage{
getPKRatioMeasure(data, dataMapping = NULL, ratioLimits = c(1.5, 2))
}
\arguments{
\item{data}{data.frame (or list of data.frames? TO BE DISCUSSED)
containing the data to be used for the plot}

\item{dataMapping}{R6 class PKRatioDataMapping
mapping of x, y and groupings}

\item{ratioLimits}{vector of numeric
containing the fold limits of PK ratios
As default limits are 1.5 and 2 folds}
}
\value{
a data.frame reporting the PK Ratio Qualification Measure
}
\description{
getPKRatioMeasure(data, dataMapping, ratioLimits) get the number of PK ratios
that are within specific limits
}
\details{
get PK-Ratio qualification measure
}
