% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy-datamapping.R
\docType{class}
\name{Grouping}
\alias{Grouping}
\title{Grouping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Grouping
}
\description{
Abstract Grouping class for Mapping colors, shape to groups
}
\keyword{datasets}
