% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-grouping.R
\name{Grouping}
\alias{Grouping}
\title{Grouping}
\description{
R6 class for mapping a \code{group} of variable(s) and their \code{label} to \code{data}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{group}}{data.frame or character defining the groups or group variables to group by}

\item{\code{label}}{character printed name of the \code{grouping}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Grouping$new()}}
\item \href{#method-getCaptions}{\code{Grouping$getCaptions()}}
\item \href{#method-clone}{\code{Grouping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Grouping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grouping$new(group, label = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{data.frame or character vector of groups}

\item{\code{label}}{character name of the \code{group}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Grouping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCaptions"></a>}}
\subsection{Method \code{getCaptions()}}{
Get the caption associated to each \code{group}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grouping$getCaptions(data, metaData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to map}

\item{\code{metaData}}{list of information on the \code{data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of characters containing the captions associated to each \code{group} of \code{data}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grouping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
