% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-groupmapping.R
\docType{class}
\name{GroupMapping}
\alias{GroupMapping}
\title{GroupMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupMapping
}
\description{
Abstract class for Grouping
}
\keyword{datasets}
