% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram-datamapping.R
\name{HistogramDataMapping}
\alias{HistogramDataMapping}
\title{HistogramDataMapping}
\description{
R6 class for mapping \code{x}, \code{verticalLineGroupings},
\code{verticalLineFunctionNames} and \code{verticalLineFunctions} variables to \code{data}
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{HistogramDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{verticalLineGroupings}}{R6 class \code{Grouping} variable}

\item{\code{verticalLineFunctionNames}}{Vector of function name
to be indicated in captions of the histogram}

\item{\code{verticalLineFunctions}}{List of functions calculated on \code{data}
to obtain vertical lines on the histogram}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HistogramDataMapping$new()}}
\item \href{#method-clone}{\code{HistogramDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="XYGDataMapping" data-id="checkMapData">}\href{../../tlf/html/XYGDataMapping.html#method-checkMapData}{\code{tlf::XYGDataMapping$checkMapData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{HistogramDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistogramDataMapping$new(
  verticalLineGroupings = NULL,
  verticalLineFunctionNames = c("mean", "median"),
  verticalLineFunctions = c(mean, median),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verticalLineGroupings}}{R6 class \code{Grouping} variable}

\item{\code{verticalLineFunctionNames}}{Vector of function name to be indicated in captions of the histogram
Default value uses \code{mean} and \code{median}.}

\item{\code{verticalLineFunctions}}{List of functions calculated on \code{data}
Default value uses \code{mean} and \code{median}.}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{HistogramDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistogramDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
