% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram-datamapping.R
\name{HistogramDataMapping}
\alias{HistogramDataMapping}
\title{HistogramDataMapping}
\description{
R6 class for mapping \code{x}, \code{bins}, \code{binwidth},\code{stack} and \code{distribution} to \code{data}
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{HistogramDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{stack}}{logical defining if histogram bars should be stacked}

\item{\code{bins}}{number of bins or binning values/methods passed on \code{ggplot2::geom_histogram}}

\item{\code{binwidth}}{width of bins passed on \code{ggplot2::geom_histogram}. Overwrites \code{bins}}

\item{\code{distribution}}{Name of distribution to fit to the data.
Only 2 distributions are currently available: \code{"normal"} and \code{"logNormal"}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HistogramDataMapping$new()}}
\item \href{#method-clone}{\code{HistogramDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="XYGDataMapping" data-id="checkMapData">}\href{../../tlf/html/XYGDataMapping.html#method-checkMapData}{\code{tlf::XYGDataMapping$checkMapData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{HistogramDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistogramDataMapping$new(
  stack = FALSE,
  bins = NULL,
  binwidth = NULL,
  distribution = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stack}}{logical defining if histogram bars should be stacked}

\item{\code{bins}}{argument passed on \code{ggplot2::geom_histogram}}

\item{\code{binwidth}}{width of bins passed on \code{ggplot2::geom_histogram}. Overwrites \code{bins}}

\item{\code{distribution}}{Name of distribution to fit to the data.
Only 2 distributions are currently available: \code{"normal"} and \code{"logNormal"}}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{HistogramDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistogramDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
