% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram-plotconfiguration.R
\name{HistogramPlotConfiguration}
\alias{HistogramPlotConfiguration}
\title{HistogramPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for histograms
}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{HistogramPlotConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HistogramPlotConfiguration$new()}}
\item \href{#method-addVerticalLines}{\code{HistogramPlotConfiguration$addVerticalLines()}}
\item \href{#method-clone}{\code{HistogramPlotConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{HistogramPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistogramPlotConfiguration$new(
  lines = NULL,
  ribbons = NULL,
  ylabel = "Count",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lines}}{\code{ThemeAestheticSelections} object defining properties for vertical lines}

\item{\code{ribbons}}{\code{ThemeAestheticSelections} object defining properties for histogram}

\item{\code{ylabel}}{Histograms default display is "Count"}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TimeProfilePlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addVerticalLines"></a>}}
\if{latex}{\out{\hypertarget{method-addVerticalLines}{}}}
\subsection{Method \code{addVerticalLines()}}{
Add statistics as line layer to a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistogramPlotConfiguration$addVerticalLines(
  plotObject,
  data,
  metaData,
  dataMapping
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{\code{ggplot} object}

\item{\code{data}}{data.frame}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{dataMapping}}{R6 class \code{HistogramDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistogramPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
