% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\name{HistogramPlotConfiguration}
\alias{HistogramPlotConfiguration}
\title{HistogramPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for histograms
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{ObsVsPredPlotConfiguration}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{QQPlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{tlf::PlotConfiguration} -> \code{HistogramPlotConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HistogramPlotConfiguration-new}{\code{HistogramPlotConfiguration$new()}}
\item \href{#method-HistogramPlotConfiguration-clone}{\code{HistogramPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistogramPlotConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-HistogramPlotConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{HistogramPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistogramPlotConfiguration$new(ylabel = "Count", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ylabel}}{Histograms default display is "Count"}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{HistogramPlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HistogramPlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HistogramPlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HistogramPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
