% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom-plots.R
\name{initializePlot}
\alias{initializePlot}
\title{initializePlot}
\usage{
initializePlot(plotConfiguration = NULL)
}
\arguments{
\item{plotConfiguration}{\code{PlotConfiguration} objecct defining labels, grid, background and watermark
This parameter is optional: the \code{tlf} library provides a default configuration according to the current theme}
}
\value{
A \code{ggplot} graphical object
}
\description{
Initialize a \code{ggplot} object and set the labels, grid, background and watermark
}
\examples{
# Initialize an empty plot
p <- initializePlot()

# Use a predifined theme
useTheme(tlfTheme)
p <- initializePlot()

# Implement a customized configuration using PlotConfiguration
config <- PlotConfiguration$new(title = "My Plot", xlabel = "x variable", ylabel = "y variable")
p <- initializePlot(config)
}
