% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\docType{class}
\name{Label}
\alias{Label}
\title{Label}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Label
}
\description{
Properties for ggplot labels
}
\section{Fields}{

\describe{
\item{\code{text}}{character}

\item{\code{font}}{Font object}
}}

\section{Methods}{

\describe{
\item{new(...)}{Initialize Label}
\item{print(...)}{Print Label properties}
\item{setFontProperties(color, size, fontFamily, fontFace)}{Set fields of font object}
}
}

\keyword{datasets}
