% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{Label}
\alias{Label}
\title{Label}
\description{
R6 class defining \code{text} and \code{font} of label
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{text}}{character text of the label}

\item{\code{font}}{\code{Font} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Label$new()}}
\item \href{#method-createPlotFont}{\code{Label$createPlotFont()}}
\item \href{#method-clone}{\code{Label$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Label} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Label$new(
  text = "",
  font = NULL,
  color = NULL,
  size = NULL,
  fontFace = NULL,
  fontFamily = NULL,
  angle = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{character text of the \code{Label} object}

\item{\code{font}}{\code{Font} object defining the font of the \code{Label} object}

\item{\code{color}}{character defining the color of the \code{Label} object}

\item{\code{size}}{numeric defining the size of the \code{Label} object}

\item{\code{fontFace}}{character defining the font face of the \code{Label} object}

\item{\code{fontFamily}}{character defining the font family of the \code{Label} object}

\item{\code{angle}}{numeric defining the angle of the \code{Label} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Label} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createPlotFont"></a>}}
\if{latex}{\out{\hypertarget{method-createPlotFont}{}}}
\subsection{Method \code{createPlotFont()}}{
Create a \code{ggplot2::element_text} directly convertible by \code{ggplot2::theme}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Label$createPlotFont()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An \code{element_text} or \code{element_blank}object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Label$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
