% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{Label}
\alias{Label}
\title{Label}
\description{
R6 class defining \code{text} and \code{font} of label
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{text}}{character text of label}

\item{\code{font}}{R6 class \code{Font} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Label$new()}}
\item \href{#method-print}{\code{Label$print()}}
\item \href{#method-setFontProperties}{\code{Label$setFontProperties()}}
\item \href{#method-clone}{\code{Label$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Label} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Label$new(
  text = "",
  font = NULL,
  color = NULL,
  size = NULL,
  fontFace = NULL,
  fontFamily = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{character text of label}

\item{\code{font}}{R6 class \code{Font} object}

\item{\code{color}}{character color of font}

\item{\code{size}}{numeric size of font}

\item{\code{fontFace}}{character face of font}

\item{\code{fontFamily}}{character family of font}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Label} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print \code{Label} properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Label$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setFontProperties"></a>}}
\subsection{Method \code{setFontProperties()}}{
Set font properties of Label
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Label$setFontProperties(
  color = self$font$color,
  size = self$font$size,
  fontFamily = self$font$fontFamily,
  fontFace = self$font$fontFace
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color}}{character color of font}

\item{\code{size}}{numeric size of font}

\item{\code{fontFamily}}{character family of font}

\item{\code{fontFace}}{character face of font}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Label$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
