% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-label.R
\name{LabelConfiguration}
\alias{LabelConfiguration}
\title{LabelConfiguration}
\description{
R6 class defining the configuration of the labels of a \code{ggplot} object
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{title}}{\code{Label} object defining the title of the plot}

\item{\code{subtitle}}{\code{Label} object defining the subtitle of the plot}

\item{\code{xlabel}}{\code{Label} object defining the xlabel of the plot}

\item{\code{ylabel}}{\code{Label} object defining the ylabel of the plot}

\item{\code{caption}}{\code{Label} object defining the caption of the plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LabelConfiguration$new()}}
\item \href{#method-updatePlot}{\code{LabelConfiguration$updatePlot()}}
\item \href{#method-clone}{\code{LabelConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{LabelConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelConfiguration$new(
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  caption = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{character or \code{Label} object defining title}

\item{\code{subtitle}}{character or \code{Label} object defining subtitle}

\item{\code{xlabel}}{character or \code{Label} object defining xlabel}

\item{\code{ylabel}}{character or \code{Label} object defining ylabel}

\item{\code{caption}}{character or \code{Label} object defining caption}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{LabelConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePlot"></a>}}
\if{latex}{\out{\hypertarget{method-updatePlot}{}}}
\subsection{Method \code{updatePlot()}}{
Update labels of a \code{ggplot} object and their properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelConfiguration$updatePlot(plotObject)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{a \code{ggplot} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
