% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-label.R
\docType{class}
\name{LabelConfiguration}
\alias{LabelConfiguration}
\title{LabelConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LabelConfiguration
}
\description{
Generic Label Configuration
}
\keyword{datasets}
