% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-legend.R
\name{LegendConfiguration}
\alias{LegendConfiguration}
\title{LegendConfiguration}
\description{
R6 class defining the legend configuration of a \code{ggplot} object
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{caption}}{of legend defined as data.frame with caption properties}

\item{\code{position}}{of legend as defined in Enum \code{LegendPositions}}

\item{\code{font}}{\code{Font} object defining the font of the legend}

\item{\code{background}}{\code{Background} object defining the background of the legend}

\item{\code{title}}{character defining title of the legend}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LegendConfiguration-new}{\code{LegendConfiguration$new()}}
\item \href{#method-LegendConfiguration-updatePlot}{\code{LegendConfiguration$updatePlot()}}
\item \href{#method-LegendConfiguration-clone}{\code{LegendConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LegendConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-LegendConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{LegendConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegendConfiguration$new(
  position = NULL,
  caption = NULL,
  title = NULL,
  font = NULL,
  background = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{position of the legend as defined by enum \code{LegendPositions}}

\item{\code{caption}}{data.frame containing the properties of the legend caption}

\item{\code{title}}{character or \code{Label} object defining the title of the legend. A value of \code{NULL} removes the title.}

\item{\code{font}}{\code{Font} object defining the font of the legend caption}

\item{\code{background}}{\code{BackgroundElement} object defining the background of the legend}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{LegendConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LegendConfiguration-updatePlot"></a>}}
\if{latex}{\out{\hypertarget{method-LegendConfiguration-updatePlot}{}}}
\subsection{Method \code{updatePlot()}}{
Update legend configuration on a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegendConfiguration$updatePlot(plotObject)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{\code{ggplot} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object with updated axis properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LegendConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LegendConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegendConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
