% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-legend.R
\name{LegendConfiguration}
\alias{LegendConfiguration}
\title{LegendConfiguration}
\description{
R6 class defining the legend configuration of a \code{ggplot} object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{position}}{character position of the legend}

\item{\code{title}}{character name of the legend}

\item{\code{caption}}{data.frame with}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LegendConfiguration$new()}}
\item \href{#method-print}{\code{LegendConfiguration$print()}}
\item \href{#method-clone}{\code{LegendConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{LegendConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegendConfiguration$new(
  position = LegendPositions$outsideRight,
  title = NULL,
  caption = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{character position of the legend.
Use enum \code{LegendPositions} to assess available legend positions.}

\item{\code{title}}{character title of the legend caption.
Default \code{NULL} does not provide any legend title.}

\item{\code{caption}}{data.frame containing the legend caption properties}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{LegendConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print legend properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegendConfiguration$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Legend properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegendConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
