% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-legend.R
\name{LegendConfiguration}
\alias{LegendConfiguration}
\title{LegendConfiguration}
\description{
R6 class defining the legend configuration of a \code{ggplot} object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{position}}{character position of the legend}

\item{\code{titles}}{R6 class \code{LegendTitles} object}

\item{\code{captions}}{R6 class \code{LegendCaptions} object}

\item{\code{values}}{R6 class \code{LegendValues} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LegendConfiguration$new()}}
\item \href{#method-setPlotLegend}{\code{LegendConfiguration$setPlotLegend()}}
\item \href{#method-clone}{\code{LegendConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{LegendConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegendConfiguration$new(
  position = NULL,
  titles = NULL,
  captions = NULL,
  values = NULL,
  dataMapping = NULL,
  data = NULL,
  metaData = NULL,
  theme = tlfEnv$currentTheme
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{character position of the legend.
Use enum \code{LegendPositions} to assess available legend positions.}

\item{\code{titles}}{R6 class \code{LegendTitles} object}

\item{\code{captions}}{R6 class \code{LegendCaptions} object}

\item{\code{values}}{R6 class \code{LegendValues} object}

\item{\code{dataMapping}}{R6 class or subclass \code{XYDataMapping}}

\item{\code{data}}{data.frame used by \code{smartMapping}}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{theme}}{R6 class \code{Theme}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{LegendConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPlotLegend"></a>}}
\subsection{Method \code{setPlotLegend()}}{
Set plot legend of a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegendConfiguration$setPlotLegend(plotObject)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{\code{ggplot} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegendConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
