% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\docType{class}
\name{LegendConfiguration}
\alias{LegendConfiguration}
\title{LegendConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LegendConfiguration
}
\description{
Generic Legend Configuration
}
\keyword{datasets}
