% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-enums.R
\docType{data}
\name{Linetypes}
\alias{Linetypes}
\title{Linetypes}
\format{
An object of class \code{list} of length 7.
}
\usage{
Linetypes
}
\description{
Enum of \code{ggplot2} linetypes
}
\seealso{
Other enum helpers: 
\code{\link{AestheticProperties}},
\code{\link{Alignments}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{VerticalJustification}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
