% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\docType{data}
\name{Linetypes}
\alias{Linetypes}
\title{Linetypes}
\format{
An object of class \code{list} of length 7.
}
\usage{
Linetypes
}
\description{
Enum of \code{ggplot2} linetypes
}
\keyword{datasets}
