% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed-data-mapping.R
\name{ObservedDataMapping}
\alias{ObservedDataMapping}
\title{ObservedDataMapping}
\description{
R6 class for mapping \code{x}, \code{y}, of observed data for a time profile plot
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{ObservedDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lloq}}{mapping lower limit of quantitation variable}

\item{\code{uncertainty}}{mapping error bars around scatter points}

\item{\code{mdv}}{mapping missing dependent variable}

\item{\code{ymin}}{mapping error bars around scatter points}

\item{\code{ymax}}{mapping error bars around scatter points}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ObservedDataMapping$new()}}
\item \href{#method-checkMapData}{\code{ObservedDataMapping$checkMapData()}}
\item \href{#method-clone}{\code{ObservedDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ObservedDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservedDataMapping$new(lloq = NULL, uncertainty = NULL, mdv = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lloq}}{mapping lower limit of quantitation variable}

\item{\code{uncertainty}}{mapping error bars around scatter points}

\item{\code{mdv}}{mapping missing dependent variable}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PKRatioDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkMapData"></a>}}
\if{latex}{\out{\hypertarget{method-checkMapData}{}}}
\subsection{Method \code{checkMapData()}}{
Check that \code{data} variables include map variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservedDataMapping$checkMapData(data, metaData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}

\item{\code{metaData}}{list containing information on \code{data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with map and \code{defaultAes} variables.
Dummy variable \code{defaultAes} is necessary to allow further modification of plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservedDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
