% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed-data-mapping.R
\name{ObservedDataMapping}
\alias{ObservedDataMapping}
\title{ObservedDataMapping}
\description{
R6 class for mapping \code{x}, \code{y}, of observed data for a time profile plot
}
\seealso{
Other DataMapping classes: 
\code{\link{BoxWhiskerDataMapping}},
\code{\link{CumulativeTimeProfileDataMapping}},
\code{\link{DDIRatioDataMapping}},
\code{\link{GroupMapping}},
\code{\link{Grouping}},
\code{\link{HistogramDataMapping}},
\code{\link{ObsVsPredDataMapping}},
\code{\link{PKRatioDataMapping}},
\code{\link{QQDataMapping}},
\code{\link{RangeDataMapping}},
\code{\link{ResVsPredDataMapping}},
\code{\link{ResVsTimeDataMapping}},
\code{\link{TimeProfileDataMapping}},
\code{\link{TornadoDataMapping}},
\code{\link{XYDataMapping}},
\code{\link{XYGDataMapping}}
}
\concept{DataMapping classes}
\section{Super classes}{
\code{tlf::XYDataMapping} -> \code{tlf::XYGDataMapping} -> \code{ObservedDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{error}}{mapping error bars around scatter points}

\item{\code{mdv}}{mapping missing dependent variable}

\item{\code{ymin}}{mapping error bars around scatter points}

\item{\code{ymax}}{mapping error bars around scatter points}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObservedDataMapping-new}{\code{ObservedDataMapping$new()}}
\item \href{#method-ObservedDataMapping-checkMapData}{\code{ObservedDataMapping$checkMapData()}}
\item \href{#method-ObservedDataMapping-clone}{\code{ObservedDataMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservedDataMapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObservedDataMapping-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ObservedDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservedDataMapping$new(
  x,
  y,
  ymin = NULL,
  ymax = NULL,
  color = NULL,
  shape = NULL,
  group = NULL,
  error = NULL,
  uncertainty = NULL,
  mdv = NULL,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{y}}{Name of y variable to map}

\item{\code{ymin}}{mapping lower end of error bars around scatter points}

\item{\code{ymax}}{mapping upper end of error bars around scatter points}

\item{\code{color}}{R6 class \code{Grouping} object or its input}

\item{\code{shape}}{R6 class \code{Grouping} object or its input}

\item{\code{group}}{R6 class \code{Grouping} object or its input}

\item{\code{error}}{mapping error bars around scatter points}

\item{\code{uncertainty}}{mapping error bars around scatter points.
Deprecated parameter replaced by \code{error}.}

\item{\code{mdv}}{mapping missing dependent variable}

\item{\code{data}}{data.frame to map used by \code{.smartMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ObservedDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservedDataMapping-checkMapData"></a>}}
\if{latex}{\out{\hypertarget{method-ObservedDataMapping-checkMapData}{}}}
\subsection{Method \code{checkMapData()}}{
Check that \code{data} variables include map variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservedDataMapping$checkMapData(data, metaData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}

\item{\code{metaData}}{list containing information on \code{data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with map and \code{defaultAes} variables.
Dummy variable \code{defaultAes} is necessary to allow further modification of plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservedDataMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObservedDataMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservedDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
