% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs-vs-pred-datamapping.R
\name{ObsVsPredDataMapping}
\alias{ObsVsPredDataMapping}
\title{ObsVsPredDataMapping}
\description{
R6 class for mapping \code{x}, \code{y}, \code{GroupMapping} and \code{obsVsPredLines} variables to \code{data}
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{ObsVsPredDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{obsVsPredValues}}{numeric vector of limits to plot}

\item{\code{lloq}}{numeric value of lower limit of quantification}

\item{\code{smoother}}{regression function name}

\item{\code{range}}{2 elements vector of x limits}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ObsVsPredDataMapping$new()}}
\item \href{#method-getObsVsPredLines}{\code{ObsVsPredDataMapping$getObsVsPredLines()}}
\item \href{#method-clone}{\code{ObsVsPredDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="XYGDataMapping" data-id="checkMapData">}\href{../../tlf/html/XYGDataMapping.html#method-checkMapData}{\code{tlf::XYGDataMapping$checkMapData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{ObsVsPredDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredDataMapping$new(
  obsVsPredValues = DefaultDataMappingValues$obsVsPred,
  lloq = NULL,
  range = NULL,
  smoother = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obsVsPredValues}}{list of values for obs vs pred plot}

\item{\code{lloq}}{numeric value of lower limit of quantification}

\item{\code{range}}{2 elements vector of x limits}

\item{\code{smoother}}{smoother function or parameter
To map a loess smoother to the plot, use \code{smoother}="loess"}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ObsVsPredDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getObsVsPredLines"></a>}}
\subsection{Method \code{getObsVsPredLines()}}{
Create a data.frame with of limits to plot
This data.frame is necessary in case if log-log plots as
\code{geom_abline} doesn't work properly in log scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredDataMapping$getObsVsPredLines(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame of data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
