% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs-vs-pred-datamapping.R
\name{ObsVsPredDataMapping}
\alias{ObsVsPredDataMapping}
\title{ObsVsPredDataMapping}
\description{
Class for mapping variables in observations vs predictions plot
}
\seealso{
Other DataMapping classes: 
\code{\link{BoxWhiskerDataMapping}},
\code{\link{DDIRatioDataMapping}},
\code{\link{GroupMapping}},
\code{\link{Grouping}},
\code{\link{HistogramDataMapping}},
\code{\link{ObservedDataMapping}},
\code{\link{PKRatioDataMapping}},
\code{\link{RangeDataMapping}},
\code{\link{ResVsPredDataMapping}},
\code{\link{ResVsTimeDataMapping}},
\code{\link{TimeProfileDataMapping}},
\code{\link{TornadoDataMapping}},
\code{\link{XYDataMapping}},
\code{\link{XYGDataMapping}}
}
\concept{DataMapping classes}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{ObsVsPredDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lines}}{list of ratio limits to plot as horizontal lines}

\item{\code{xmin}}{mapping of upper value of error bars around scatter points}

\item{\code{xmax}}{mapping of lower value of error bars around scatter points}

\item{\code{smoother}}{regression function name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObsVsPredDataMapping-new}{\code{ObsVsPredDataMapping$new()}}
\item \href{#method-ObsVsPredDataMapping-clone}{\code{ObsVsPredDataMapping$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tlf" data-topic="PKRatioDataMapping" data-id="checkMapData"><a href='../../tlf/html/PKRatioDataMapping.html#method-PKRatioDataMapping-checkMapData'><code>tlf::PKRatioDataMapping$checkMapData()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsVsPredDataMapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObsVsPredDataMapping-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ObsVsPredDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredDataMapping$new(
  x = NULL,
  y = NULL,
  xmin = NULL,
  xmax = NULL,
  lines = DefaultDataMappingValues$obsVsPred,
  smoother = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{y}}{Name of y variable to map}

\item{\code{xmin}}{mapping of upper value of error bars around scatter points}

\item{\code{xmax}}{mapping of lower value of error bars around scatter points}

\item{\code{lines}}{list of lines to plot}

\item{\code{smoother}}{smoother function or parameter
To map a loess smoother to the plot, use \code{smoother}="loess"}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ObsVsPredDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsVsPredDataMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObsVsPredDataMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
