% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs-vs-pred-plotconfiguration.R
\name{ObsVsPredPlotConfiguration}
\alias{ObsVsPredPlotConfiguration}
\title{ObsVsPredPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for Obs vs Pred plots
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{ObsVsPredPlotConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ObsVsPredPlotConfiguration$new()}}
\item \href{#method-clone}{\code{ObsVsPredPlotConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ObsVsPredPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredPlotConfiguration$new(
  lines = NULL,
  points = NULL,
  errorbars = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lines}}{\code{ThemeAestheticSelections} object defining properties for lines}

\item{\code{points}}{\code{ThemeAestheticSelections} object defining properties for scatter points}

\item{\code{errorbars}}{\code{ThemeAestheticSelections} object defining properties for error bars}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ObsVsPredPlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
