% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs-vs-pred-plotconfiguration.R
\name{ObsVsPredPlotConfiguration}
\alias{ObsVsPredPlotConfiguration}
\title{ObsVsPredPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for Obs vs Pred plots
}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{ObsVsPredPlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{obsVsPredCaption}}{list of properties for obs vs pred plot specific features}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ObsVsPredPlotConfiguration$new()}}
\item \href{#method-addSmoother}{\code{ObsVsPredPlotConfiguration$addSmoother()}}
\item \href{#method-addObsVsPred}{\code{ObsVsPredPlotConfiguration$addObsVsPred()}}
\item \href{#method-clone}{\code{ObsVsPredPlotConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="PlotConfiguration" data-id="print">}\href{../../tlf/html/PlotConfiguration.html#method-print}{\code{tlf::PlotConfiguration$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{ObsVsPredPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredPlotConfiguration$new(
  obsVsPredCaption = getDefaultCaptionFor("obsVsPred"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obsVsPredCaption}}{list of properties for DDI ratio plot specific features}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{obsVsPredProperties} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addSmoother"></a>}}
\subsection{Method \code{addSmoother()}}{
Add smoother layer to a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredPlotConfiguration$addSmoother(plotObject, data, metaData, dataMapping)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{\code{ggplot} object}

\item{\code{data}}{data.frame}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{dataMapping}}{R6 class \code{ObsVsPredDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object with smoother
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addObsVsPred"></a>}}
\subsection{Method \code{addObsVsPred()}}{
Add obs vs pred as scatter layer to a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredPlotConfiguration$addObsVsPred(
  plotObject,
  data,
  metaData,
  dataMapping
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{\code{ggplot} object}

\item{\code{data}}{data.frame}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{dataMapping}}{R6 class \code{ObsVsPredDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
