% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-sub-classes.R
\name{ObsVsPredPlotConfiguration}
\alias{ObsVsPredPlotConfiguration}
\title{ObsVsPredPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for Obs vs Pred plots
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{AxisConfiguration}},
\code{\link{BackgroundConfiguration}},
\code{\link{BackgroundElement}},
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{CumulativeTimeProfilePlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{ExportConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{LabelConfiguration}},
\code{\link{LegendConfiguration}},
\code{\link{LineElement}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{PlotGridConfiguration}},
\code{\link{QQPlotConfiguration}},
\code{\link{ResVsPredPlotConfiguration}},
\code{\link{ResVsTimePlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}},
\code{\link{XAxisConfiguration}},
\code{\link{YAxisConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{ObsVsPredPlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{defaultSymmetricAxes}}{Default option setting symmetric xAxis and/or yAxis limits when creating a \code{ObsVsPredPlotConfiguration} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObsVsPredPlotConfiguration-new}{\code{ObsVsPredPlotConfiguration$new()}}
\item \href{#method-ObsVsPredPlotConfiguration-getKeyGlyph}{\code{ObsVsPredPlotConfiguration$getKeyGlyph()}}
\item \href{#method-ObsVsPredPlotConfiguration-clone}{\code{ObsVsPredPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsVsPredPlotConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObsVsPredPlotConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ObsVsPredPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredPlotConfiguration$new(
  foldLinesLegend = FALSE,
  foldLinesLegendDiagonal = FALSE,
  legendPosition = "insideTopLeft",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}

\item{\code{colorPalette}}{color palette property from \code{ggplot2}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{CumulativeTimeProfilePlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsVsPredPlotConfiguration-getKeyGlyph"></a>}}
\if{latex}{\out{\hypertarget{method-ObsVsPredPlotConfiguration-getKeyGlyph}{}}}
\subsection{Method \code{getKeyGlyph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredPlotConfiguration$getKeyGlyph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsVsPredPlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObsVsPredPlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsVsPredPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
