% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-pkratio.R, R/pkratio-datamapping.R
\docType{class}
\name{PKRatioDataMapping}
\alias{PKRatioDataMapping}
\title{PKRatioDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PKRatioDataMapping

PKRatioDataMapping
}
\description{
Data Mapping for PKRatio

Data Mapping for PKRatio
}
\keyword{datasets}
