% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-pkratio.R
\name{PKRatioDataMapping}
\alias{PKRatioDataMapping}
\title{PKRatioDataMapping}
\description{
R6 class for mapping \code{x}, \code{y}, \code{GroupMapping} and \code{pkRatioLines} variables to \code{data}
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{PKRatioDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pkRatioLines}}{numeric vector of ratio limits to plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PKRatioDataMapping$new()}}
\item \href{#method-clone}{\code{PKRatioDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="XYGDataMapping" data-id="checkMapData">}\href{../../tlf/html/XYGDataMapping.html#method-checkMapData}{\code{tlf::XYGDataMapping$checkMapData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{PKRatioDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKRatioDataMapping$new(pkRatioLines = c(1, 1.5, 1/1.5, 2, 1/2), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pkRatioLines}}{numeric vector of ratio limits to plot}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PKRatioDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKRatioDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
