% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkratio-plotconfiguration.R
\name{PKRatioPlotConfiguration}
\alias{PKRatioPlotConfiguration}
\title{PKRatioPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for PK ratio plots
}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{PKRatioPlotConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PKRatioPlotConfiguration$new()}}
\item \href{#method-clone}{\code{PKRatioPlotConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{PKRatioPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKRatioPlotConfiguration$new(
  lines = NULL,
  points = NULL,
  errorbars = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lines}}{\code{ThemeAestheticSelections} object defining properties for PK ratio horizontal lines}

\item{\code{points}}{\code{ThemeAestheticSelections} object defining properties for PK ratio scatter points}

\item{\code{errorbars}}{\code{ThemeAestheticSelections} object defining properties for PK ratio error bars}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PKRatioPlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKRatioPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
