% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkratio-plotconfiguration.R
\name{PKRatioPlotConfiguration}
\alias{PKRatioPlotConfiguration}
\title{PKRatioPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for PK ratio plots
}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{PKRatioPlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pkRatioCaption}}{list of properties for PK ratio plot specific features}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PKRatioPlotConfiguration$new()}}
\item \href{#method-clone}{\code{PKRatioPlotConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="PlotConfiguration" data-id="print">}\href{../../tlf/html/PlotConfiguration.html#method-print}{\code{tlf::PlotConfiguration$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{PKRatioPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKRatioPlotConfiguration$new(
  pkRatioCaption = getDefaultCaptionFor("pkRatio"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pkRatioCaption}}{list of properties for PK ratio plot specific features}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PKRatioPlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKRatioPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
