% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkratio-plotconfiguration.R
\docType{class}
\name{PKRatioPlotConfiguration}
\alias{PKRatioPlotConfiguration}
\title{PKRatioPlotConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PKRatioPlotConfiguration
}
\description{
Plot Configuration for PKRatio
}
\keyword{datasets}
