% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\name{PKRatioPlotConfiguration}
\alias{PKRatioPlotConfiguration}
\title{PKRatioPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for PK ratio plots
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{ObsVsPredPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{tlf::PlotConfiguration} -> \code{PKRatioPlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{defaultYScale}}{Default yAxis scale value when creating a \code{PKRatioPlotConfiguration} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PKRatioPlotConfiguration-clone}{\code{PKRatioPlotConfiguration$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tlf" data-topic="PlotConfiguration" data-id="initialize"><a href='../../tlf/html/PlotConfiguration.html#method-PlotConfiguration-initialize'><code>tlf::PlotConfiguration$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKRatioPlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PKRatioPlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKRatioPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
