% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxwhisker.R
\name{plotBoxWhisker}
\alias{plotBoxWhisker}
\title{plotBoxWhisker}
\usage{
plotBoxWhisker(data, metaData = NULL, dataMapping = NULL,
  plotConfiguration = NULL)
}
\arguments{
\item{data}{data.frame (or list of data.frames? TO BE DISCUSSED)
containing the data to be used for the plot}

\item{metaData}{list of lists (structure TO BE DISCUSSED)
containing complementary information to data (e.g. unit)}

\item{dataMapping}{R6 class BoxWhiskerDataMapping
mapping of x, y axes + mapping of colorGrouping, sizeGrouping, shapeGrouping}

\item{plotConfiguration}{R6 class BoxWhiskerConfiguration
Plot Configuration defining title, subtitle, xlabel, ylabel watermark, and legend}
}
\value{
a ggplot graphical object
}
\description{
plotBoxWhisker(data, metaData, dataMapping, plotConfiguration)
}
\details{
Create a Box Whisker plot
}
