% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\docType{class}
\name{PlotConfiguration}
\alias{PlotConfiguration}
\title{PlotConfiguration}
\usage{
PlotConfiguration
}
\description{
Class for Plot Configuration
}
\section{Fields}{

\describe{
\item{\code{legend}}{\code{LegendConfiguration} R6 class defining legend properties}

\item{\code{xAxis}}{\code{XAxisConfiguration} R6 class defining X-axis properties}

\item{\code{yAxis}}{\code{YAxisConfiguration} R6 class defining Y-axis properties}

\item{\code{background}}{\code{BackgroundConfiguration} R6 class defining background properties}

\item{\code{saveConfiguration}}{\code{SaveConfiguration} R6 class defining save properties}

\item{\code{theme}}{\code{Theme} R6 class defining theme aesthetic properties}
}}

\section{Methods}{

\describe{
\item{new(...)}{Initialize PlotConfiguration}
\item{setPlotProperties(plotObject)}{Apply plot labels properties to plotObject}
\item{setPlotBackground(plotObject)}{Apply background properties to plotObject}
\item{savePlot(plotObject)}{Save plotObject into a file.}
}
}

\keyword{datasets}
