% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\docType{class}
\name{PlotConfiguration}
\alias{PlotConfiguration}
\title{PlotConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PlotConfiguration
}
\description{
Generic Plot Configuration
}
\keyword{datasets}
