% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\name{PlotConfiguration}
\alias{PlotConfiguration}
\title{PlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{labels}}{R6 class \code{LabelConfiguration} defining labels properties}

\item{\code{legend}}{R6 class \code{LegendConfiguration} defining legend properties}

\item{\code{xAxis}}{R6 class \code{XAxisConfiguration} defining X-axis properties}

\item{\code{yAxis}}{R6 class \code{YAxisConfiguration} defining Y-axis properties}

\item{\code{background}}{R6 class \code{BackgroundConfiguration} defining background properties}

\item{\code{export}}{R6 class \code{ExportConfiguration} defining export properties}

\item{\code{theme}}{\code{Theme} R6 class defining theme aesthetic properties}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlotConfiguration$new()}}
\item \href{#method-print}{\code{PlotConfiguration$print()}}
\item \href{#method-clone}{\code{PlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{PlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotConfiguration$new(
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  legend = NULL,
  legendTitle = NULL,
  legendPosition = NULL,
  xAxis = NULL,
  xScale = NULL,
  xLimits = NULL,
  yAxis = NULL,
  yScale = NULL,
  yLimits = NULL,
  background = NULL,
  watermark = NULL,
  export = NULL,
  format = NULL,
  width = NULL,
  height = NULL,
  units = NULL,
  data = NULL,
  metaData = NULL,
  dataMapping = NULL,
  theme = tlfEnv$currentTheme
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{R6 class \code{Label} object}

\item{\code{subtitle}}{R6 class \code{Label} object}

\item{\code{xlabel}}{R6 class \code{Label} object}

\item{\code{ylabel}}{R6 class \code{Label} object}

\item{\code{legend}}{R6 class \code{LegendConfiguration} object defining legend properties}

\item{\code{legendTitle}}{character legend title}

\item{\code{legendPosition}}{character legend position.
Use Enum \code{LegendPositions} to get a list of available to legend positions.}

\item{\code{xAxis}}{R6 class \code{XAxisConfiguration} object defining X-axis properties}

\item{\code{xScale}}{character defining X-axis scale. Use enum \code{Scaling} to access predefined scales.}

\item{\code{xLimits}}{numeric vector of X-axis limits}

\item{\code{yAxis}}{R6 class \code{YAxisConfiguration} object defining X-axis properties}

\item{\code{yScale}}{character defining Y-axis scale. Use enum \code{Scaling} to access predefined scales.}

\item{\code{yLimits}}{numeric vector of Y-axis limits}

\item{\code{background}}{R6 class \code{BackgroundConfiguration} defining background properties}

\item{\code{watermark}}{R6 class \code{Label} object defining watermark background}

\item{\code{export}}{R6 class \code{SaveConfiguration} defining saving properties}

\item{\code{format}}{character defining the format of the file to be saved}

\item{\code{width}}{numeric values defining the width in \code{units} of the plot dimensions after saving}

\item{\code{height}}{numeric values defining the height in \code{units} of the plot dimensions after saving}

\item{\code{units}}{character defining the unit of the saving dimension}

\item{\code{data}}{data.frame used by \code{smartMapping}}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{dataMapping}}{R6 class or subclass \code{XYDataMapping}}

\item{\code{theme}}{R6 class \code{Theme}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print plot configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotConfiguration$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Plot configuration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
