% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ddiratio.R
\name{plotDDIRatio}
\alias{plotDDIRatio}
\title{plotDDIRatio}
\usage{
plotDDIRatio(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  residualsVsObserved = NULL,
  foldDistance = NULL,
  deltaGuest = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{dataMapping}{A \code{DDIRatioDataMapping} object mapping \code{x}, \code{y} and aesthetic groups to their variable names of \code{data}.}

\item{plotConfiguration}{An optional \code{DDIRatioPlotConfiguration} object defining labels, grid, background and watermark.}

\item{residualsVsObserved}{Optional logical value defining
if DDI Ratio plot is drawn as residuals vs observed, instead of predicted vs observed.}

\item{foldDistance}{Numeric values of fold distance lines to display in log plots.
This argument is internally translated into \code{lines} field of \code{dataMapping}.
\strong{Caution}: this argument is meant for log scaled plots and since fold distance is a ratio it is expected positive.
In particular, line of identity corresponds to a \code{foldDistance} of \code{1}.}

\item{deltaGuest}{Numeric value parameter of Guest function}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layer}
}
\value{
A \code{ggplot} object
}
\description{
Producing DDI Ratio plots
}
\examples{
# Produce DDI Ratio plot
ddiData <- data.frame(x = c(1, 2, 1, 2, 3), y = c(5, 0.2, 2, 3, 4))

plotDDIRatio(data = ddiData, dataMapping = DDIRatioDataMapping$new(x = "x", y = "y"))

# Produce DDI Ratio plot with user-defined horizontal lines
plotDDIRatio(
  data = ddiData,
  dataMapping = DDIRatioDataMapping$new(x = "x", y = "y"),
  foldDistance = c(1, 10),
  deltaGuest = 1.25,
  residualsVsObserved = TRUE
)

}
\references{
For examples, see:
\url{https://www.open-systems-pharmacology.org/TLF-Library/articles/pk-ratio-vignette.html}
}
\seealso{
Other molecule plots: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotCumulativeTimeProfile}()},
\code{\link{plotGrid}()},
\code{\link{plotHistogram}()},
\code{\link{plotObsVsPred}()},
\code{\link{plotObservedTimeProfile}()},
\code{\link{plotPKRatio}()},
\code{\link{plotPieChart}()},
\code{\link{plotQQ}()},
\code{\link{plotResVsPred}()},
\code{\link{plotResVsTime}()},
\code{\link{plotSimulatedTimeProfile}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotTornado}()}
}
\concept{molecule plots}
