% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pkratio.R
\name{plotPKRatio}
\alias{plotPKRatio}
\title{plotPKRatio}
\usage{
plotPKRatio(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{dataMapping}{A \code{PKRatioDataMapping} object mapping \code{x}, \code{y} and aesthetic groups to their variable names of \code{data}.}

\item{plotConfiguration}{An optional \code{PKRatioPlotConfiguration} object defining labels, grid, background and watermark.}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layer}
}
\value{
A \code{ggplot} object
}
\description{
Producing PK Ratio plots
}
\examples{
# Produce PK Ratio plot
pkData <- data.frame(x = c(1, 2, 1, 2, 3), y = c(5, 0.2, 2, 3, 4))

plotPKRatio(data = pkData, dataMapping = PKRatioDataMapping$new(x = "x", y = "y"))

}
\references{
For examples, see:
\url{https://www.open-systems-pharmacology.org/TLF-Library/articles/pk-ratio-vignette.html}
}
\seealso{
Other molecule plots: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotDDIRatio}()},
\code{\link{plotHistogram}()},
\code{\link{plotObsVsPred}()},
\code{\link{plotResVsPred}()},
\code{\link{plotResVsTime}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotTornado}()}
}
\concept{molecule plots}
