% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res-vs-pred.R
\name{plotResVsPred}
\alias{plotResVsPred}
\title{plotResVsPred}
\usage{
plotResVsPred(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  smoother = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{dataMapping}{A \code{ResVsPredDataMapping} object mapping \code{x}, \code{y} and aesthetic groups to their variable names of \code{data}.}

\item{plotConfiguration}{An optional \code{ResVsPredConfiguration} object defining labels, grid, background and watermark.}

\item{smoother}{Optional name of smoother function:
\itemize{
\item \code{"loess"} for loess regression
\item \code{"lm"} for linear regression
}}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layer}
}
\value{
A \code{ggplot} object
}
\description{
Producing residuals vs predicted plots
}
\examples{
# Produce Obs vs Pred plot
resVsPredData <- data.frame(x = c(1, 2, 1, 2, 3), y = c(5, 0.2, 2, 3, 4))

plotResVsPred(data = resVsPredData, dataMapping = ResVsPredDataMapping$new(x = "x", y = "y"))

# Produce Res vs Pred plot with linear regression
plotResVsPred(
data = resVsPredData, 
dataMapping = ResVsPredDataMapping$new(x = "x", y = "y"),
smoother = "lm"
)
}
\seealso{
Other molecule plots: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotDDIRatio}()},
\code{\link{plotHistogram}()},
\code{\link{plotObsVsPred}()},
\code{\link{plotPKRatio}()},
\code{\link{plotResVsTime}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotTornado}()}
}
\concept{molecule plots}
