% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res-vs-pred.R
\name{plotResVsPred}
\alias{plotResVsPred}
\title{plotResVsPred}
\usage{
plotResVsPred(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  smoother = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{data.frame containing the data to be used for the plot}

\item{metaData}{list of lists
containing complementary information to data (e.g. their unit and dimension).
This parameter is optional.}

\item{dataMapping}{\code{ObsVsPredDataMapping} class or subclass mapping x and y variables to \code{data} variable names.
\code{dataMapping} provides also the values of the identity and fold errors lines.
This parameter is optional: the \code{tlf} library provides a smart mapping if only \code{data} is provided
and default values of the identity and fold errors lines.}

\item{plotConfiguration}{\code{ObsVsPredConfiguration} class or subclass defining labels, grid, background and watermark
This parameter is optional: the \code{tlf} library provides a default configuration according to the current theme}

\item{smoother}{smoother Name of smoother function: "loess" or "lm"}

\item{plotObject}{\code{ggplot} graphical object to which the Observations vs Predictions plot layer is added
This parameter is optional: the \code{tlf} library will initialize an empty plot if the parameter is NULL or not provided}
}
\value{
A \code{ggplot} graphical object
}
\description{
Add Residuals vs Predictions / Time plot layers to a \code{ggplot} graphical object.
Residuals vs Predictions are plotted as a scatter plot.
}
