% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-timeprofile.R
\name{plotTimeProfile}
\alias{plotTimeProfile}
\title{plotTimeProfile}
\usage{
plotTimeProfile(
  data,
  metaData = NULL,
  dataMapping = NULL,
  observedData = NULL,
  dataMappingForObserved = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{data.frame containing the data to be used for the plot}

\item{metaData}{list containing complementary information to data (e.g. unit)}

\item{dataMapping}{\code{TimeProfileDataMapping} object}

\item{plotConfiguration}{\code{TimeProfilePlotConfiguration} object}

\item{plotObject}{ggplot object, if null creates new plot, if not add time profile layers to ggplot}
}
\value{
a ggplot graphical object
}
\description{
plotTimeProfile(data, metaData = NULL, dataMapping = NULL, plotConfiguration = NULL, plotObject = NULL)
}
