% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-tornado.R
\name{plotTornado}
\alias{plotTornado}
\title{plotTornado}
\usage{
plotTornado(
  data = NULL,
  metaData = NULL,
  x = NULL,
  y = NULL,
  sorted = NULL,
  colorPalette = NULL,
  bar = TRUE,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{A data.frame to use for plot.}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{x}{Numeric values to plot along the \code{x} axis. Only used instead of \code{data} if \code{data} is \code{NULL}.}

\item{y}{Character values to plot along the \code{y} axis. Only used instead of \code{data} if \code{data} is \code{NULL}.}

\item{sorted}{Optional logical value defining if \code{y} values are sorted by absolute values of \code{x}.}

\item{colorPalette}{Optional character values defining a \code{ggplot2} colorPalette (e.g. \code{"Spectral"})}

\item{bar}{Optional logical value setting tornado plot as bar plot instead of scatter plot.}

\item{dataMapping}{A \code{TornadoDataMapping} object mapping \code{x}, \code{y} and aesthetic groups to their variable names of \code{data}.}

\item{plotConfiguration}{An optional \code{TornadoPlotConfiguration} object defining labels, grid, background and watermark.}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layer}
}
\value{
A \code{ggplot} object
}
\description{
Producing tornado plots
}
\examples{
# Produce a tornado plot
plotTornado(x = c(2, -1, 3), y = c("A", "B", "C"))

# Produce a tornado plot as scatter plot
plotTornado(x = c(2, -1, 3), y = c("A", "B", "C"), bar = FALSE)

# Produce a tornado plot as is (no sorting)
plotTornado(x = c(2, -1, 3), y = c("A", "B", "C"), sorted = FALSE)

}
\seealso{
Other molecule plots: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotCumulativeTimeProfile}()},
\code{\link{plotDDIRatio}()},
\code{\link{plotHistogram}()},
\code{\link{plotObsVsPred}()},
\code{\link{plotPKRatio}()},
\code{\link{plotQQ}()},
\code{\link{plotResVsPred}()},
\code{\link{plotResVsTime}()},
\code{\link{plotTimeProfile}()}
}
\concept{molecule plots}
