% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-range.R
\name{RangeDataMapping}
\alias{RangeDataMapping}
\title{RangeDataMapping}
\description{
R6 class for mapping \code{x}, \code{ymin} and \code{ymax} variable to \code{data}
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{RangeDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ymin}}{Name of ymin variable to map}

\item{\code{ymax}}{Name of ymax variable to map}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RangeDataMapping$new()}}
\item \href{#method-checkMapData}{\code{RangeDataMapping$checkMapData()}}
\item \href{#method-clone}{\code{RangeDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{RangeDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RangeDataMapping$new(
  x = NULL,
  ymin = NULL,
  ymax = NULL,
  groupMapping = NULL,
  color = NULL,
  fill = NULL,
  linetype = NULL,
  shape = NULL,
  size = NULL,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{ymin}}{Name of ymin variable to map}

\item{\code{ymax}}{Name of ymax variable to map}

\item{\code{groupMapping}}{R6 class \code{GroupMapping} object}

\item{\code{color}}{R6 class \code{Grouping} object or its input}

\item{\code{fill}}{R6 class \code{Grouping} object or its input}

\item{\code{linetype}}{R6 class \code{Grouping} object or its input}

\item{\code{shape}}{R6 class \code{Grouping} object or its input}

\item{\code{size}}{R6 class \code{Grouping} object or its input}

\item{\code{data}}{data.frame to map used by \code{smartMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{RangeDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkMapData"></a>}}
\subsection{Method \code{checkMapData()}}{
Check that \code{data} variables include map variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RangeDataMapping$checkMapData(data, metaData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}

\item{\code{metaData}}{list containing information on \code{data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with map and \code{defaultAes} variables.
Dummy variable \code{defaultAes} is necessary to allow further modification of plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RangeDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
