% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs-vs-pred-datamapping.R
\name{ResVsPredDataMapping}
\alias{ResVsPredDataMapping}
\title{ResVsPredDataMapping}
\description{
Class for mapping variables in residuals vs predictions/time plot
}
\seealso{
Other DataMapping classes: 
\code{\link{BoxWhiskerDataMapping}},
\code{\link{CumulativeTimeProfileDataMapping}},
\code{\link{DDIRatioDataMapping}},
\code{\link{GroupMapping}},
\code{\link{Grouping}},
\code{\link{HistogramDataMapping}},
\code{\link{ObsVsPredDataMapping}},
\code{\link{ObservedDataMapping}},
\code{\link{PKRatioDataMapping}},
\code{\link{QQDataMapping}},
\code{\link{RangeDataMapping}},
\code{\link{ResVsTimeDataMapping}},
\code{\link{TimeProfileDataMapping}},
\code{\link{TornadoDataMapping}},
\code{\link{XYDataMapping}},
\code{\link{XYGDataMapping}}
}
\concept{DataMapping classes}
\section{Super classes}{
\code{tlf::XYDataMapping} -> \code{tlf::XYGDataMapping} -> \code{tlf::ObsVsPredDataMapping} -> \code{ResVsPredDataMapping}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResVsPredDataMapping-clone}{\code{ResVsPredDataMapping$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tlf" data-topic="ObsVsPredDataMapping" data-id="checkMapData"><a href='../../tlf/html/ObsVsPredDataMapping.html#method-ObsVsPredDataMapping-checkMapData'><code>tlf::ObsVsPredDataMapping$checkMapData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="ObsVsPredDataMapping" data-id="initialize"><a href='../../tlf/html/ObsVsPredDataMapping.html#method-ObsVsPredDataMapping-initialize'><code>tlf::ObsVsPredDataMapping$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResVsPredDataMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResVsPredDataMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResVsPredDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
