% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\name{ResVsPredPlotConfiguration}
\alias{ResVsPredPlotConfiguration}
\title{ResVsPredPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for Res vs Pred/Time plots
}
\section{Super class}{
\code{tlf::PlotConfiguration} -> \code{ResVsPredPlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{defaultSymmetricAxes}}{Default option setting symmetric xAxis and/or yAxis limits when creating a \code{ResVsPredPlotConfiguration} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResVsPredPlotConfiguration-clone}{\code{ResVsPredPlotConfiguration$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tlf" data-topic="PlotConfiguration" data-id="initialize"><a href='../../tlf/html/PlotConfiguration.html#method-PlotConfiguration-initialize'><code>tlf::PlotConfiguration$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResVsPredPlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResVsPredPlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResVsPredPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
