% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\name{ResVsTimePlotConfiguration}
\alias{ResVsTimePlotConfiguration}
\title{ResVsTimePlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for Res vs Pred/Time plots
}
\section{Super classes}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{\link[tlf:ResVsPredPlotConfiguration]{tlf::ResVsPredPlotConfiguration}} -> \code{ResVsTimePlotConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResVsTimePlotConfiguration-clone}{\code{ResVsTimePlotConfiguration$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tlf" data-topic="PlotConfiguration" data-id="initialize"><a href='../../tlf/html/PlotConfiguration.html#method-PlotConfiguration-initialize'><code>tlf::PlotConfiguration$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResVsTimePlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResVsTimePlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResVsTimePlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
