% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-save.R
\name{SaveConfiguration}
\alias{SaveConfiguration}
\title{saveConfiguration}
\description{
R6 class defining the save configuration of a \code{ggplot} object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filename}}{character defining the name of the file to be saved}

\item{\code{width}}{numeric values defining the width in \code{units} of the plot dimensions after saving}

\item{\code{height}}{numeric values defining the height in \code{units} of the plot dimensions after saving}

\item{\code{units}}{character defining the unit of the saving dimension}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SaveConfiguration$new()}}
\item \href{#method-print}{\code{SaveConfiguration$print()}}
\item \href{#method-savePlot}{\code{SaveConfiguration$savePlot()}}
\item \href{#method-clone}{\code{SaveConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{SaveConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveConfiguration$new(
  filename = "tlf-plot.png",
  width = 20,
  height = 12,
  units = "cm",
  format = NULL,
  path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{character defining the name of the file to be saved}

\item{\code{width}}{numeric values defining the width in \code{units} of the plot dimensions after saving}

\item{\code{height}}{numeric values defining the height in \code{units} of the plot dimensions after saving}

\item{\code{units}}{character defining the unit of the saving dimension}

\item{\code{format}}{character defining the name of the file to be saved.
If \code{NULL} use "png" format.}

\item{\code{path}}{character defining the path where to save the file.
If \code{NULL} use current path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SaveConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print file save properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveConfiguration$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-savePlot"></a>}}
\subsection{Method \code{savePlot()}}{
Save \code{ggplot} object into a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveConfiguration$savePlot(plotObject)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{a \code{ggplot} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaveConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
