% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\docType{data}
\name{Scaling}
\alias{Scaling}
\title{How should comparison of entities be performed}
\format{An object of class \code{list} of length 2.}
\usage{
Scaling
}
\description{
How should comparison of entities be performed
}
\keyword{datasets}
