% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\docType{data}
\name{Scaling}
\alias{Scaling}
\title{How should comparison of entities be performed}
\format{An object of class \code{list} of length 21.}
\usage{
Scaling
}
\description{
Built-in transformations include
"asn", "atanh", "boxcox", "date", "exp", "hms", "identity",
"log", "log10", "log1p", "log2", "logit", "modulus", "probability",
"probit", "pseudo_log", "reciprocal", "reverse", "sqrt" and "time"
}
\keyword{datasets}
