% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\docType{data}
\name{Scaling}
\alias{Scaling}
\title{Scaling}
\format{An object of class \code{list} of length 8.}
\usage{
Scaling
}
\description{
Pre-defined transformation of axes
Not that built-in transformations from ggplot2 includes more transformations
}
\keyword{datasets}
