% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enums.R
\docType{data}
\name{Scaling}
\alias{Scaling}
\title{Scaling}
\format{
An object of class \code{list} of length 9.
}
\usage{
Scaling
}
\description{
Helper enum of predefined transformations of axes Note that the
transformations will be translated internally into \code{ggplot2}
transformations. \code{ggplot2} includes more transformations than what is
available in this enum.
}
\examples{
# Continuous linear/identity scale
Scaling$identity
Scaling$lin

# Continuous log10 scale
Scaling$log

# Continuous natural logarithm (ln) scale (base is *e*)
Scaling$ln

# Discrete scale for categrical data such as boxplot and tornado plot data
Scaling$discrete

# Reverse continuous linear scale to switch end and beginning of linear scale
Scaling$reverse

# Continusous square root scale
Scaling$sqrt

# Time scale for POSIXlt or POSIXct data
Scaling$time

# Date scale for POSIXlt or POSIXct data
Scaling$date

}
\seealso{
Other enum helpers: 
\code{\link{AestheticFields}},
\code{\link{AestheticProperties}},
\code{\link{AestheticSelectionKeys}},
\code{\link{Alignments}},
\code{\link{AtomPlots}},
\code{\link{ColorMaps}},
\code{\link{DataMappings}},
\code{\link{DefaultDataMappingValues}},
\code{\link{Directions}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{LegendTypes}},
\code{\link{Linetypes}},
\code{\link{MoleculePlots}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{PlotConfigurations}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{TickLabelTransforms}},
\code{\link{VerticalJustification}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
