% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-background.R
\name{setBackgroundPlotArea}
\alias{setBackgroundPlotArea}
\title{setBackgroundPlotArea}
\usage{
setBackgroundPlotArea(
  plotObject,
  fill = NULL,
  color = NULL,
  linetype = NULL,
  size = NULL
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{fill}{Optional character values defining the color of the background.
See \code{grDevices::colors()} to get names of colors}

\item{color}{Optional character values defining the color of the background frame.
See \code{grDevices::colors()} to get names of colors}

\item{linetype}{Optional character values defining the linetype of the background frame.
See enum \code{Linetypes} to get names of linetype.}

\item{size}{Optional numeric values defining the size of the background frame.}
}
\value{
A \code{ggplot} object
}
\description{
Set background plot area properties of a \code{ggplot} object
}
\examples{
# Set background of a scatter plot
p <- addScatter(x = c(1, 2, 1, 2, 3), y = c(5, 0, 2, 3, 4))

setBackgroundPlotArea(p, fill = "yellowgreen", color = "red", linetype = "dotted")

}
