% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tlf.R
\name{setDefaultAggregationBins}
\alias{setDefaultAggregationBins}
\title{setDefaultAggregationBins}
\usage{
setDefaultAggregationBins(bins = NULL)
}
\arguments{
\item{bins}{Number of bins if value, edges if vector or binning function if function}
}
\description{
Set default aggregation bins of tlf environment
}
\examples{
# Set default number of bins
plotHistogram(x = rnorm(1000))

setDefaultAggregationBins(21)
plotHistogram(x = rnorm(1000))

}
