% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultExportParameters}
\alias{setDefaultExportParameters}
\title{setDefaultExportParameters}
\usage{
setDefaultExportParameters(
  format = NULL,
  width = NULL,
  height = NULL,
  units = NULL,
  dpi = NULL,
  name = NULL
)
}
\arguments{
\item{format}{file format of the exported plots}

\item{width}{plot width in \code{unit}}

\item{height}{plot height in \code{unit}}

\item{units}{units of \code{width} and \code{height}}

\item{dpi}{units of \code{width} and \code{height}}

\item{name}{base file name of the exported plots}
}
\description{
Set default tlf properties for exporting/saving plots
}
