% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultLegendTitle}
\alias{setDefaultLegendTitle}
\title{setDefaultLegendTitle}
\usage{
setDefaultLegendTitle(title)
}
\arguments{
\item{title}{Character or \code{Label} object}
}
\description{
Set default legend title of tlf environment
}
