% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf-env.R
\name{setDefaultWatermark}
\alias{setDefaultWatermark}
\title{setDefaultWatermark}
\usage{
setDefaultWatermark(watermark = NULL)
}
\arguments{
\item{watermark}{character or Label class object}
}
\description{
Set default watermark value for current theme
}
