% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegend}
\alias{setLegend}
\title{SetLegend}
\usage{
setLegend(plotHandle, legendPosition, legendColorCaptions = NULL,
  legendSizeCaptions = NULL, legendShapeCaptions = NULL)
}
\arguments{
\item{plotHandle}{Graphical object created from ggplot}

\item{legendPosition}{element of legendPositions list}

\item{legendColorCaptions}{captions of the legend}
}
\value{
Current plot with Legend Layer
}
\description{
SetLegend set a Legend Position and labels in a plot
}
