% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-utils.R, R/utilities-legend.R
\name{setLegend}
\alias{setLegend}
\title{SetLegendPosition}
\usage{
setLegend(plotHandle, legendPosition, legendColorCaptions = NULL,
  legendSizeCaptions = NULL, legendShapeCaptions = NULL)

setLegend(plotHandle, legendPosition, legendColorCaptions = NULL,
  legendSizeCaptions = NULL, legendShapeCaptions = NULL)
}
\arguments{
\item{plotHandle}{Graphical object created from ggplot}

\item{legendPosition}{element of legendPositions list}

\item{legendColorCaptions}{captions of the legend}

\item{plotConfiguration}{plot Configuration for legend}

\item{dataMapping}{mapping (might be overwritten to use plotConfiguration instead)}

\item{plotHandle}{Graphical object created from ggplot}
}
\value{
Current plot with Legend Layer

Current plot with Legend Layer
}
\description{
SetLegend set a Legend Position and labels in a plot

SetLegend set a Legend Position and labels in a plot
}
