% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-utils.R
\name{setLegend}
\alias{setLegend}
\title{SetLegendPosition}
\usage{
setLegend(plotHandle, plotConfiguration, dataMapping)
}
\arguments{
\item{plotHandle}{Graphical object created from ggplot}

\item{plotConfiguration}{plot Configuration for legend}

\item{dataMapping}{mapping (might be overwritten to use plotConfiguration instead)}
}
\value{
Current plot with Legend Layer
}
\description{
SetLegend set a Legend Position and labels in a plot
}
