% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegend}
\alias{setLegend}
\title{setLegend}
\usage{
setLegend(
  plotObject,
  position = NULL,
  title = NULL,
  font = NULL,
  caption = NULL
)
}
\arguments{
\item{plotObject}{Graphical object created from ggplot}

\item{position}{legend position.
Use enum \code{LegendPositions} to access the list of legend positions.}

\item{title}{character or \code{Label} object}

\item{font}{\code{Font} object defining legend font}

\item{caption}{data.frame containing the caption properties of the legend}
}
\value{
A \code{ggplot} object
}
\description{
Set legend position, title, font and/or caption
}
