% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegend}
\alias{setLegend}
\title{SetLegend}
\usage{
setLegend(plotObject, legendPosition, legendTitles = NULL,
  legendCaptions = NULL, legendValues = NULL)
}
\arguments{
\item{plotObject}{Graphical object created from ggplot}

\item{legendPosition}{element of legendPositions list}

\item{legendCaptions}{captions of the legend (as a list with LegendTypes attribute)}

\item{legendValues}{values of the legend aesthetic attributes}
}
\value{
Current plot with Legend Layer
}
\description{
SetLegend set a Legend Position and labels in a plot
}
