% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegendPosition}
\alias{setLegendPosition}
\title{SetLegendPosition}
\usage{
setLegendPosition(plotObject,
  legendPosition = LegendPositions$outsideRight, Subplot.Index = NA)
}
\arguments{
\item{legendPosition}{element of legendPositions list}

\item{Subplot.Index}{(optional) OPTION NOT IMPLEMENTED YET}

\item{plotHandle}{Graphical object created from ggplot}
}
\value{
Current plot with Legend Layer
}
\description{
SetLegendPosition set a Legend Position as defined by legendPosition key
Default legend setting is within plot on the top right corner
}
