% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegendPosition}
\alias{setLegendPosition}
\title{setLegendPosition}
\usage{
setLegendPosition(plotObject, position = NULL)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{position}{legend position as defined in helper enum \code{LegendPositions}}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the legend position
}
\seealso{
LegendPositions
}
