% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegendPosition}
\alias{setLegendPosition}
\title{setLegendTitle}
\usage{
setLegendPosition(plotObject, position)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{position}{legend position.
Use enum \code{LegendPositions} to access the list of legend positions.}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the legend position
}
