% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLegend.R
\name{setLegendPosition}
\alias{setLegendPosition}
\title{SetLegendPosition}
\usage{
setLegendPosition(PlotHandle, Location = "inside", X.Location = NA,
  Y.Location = NA, Subplot.Index = NA)
}
\arguments{
\item{PlotHandle}{Graphical object created from ggplot}

\item{Location}{Inside/Oustide/None}

\item{X.Location}{NA/Left/Center/Right}

\item{Y.Location}{NA/Top/Center/Bottom}

\item{Subplot.Index}{(optional) OPTION NOT IMPLEMENTED YET}
}
\value{
Current plot with Legend Layer
}
\description{
SetLegendPosition set a Legend Position in a plot (case insensitive)
Default legend setting is within plot on the top right corner
}
