% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegendTitle}
\alias{setLegendTitle}
\title{setLegendTitle}
\usage{
setLegendTitle(plotObject, title = NULL)
}
\arguments{
\item{plotObject}{ggplot object}

\item{title}{character or \code{Label} object}
}
\value{
A ggplot object
}
\description{
Set legend title
}
