% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegendTitle}
\alias{setLegendTitle}
\title{setLegendTitle}
\usage{
setLegendTitle(plotObject, title)
}
\arguments{
\item{plotObject}{\code{ggplot} graphical object}

\item{title}{title of legend}
}
\value{
A \code{ggplot} graphical object
}
\description{
Set the legend title
}
