% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-legend.R
\name{setLegendTitle}
\alias{setLegendTitle}
\title{setLegendTitle}
\usage{
setLegendTitle(
  plotObject,
  title = NULL,
  color = NULL,
  size = NULL,
  fontFace = NULL,
  angle = NULL
)
}
\arguments{
\item{plotObject}{ggplot object}

\item{title}{character or \code{Label} object}

\item{color}{color of legend font}

\item{size}{size of legend font}

\item{fontFace}{color of legend font}

\item{angle}{angle of legend font}
}
\value{
A ggplot object
}
\description{
Set legend title
}
