% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-export.R
\name{setPlotExport}
\alias{setPlotExport}
\title{setPlotExport}
\usage{
setPlotExport(
  plotObject,
  name = NULL,
  format = NULL,
  width = NULL,
  height = NULL,
  units = NULL,
  dpi = NULL
)
}
\arguments{
\item{plotObject}{Graphical object created from ggplot}

\item{name}{character defining the name of the file to be saved (without extension)}

\item{format}{character defining the format of the file to be saved.}

\item{width}{numeric values defining the width in \code{units} of the plot dimensions after saving}

\item{height}{numeric values defining the height in \code{units} of the plot dimensions after saving}

\item{units}{character defining the unit of the saving dimension}

\item{dpi}{numeric value defining plot resolution (dots per inch)}
}
\value{
ggplot object with updated saving properties
}
\description{
Set plot export properties
}
