% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-label.R
\name{setPlotLabels}
\alias{setPlotLabels}
\title{setPlotLabels}
\usage{
setPlotLabels(
  plotObject,
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  caption = NULL
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{title}{A character value or \code{Label} object}

\item{subtitle}{A character value or \code{Label} object}

\item{xlabel}{A character value or \code{Label} object}

\item{ylabel}{A character value or \code{Label} object}

\item{caption}{A character value or \code{Label} object}
}
\value{
A \code{ggplot} object
}
\description{
Set labels properties on a ggplot object
}
\examples{
# Set labels of a scatter plot
p <- addScatter(x = c(1, 2, 1, 2, 3), y = c(5, 0, 2, 3, 4))

setPlotLabels(p, xlabel = "new x label", ylabel = "new y label")

# Set labels using Label object
setPlotLabels(p, ylabel = Label$new(text = "red y label", color = "red"))

}
