% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-label.R
\name{setPlotLabels}
\alias{setPlotLabels}
\title{setPlotLabels}
\usage{
setPlotLabels(
  plotObject,
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL
)
}
\arguments{
\item{plotObject}{ggplot object to set}

\item{title}{character or \code{Label} object}

\item{subtitle}{character or \code{Label} object}

\item{xlabel}{character or \code{Label} object}

\item{ylabel}{character or \code{Label} object}
}
\value{
ggplot object
}
\description{
Set labels properties on a ggplot object
}
