% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-background.R
\name{setWatermark}
\alias{setWatermark}
\title{setWatermark}
\usage{
setWatermark(plotObject, watermark = NULL, angle = 30, alpha = 0.4)
}
\arguments{
\item{plotObject}{ggplot object to which the watermark is set}

\item{watermark}{character or Label class object}

\item{angle}{Angle in degree from horizontal of the watermark label. Default angle is 30 degrees.}

\item{alpha}{Transparency of the watermark label.
Alpha is a numeric between 0 and 1: 0 label is totally transparent, 1 label is totally opaque.
Default alpha is 0.6.}
}
\value{
ggplot object to which the watermark is added.
}
\description{
setWatermark
}
