% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark-utils.R
\name{setWatermark}
\alias{setWatermark}
\title{setWatermark}
\usage{
setWatermark(plotHandle, label, angle = 30)
}
\arguments{
\item{plotHandle}{ggplot object}

\item{label}{to be written in watermark (class Label and character allowed)}

\item{angle}{(optional argument) angle of the watermark}
}
\value{
plotHandle ggplot object with watermark background as a grob
}
\description{
setWatermark set watermark as defined by label
}
