% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-watermark.R
\name{setWatermark}
\alias{setWatermark}
\title{setWatermark}
\usage{
setWatermark(plotObject, label, angle = 30, alpha = 0.4)
}
\arguments{
\item{plotObject}{ggplot object}

\item{label}{to be written in watermark (class Label and character allowed)}

\item{angle}{(optional argument) angle of the watermark}

\item{alpha}{(optional argument) transparency of the watermark}
}
\value{
plotObject ggplot object with watermark background as a grob
}
\description{
setWatermark set watermark as defined by label
}
