% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-background.R
\name{setWatermark}
\alias{setWatermark}
\title{setWatermark}
\usage{
setWatermark(
  plotObject,
  watermark = NULL,
  color = NULL,
  size = NULL,
  angle = NULL,
  alpha = NULL
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}

\item{watermark}{A character value or a \code{Label} object}

\item{color}{Color of the watermark.}

\item{size}{Size of the watermark.}

\item{angle}{Angle of the watermark (in degree).}

\item{alpha}{Numeric value between 0 and 1 corresponding to transparency of the watermark
The closer to 0, the more transparent the watermark is.
The closer to 1, the more opaque the watermark is.}
}
\value{
A \code{ggplot} object
}
\description{
Set the watermark of a \code{ggplot} object.
Unlike \code{addWatermark}, the watermark layer is overridden by \code{setWatermark}.
}
\examples{
# Add a watermark to an empty plot
p <- initializePlot()
setWatermark(p, "watermark")

# Watermark with font properties
watermarkLabel <- Label$new(text = "watermark", color = "blue")
setWatermark(p, watermarkLabel)

# Horizontal watermark
setWatermark(p, watermarkLabel, angle = 0)

# Watermark totally opaque
setWatermark(p, watermarkLabel, alpha = 1)

}
