% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{setXAxis}
\alias{setXAxis}
\title{setXAxis}
\usage{
setXAxis(
  plotObject,
  scale = NULL,
  valuesLimits = NULL,
  axisLimits = NULL,
  limits = lifecycle::deprecated(),
  ticks = NULL,
  ticklabels = NULL,
  minorTicks = NULL,
  font = NULL,
  expand = NULL
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object to set X-axis properties}

\item{scale}{Scale of axis. Use enum \code{Scaling} to access names of scales.}

\item{valuesLimits}{Optional numeric values of values limits}

\item{axisLimits}{Optional numeric values of axis limits}

\item{limits}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. Replaced by axisLimits argument.}

\item{ticks}{Optional values or function for axis ticks}

\item{ticklabels}{Optional values or function for axis ticklabels}

\item{minorTicks}{Optional values or function for axis minor ticks}

\item{font}{A \code{Font} object  defining font of ticklabels}

\item{expand}{Logical defining if data is expanded until axis}
}
\value{
A \code{ggplot} object
}
\description{
Set X-axis properties of a \code{ggplot} object
}
\examples{
myPlot <- addLine(x = c(1, 2, 3), y = c(10, 50, 100))

# Set x-axis in log scale
setXAxis(myPlot, scale = Scaling$log)

# Set x-axis ticklabels to Greek letters
setXAxis(myPlot, ticks = c(1, 2, 3), ticklabels = parse(text = c("alpha", "beta", "gamma")))

# Set x-axis limits
setXAxis(myPlot, axisLimits = c(1, 2.5))

# Set x-axis fonts
setXAxis(myPlot, font = Font$new(color = "blue", size = 14))
}
