% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{setXAxis}
\alias{setXAxis}
\title{setXAxis}
\usage{
setXAxis(
  plotObject,
  scale = NULL,
  limits = NULL,
  ticks = NULL,
  ticklabels = NULL
)
}
\arguments{
\item{plotObject}{ggplot object to set X-axis properties}

\item{scale}{Scale of X-axis. Use enum \code{Scaling} to access names of scales.}

\item{limits}{X-axis limits}

\item{ticks}{X-axis ticks}

\item{ticklabels}{X-axis ticklabels}
}
\value{
ggplot object with updated X-axis
}
\description{
Set X-axis properties on a ggplot object
}
