% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{setYAxis}
\alias{setYAxis}
\title{setYAxis}
\usage{
setYAxis(
  plotObject,
  scale = NULL,
  limits = NULL,
  ticks = NULL,
  ticklabels = NULL,
  font = NULL
)
}
\arguments{
\item{plotObject}{ggplot object to set Y-axis properties}

\item{scale}{Scale of Y-axis. Use enum \code{Scaling} to access names of scales.}

\item{limits}{Y-axis limits}

\item{ticks}{Y-axis ticks}

\item{ticklabels}{Y-axis ticklabels}

\item{font}{\code{Font} object  defining font of ticklabels}
}
\value{
ggplot object with updated Y-axis
}
\description{
Set Y-axis properties on a ggplot object
}
\examples{
myPlot <- addLine(x = c(1, 2, 3), y = c(10, 50, 100))

# Set y-axis in log scale
setYAxis(myPlot, scale = Scaling$log)

# Set y-axis ticklabels to Greek letters
setYAxis(myPlot, ticks = c(10, 50, 100), ticklabels = parse(text = c("alpha", "beta", "gamma")))

# Set y-axis limits
setYAxis(myPlot, limits = c(10, 75))

# Set y-axis fonts
setYAxis(myPlot, font = Font$new(color = "blue", size = 14))
}
