% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-axis.R
\name{setYAxis}
\alias{setYAxis}
\title{setYAxis}
\usage{
setYAxis(
  plotObject,
  scale = NULL,
  limits = NULL,
  ticks = NULL,
  ticklabels = NULL,
  minorTicks = NULL,
  font = NULL,
  expand = NULL
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object to set X-axis properties}

\item{scale}{Scale of axis. Use enum \code{Scaling} to access names of scales.}

\item{limits}{Optional numeric values of axis limits}

\item{ticks}{Optional values or function for axis ticks}

\item{ticklabels}{Optional values or function for axis ticklabels}

\item{minorTicks}{Optional values or function for axis minor ticks}

\item{font}{A \code{Font} object  defining font of ticklabels}

\item{expand}{Logical defining if data is expanded until axis}
}
\value{
A \code{ggplot} object
}
\description{
Set Y-axis properties of a \code{ggplot} object
}
\examples{
myPlot <- addLine(x = c(1, 2, 3), y = c(10, 50, 100))

# Set y-axis in log scale
setYAxis(myPlot, scale = Scaling$log)

# Set y-axis ticklabels to Greek letters
setYAxis(myPlot, ticks = c(10, 50, 100), ticklabels = parse(text = c("alpha", "beta", "gamma")))

# Set y-axis limits
setYAxis(myPlot, limits = c(10, 75))

# Set y-axis fonts
setYAxis(myPlot, font = Font$new(color = "blue", size = 14))
}
