% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\docType{data}
\name{Shapes}
\alias{Shapes}
\title{Shapes}
\format{
An object of class \code{list} of length 9.
}
\usage{
Shapes
}
\description{
List of some \code{ggplot2} shapes
}
\keyword{datasets}
