% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aesthetics.R
\docType{data}
\name{Shapes}
\alias{Shapes}
\title{Shapes}
\format{
An object of class \code{list} of length 38.
}
\usage{
Shapes
}
\description{
List of some \code{ggplot2} shapes
}
\seealso{
Other enum helpers: 
\code{\link{AestheticProperties}},
\code{\link{LegendPositions}},
\code{\link{Linetypes}},
\code{\link{Scaling}}
}
\concept{enum helpers}
\keyword{datasets}
