% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{class}
\name{Theme}
\alias{Theme}
\title{Theme}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Theme
}
\description{
Theme class sets of font properties for texts
}
\keyword{datasets}
