% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{Theme}
\alias{Theme}
\title{Theme}
\description{
R6 class defining theme
}
\section{Super class}{
\code{\link[tlf:ThemeFont]{tlf::ThemeFont}} -> \code{Theme}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{watermark}}{list of font properties for watermark}

\item{\code{background}}{list of aesthetic properties for background configuration}

\item{\code{aesProperties}}{list of aesthetic properties for plots in general}

\item{\code{defaultCaption}}{aesthetic properties for specific plots}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Theme$new()}}
\item \href{#method-clone}{\code{Theme$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Theme} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$new(
  themesProperties = tlfEnvThemesProperties$default,
  labelColors = NULL,
  labelBaseSize = 14,
  background = NULL,
  watermark = NULL,
  aesProperties = NULL,
  defaultCaption = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{themesProperties}}{list of aesthetic properties}

\item{\code{labelColors}}{list of colors for each label}

\item{\code{labelBaseSize}}{numeric value for theme base size}

\item{\code{background}}{list of aesthetic properties for background configuration}

\item{\code{watermark}}{list of font properties for watermark}

\item{\code{aesProperties}}{list of aesthetic properties for plots in general}

\item{\code{defaultCaption}}{list of aesthetic properties for specific plot features}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Theme} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Theme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
