% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{class}
\name{ThemeAesProperties}
\alias{ThemeAesProperties}
\title{ThemeAesProperties}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ThemeAesProperties
}
\description{
Theme aesthetics class to set aesthetic properties plots
}
\keyword{datasets}
