% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{class}
\name{ThemeFont}
\alias{ThemeFont}
\title{ThemeFont}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ThemeFont
}
\description{
ThemeFont class sets of font properties for texts
}
\keyword{datasets}
