% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ThemeFont}
\alias{ThemeFont}
\title{ThemeFont}
\description{
R6 class defining theme font properties
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{\code{Font} object for font properties title}

\item{\code{subtitle}}{\code{Font} object for font properties of subtitle}

\item{\code{xlabel}}{\code{Font} object for font properties of xlabel}

\item{\code{ylabel}}{\code{Font} object for font properties of ylabel}

\item{\code{caption}}{\code{Font} object for font properties of caption}

\item{\code{watermark}}{\code{Font} object for font properties of watermark}

\item{\code{legendTitle}}{\code{Font} object for font properties of legend title}

\item{\code{legend}}{\code{Font} object for font properties of legend}

\item{\code{xAxis}}{\code{Font} object for font properties of xAxis}

\item{\code{yAxis}}{\code{Font} object for font properties of yAxis}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ThemeFont$new()}}
\item \href{#method-toJson}{\code{ThemeFont$toJson()}}
\item \href{#method-clone}{\code{ThemeFont$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ThemeFont} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeFont$new(
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  caption = NULL,
  watermark = NULL,
  legendTitle = NULL,
  legend = NULL,
  xAxis = NULL,
  yAxis = NULL,
  baseColor = "black",
  baseSize = 12,
  baseFace = "plain",
  baseFamily = "",
  baseAngle = 0,
  baseAlign = "center"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{\code{Font} object or list for font properties title}

\item{\code{subtitle}}{\code{Font} object or list for font properties of subtitle}

\item{\code{xlabel}}{\code{Font} object or list for font properties of xlabel}

\item{\code{ylabel}}{\code{Font} object or list for font properties of ylabel}

\item{\code{caption}}{\code{Font} object or list for font properties of caption}

\item{\code{watermark}}{\code{Font} object or list for font properties of watermark}

\item{\code{legendTitle}}{\code{Font} object or list for font properties of legend title}

\item{\code{legend}}{\code{Font} object or list for font properties of legend}

\item{\code{xAxis}}{\code{Font} object or list for font properties of xAxis}

\item{\code{yAxis}}{\code{Font} object or list for font properties of yAxis}

\item{\code{baseColor}}{name of base color of undefined fonts. Default is "black".}

\item{\code{baseSize}}{base size of undefined fonts. Default is 12.}

\item{\code{baseFace}}{name of base face of undefined fonts. Default is "plain".}

\item{\code{baseFamily}}{name of base family of undefined fonts. Default is "".}

\item{\code{baseAngle}}{base angle of undefined fonts. Default is 0 degree.}

\item{\code{baseAlign}}{base alignment of undefined fonts. Default is "center".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ThemeFont} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJson"></a>}}
\if{latex}{\out{\hypertarget{method-toJson}{}}}
\subsection{Method \code{toJson()}}{
Translate object into a json list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeFont$toJson()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list that can be saved into a json file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeFont$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
