% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ThemePlotConfigurations}
\alias{ThemePlotConfigurations}
\title{ThemePlotConfigurations}
\description{
R6 class defining theme of plot configuration objects
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{addScatter}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addScatter()}}

\item{\code{addLine}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addLine()}}

\item{\code{addRibbon}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addRibbon()}}

\item{\code{addErrorbar}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addErrorbar()}}

\item{\code{plotPKRatio}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotPKRatio()}}

\item{\code{plotDDIRatio}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotDDIRatio()}}

\item{\code{plotTimeProfile}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotTimeProfile()}}

\item{\code{plotObsVsPred}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotObsVsPred()}}

\item{\code{plotBoxWhisker}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotBoxWhisker()}}

\item{\code{plotTornado}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotTornado()}}

\item{\code{plotHistogram}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotHistogram()}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ThemePlotConfigurations-new}{\code{ThemePlotConfigurations$new()}}
\item \href{#method-ThemePlotConfigurations-toJson}{\code{ThemePlotConfigurations$toJson()}}
\item \href{#method-ThemePlotConfigurations-clone}{\code{ThemePlotConfigurations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemePlotConfigurations-new"></a>}}
\if{latex}{\out{\hypertarget{method-ThemePlotConfigurations-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ThemePlotConfigurations} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemePlotConfigurations$new(
  addScatter = NULL,
  addLine = NULL,
  addRibbon = NULL,
  addErrorbar = NULL,
  plotPKRatio = NULL,
  plotDDIRatio = NULL,
  plotTimeProfile = NULL,
  plotObsVsPred = NULL,
  plotBoxWhisker = NULL,
  plotTornado = NULL,
  plotHistogram = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{addScatter}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addScatter()}}

\item{\code{addLine}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addLine()}}

\item{\code{addRibbon}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addRibbon()}}

\item{\code{addErrorbar}}{theme properties for \code{PlotConfiguration} objects as used in function \code{addErrorbar()}}

\item{\code{plotPKRatio}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotPKRatio()}}

\item{\code{plotDDIRatio}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotDDIRatio()}}

\item{\code{plotTimeProfile}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotTimeProfile()}}

\item{\code{plotObsVsPred}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotObsVsPred()}}

\item{\code{plotBoxWhisker}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotBoxWhisker()}}

\item{\code{plotTornado}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotTornado()}}

\item{\code{plotHistogram}}{theme properties for \code{PlotConfiguration} objects as used in function \code{plotHistogram()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ThemePlotConfigurations} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemePlotConfigurations-toJson"></a>}}
\if{latex}{\out{\hypertarget{method-ThemePlotConfigurations-toJson}{}}}
\subsection{Method \code{toJson()}}{
Translate object into a json list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemePlotConfigurations$toJson()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list that can be saved into a json file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThemePlotConfigurations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ThemePlotConfigurations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemePlotConfigurations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
