% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enums.R
\docType{data}
\name{TickLabelTransforms}
\alias{TickLabelTransforms}
\title{TickLabelTransforms}
\format{
An object of class \code{list} of length 9.
}
\usage{
TickLabelTransforms
}
\description{
List of all available tick label transformation names
}
\seealso{
Other enum helpers: 
\code{\link{AestheticFields}},
\code{\link{AestheticProperties}},
\code{\link{AestheticSelectionKeys}},
\code{\link{Alignments}},
\code{\link{AtomPlots}},
\code{\link{ColorMaps}},
\code{\link{ColorPalettes}},
\code{\link{DataMappings}},
\code{\link{DefaultDataMappingValues}},
\code{\link{Directions}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LegendPositions}},
\code{\link{LegendTypes}},
\code{\link{Linetypes}},
\code{\link{MoleculePlots}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{PlotConfigurations}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{VerticalJustification}},
\code{\link{tlfSettingNames}},
\code{\link{tlfStatFunctions}}
}
\concept{enum helpers}
\keyword{datasets}
