% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-timeprofile.R
\docType{class}
\name{TimeProfileDataMapping}
\alias{TimeProfileDataMapping}
\title{TimeProfileDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TimeProfileDataMapping
}
\description{
Data Mapping for Time Profile
}
\keyword{datasets}
