% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-timeprofile.R
\name{TimeProfileDataMapping}
\alias{TimeProfileDataMapping}
\title{TimeProfileDataMapping}
\description{
R6 class defining the configuration of a \code{ggplot} object for time profile plot
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{TimeProfileDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lloq}}{numeric value of lower limit of quantification}

\item{\code{isRangeTimeProfile}}{logical to set aggregation of data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TimeProfileDataMapping$new()}}
\item \href{#method-clone}{\code{TimeProfileDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="XYGDataMapping" data-id="checkMapData">}\href{../../tlf/html/XYGDataMapping.html#method-checkMapData}{\code{tlf::XYGDataMapping$checkMapData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{TimeProfileDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$new(lloq = NULL, isRangeTimeProfile = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lloq}}{numeric value of lower limit of quantification}

\item{\code{isRangeTimeProfile}}{Name of x variable to map}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TimeProfileDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
