% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-timeprofile.R
\name{TimeProfileDataMapping}
\alias{TimeProfileDataMapping}
\title{TimeProfileDataMapping}
\description{
R6 class defining the configuration of a \code{ggplot} object for time profile plot
}
\section{Super classes}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{\link[tlf:XYGDataMapping]{tlf::XYGDataMapping}} -> \code{TimeProfileDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ymin}}{Name of ymin variable to map}

\item{\code{ymax}}{Name of ymax variable to map}

\item{\code{LLOQ}}{values of lower limit of quatification}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TimeProfileDataMapping$new()}}
\item \href{#method-checkMapData}{\code{TimeProfileDataMapping$checkMapData()}}
\item \href{#method-clone}{\code{TimeProfileDataMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{TimeProfileDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$new(
  LLOQ = NULL,
  x,
  y = NULL,
  ymin = NULL,
  ymax = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LLOQ}}{values of lower limit of quatification}

\item{\code{x}}{Name of x variable to map}

\item{\code{y}}{Name of y variable to map}

\item{\code{ymin}}{Name of ymin variable to map}

\item{\code{ymax}}{Name of ymax variable to map}

\item{\code{...}}{parameters inherited from \code{XYGDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TimeProfileDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkMapData"></a>}}
\subsection{Method \code{checkMapData()}}{
Check that \code{data} variables include map variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$checkMapData(data, metaData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}

\item{\code{metaData}}{list containing information on \code{data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with map and \code{defaultAes} variables.
Dummy variable \code{defaultAes} is necessary to allow further modification of plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
