% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeprofile-datamapping.R
\name{TimeProfileDataMapping}
\alias{TimeProfileDataMapping}
\title{TimeProfileDataMapping}
\description{
R6 class defining the configuration of a \code{ggplot} object for time profile plot
}
\seealso{
Other DataMapping classes: 
\code{\link{BoxWhiskerDataMapping}},
\code{\link{CumulativeTimeProfileDataMapping}},
\code{\link{DDIRatioDataMapping}},
\code{\link{GroupMapping}},
\code{\link{Grouping}},
\code{\link{HistogramDataMapping}},
\code{\link{ObsVsPredDataMapping}},
\code{\link{ObservedDataMapping}},
\code{\link{PKRatioDataMapping}},
\code{\link{PieChartDataMapping}},
\code{\link{QQDataMapping}},
\code{\link{RangeDataMapping}},
\code{\link{ResVsPredDataMapping}},
\code{\link{ResVsTimeDataMapping}},
\code{\link{TornadoDataMapping}},
\code{\link{XYDataMapping}},
\code{\link{XYGDataMapping}}
}
\concept{DataMapping classes}
\section{Super classes}{
\code{tlf::XYDataMapping} -> \code{tlf::XYGDataMapping} -> \code{tlf::RangeDataMapping} -> \code{TimeProfileDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{y2Axis}}{Name of y2Axis variable to map}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeProfileDataMapping-new}{\code{TimeProfileDataMapping$new()}}
\item \href{#method-TimeProfileDataMapping-checkMapData}{\code{TimeProfileDataMapping$checkMapData()}}
\item \href{#method-TimeProfileDataMapping-requireDualAxis}{\code{TimeProfileDataMapping$requireDualAxis()}}
\item \href{#method-TimeProfileDataMapping-getLeftAxis}{\code{TimeProfileDataMapping$getLeftAxis()}}
\item \href{#method-TimeProfileDataMapping-getRightAxis}{\code{TimeProfileDataMapping$getRightAxis()}}
\item \href{#method-TimeProfileDataMapping-clone}{\code{TimeProfileDataMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeProfileDataMapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeProfileDataMapping-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TimeProfileDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$new(
  x = NULL,
  y = NULL,
  ymin = NULL,
  ymax = NULL,
  group = NULL,
  y2Axis = NULL,
  color = NULL,
  fill = NULL,
  linetype = NULL,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{y}}{Name of y variable to map}

\item{\code{ymin}}{Name of ymin variable to map}

\item{\code{ymax}}{Name of ymax variable to map}

\item{\code{group}}{R6 class \code{Grouping} object or its input}

\item{\code{y2Axis}}{Name of y2Axis variable to map}

\item{\code{color}}{R6 class \code{Grouping} object or its input}

\item{\code{fill}}{R6 class \code{Grouping} object or its input}

\item{\code{linetype}}{R6 class \code{Grouping} object or its input}

\item{\code{data}}{data.frame to map used by \code{.smartMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{RangeDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeProfileDataMapping-checkMapData"></a>}}
\if{latex}{\out{\hypertarget{method-TimeProfileDataMapping-checkMapData}{}}}
\subsection{Method \code{checkMapData()}}{
Check that \code{data} variables include map variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$checkMapData(data, metaData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}

\item{\code{metaData}}{list containing information on \code{data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with map and \code{legendLabels} variables.
Dummy variable \code{legendLabels} is necessary to allow further modification of plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeProfileDataMapping-requireDualAxis"></a>}}
\if{latex}{\out{\hypertarget{method-TimeProfileDataMapping-requireDualAxis}{}}}
\subsection{Method \code{requireDualAxis()}}{
Assess if \code{data} require a dual axis plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$requireDualAxis(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeProfileDataMapping-getLeftAxis"></a>}}
\if{latex}{\out{\hypertarget{method-TimeProfileDataMapping-getLeftAxis}{}}}
\subsection{Method \code{getLeftAxis()}}{
Render NA values for all right axis data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$getLeftAxis(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data.frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame to be plotted in left axis
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeProfileDataMapping-getRightAxis"></a>}}
\if{latex}{\out{\hypertarget{method-TimeProfileDataMapping-getRightAxis}{}}}
\subsection{Method \code{getRightAxis()}}{
Render NA values for all left axis data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$getRightAxis(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data.frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame to be plotted in right axis
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeProfileDataMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeProfileDataMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfileDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
