% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-timeprofile.R
\docType{class}
\name{TimeProfilePlotConfiguration}
\alias{TimeProfilePlotConfiguration}
\title{TimeProfilePlotConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TimeProfilePlotConfiguration
}
\description{
Plot Configuration for PKRatio
}
\keyword{datasets}
