% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-timeprofile.R
\name{TimeProfilePlotConfiguration}
\alias{TimeProfilePlotConfiguration}
\title{TimeProfilePlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for time profile plots
}
\section{Super classes}{
\code{\link[tlf:LabelConfiguration]{tlf::LabelConfiguration}} -> \code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{TimeProfilePlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{timeProfileProperties}}{list of properties for time profile plot specific features}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TimeProfilePlotConfiguration$new()}}
\item \href{#method-addLLOQLines}{\code{TimeProfilePlotConfiguration$addLLOQLines()}}
\item \href{#method-addTimeProfiles}{\code{TimeProfilePlotConfiguration$addTimeProfiles()}}
\item \href{#method-clone}{\code{TimeProfilePlotConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="LabelConfiguration" data-id="setPlotLabels">}\href{../../tlf/html/LabelConfiguration.html#method-setPlotLabels}{\code{tlf::LabelConfiguration$setPlotLabels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="PlotConfiguration" data-id="savePlot">}\href{../../tlf/html/PlotConfiguration.html#method-savePlot}{\code{tlf::PlotConfiguration$savePlot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="PlotConfiguration" data-id="setPlotBackground">}\href{../../tlf/html/PlotConfiguration.html#method-setPlotBackground}{\code{tlf::PlotConfiguration$setPlotBackground()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tlf" data-topic="PlotConfiguration" data-id="setPlotProperties">}\href{../../tlf/html/PlotConfiguration.html#method-setPlotProperties}{\code{tlf::PlotConfiguration$setPlotProperties()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{TimeProfilePlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfilePlotConfiguration$new(
  timeProfileProperties = tlfEnv$currentTheme$timeProfile,
  title = "Time Profile Plot",
  subtitle = paste("Date:", format(Sys.Date(), "\%y-\%m-\%d")),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeProfileProperties}}{list of properties for PK ratio plot specific features}

\item{\code{title}}{R6 class \code{Label} object}

\item{\code{subtitle}}{R6 class \code{Label} object}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}

\item{\code{xlabel}}{R6 class \code{Label} object}

\item{\code{ylabel}}{R6 class \code{Label} object}

\item{\code{legend}}{R6 class \code{LegendConfiguration} object defining legend properties}

\item{\code{legendTitles}}{List of legend titles}

\item{\code{xAxis}}{R6 class \code{XAxisConfiguration} object defining X-axis properties}

\item{\code{xScale}}{character defining X-axis scale. Use enum \code{Scaling} to access predefined scales.}

\item{\code{xLimits}}{numeric vector of X-axis limits}

\item{\code{yAxis}}{R6 class \code{YAxisConfiguration} object defining X-axis properties}

\item{\code{yScale}}{character defining Y-axis scale. Use enum \code{Scaling} to access predefined scales.}

\item{\code{yLimits}}{numeric vector of Y-axis limits}

\item{\code{background}}{R6 class \code{BackgroundConfiguration} defining background properties}

\item{\code{watermark}}{R6 class \code{Label} object defining watermark background}

\item{\code{saveConfiguration}}{R6 class \code{SaveConfiguration} defining saving properties}

\item{\code{filename}}{character defining the name of the file to be saved}

\item{\code{width}}{numeric values defining the width in \code{units} of the plot dimensions after saving}

\item{\code{height}}{numeric values defining the height in \code{units} of the plot dimensions after saving}

\item{\code{units}}{character defining the unit of the saving dimension}

\item{\code{data}}{data.frame used by \code{smartMapping}}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{dataMapping}}{R6 class or subclass \code{XYGDataMapping}}

\item{\code{theme}}{R6 class \code{Theme}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TimeProfilePlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addLLOQLines"></a>}}
\subsection{Method \code{addLLOQLines()}}{
Add LLOQ limits as line layer to a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfilePlotConfiguration$addLLOQLines(plotObject, metaData, dataMapping)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{\code{ggplot} object}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{dataMapping}}{R6 class \code{TimeProfileDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object with PK ratio limits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addTimeProfiles"></a>}}
\subsection{Method \code{addTimeProfiles()}}{
Add time profiles as a combination of scatter, line, ribbon and errorbar layers to a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfilePlotConfiguration$addTimeProfiles(
  plotObject,
  data,
  metaData,
  dataMapping
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{\code{ggplot} object}

\item{\code{data}}{data.frame}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{dataMapping}}{R6 class \code{TimeProfileDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfilePlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
