% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeprofile-plotconfiguration.R
\name{TimeProfilePlotConfiguration}
\alias{TimeProfilePlotConfiguration}
\title{TimeProfilePlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for time profile plots
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{AxisConfiguration}},
\code{\link{BackgroundConfiguration}},
\code{\link{BackgroundElement}},
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{CumulativeTimeProfilePlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{ExportConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{LabelConfiguration}},
\code{\link{LegendConfiguration}},
\code{\link{LineElement}},
\code{\link{ObsVsPredPlotConfiguration}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PieChartPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{PlotGridConfiguration}},
\code{\link{QQPlotConfiguration}},
\code{\link{ResVsPredPlotConfiguration}},
\code{\link{ResVsTimePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}},
\code{\link{XAxisConfiguration}},
\code{\link{YAxisConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{tlf::PlotConfiguration} -> \code{TimeProfilePlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lloqDirection}}{Whether to draw LLOQ lines for x (vertical), y (horizontal) or x and y (both).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{y2Axis}}{\code{YAxisConfiguration} object defining properties of y2-axis}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeProfilePlotConfiguration-new}{\code{TimeProfilePlotConfiguration$new()}}
\item \href{#method-TimeProfilePlotConfiguration-clone}{\code{TimeProfilePlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeProfilePlotConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeProfilePlotConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TimeProfilePlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfilePlotConfiguration$new(
  ...,
  y2label = NULL,
  y2Axis = NULL,
  y2Scale = NULL,
  y2ValuesLimits = NULL,
  y2AxisLimits = NULL,
  y2Limits = lifecycle::deprecated(),
  lloqDirection = "horizontal",
  data = NULL,
  metaData = NULL,
  dataMapping = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}

\item{\code{y2label}}{character or \code{Label} object defining plot y2label}

\item{\code{y2Axis}}{\code{YAxisConfiguration} object defining y-axis properties}

\item{\code{y2Scale}}{name of y2-axis scale. Use enum \code{Scaling} to access predefined scales.}

\item{\code{y2ValuesLimits}}{numeric vector of length 2 defining y values limits}

\item{\code{y2AxisLimits}}{numeric vector of length 2 defining y axis limits}

\item{\code{y2Limits}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. Replaced by y2AxisLimits argument.}

\item{\code{lloqDirection}}{Whether to draw LLOQ lines for x (vertical), y (horizontal) or x and y (both).}

\item{\code{data}}{data.frame used by \code{.smartMapping}}

\item{\code{metaData}}{list of information on \code{data}}

\item{\code{dataMapping}}{R6 class or subclass \code{TimeProfileDataMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TimeProfilePlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeProfilePlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeProfilePlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfilePlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
