% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-timeprofile.R
\docType{class}
\name{TimeProfilePlotConfiguration}
\alias{TimeProfilePlotConfiguration}
\title{TimeProfilePlotConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TimeProfilePlotConfiguration
}
\description{
Class for TimeProfile Plot Configuration
}
\section{Fields}{

\describe{
\item{\code{legend}}{R6 class defining legendConfiguration}

\item{\code{xAxis}}{R6 class defining xAxisConfiguration}

\item{\code{yAxis}}{R6 class defining yAxisConfiguration}

\item{\code{background}}{R6 class defining backgroundConfiguration}

\item{\code{theme}}{R6 class defining theme aesthtic properties}

\item{\code{filename}}{Name of the saved plot}

\item{\code{timeProfileProperties}}{Properties of time profile plot specific features}
}}

\section{Methods}{

\describe{
\item{new(...)}{Initialize TimeProfilePlotConfiguration}
\item{setPlotProperties(plotObject)}{Apply properties of plot labels.}
\item{setPlotBackground(plotObject)}{Apply background properties to plot.}
\item{savePlot(plotObject)}{Save ggplot as file.}
\item{addLLOQLines(plotObject, metaData, dataMapping)}{Add LLOQ horizontal lines to plot.}
\item{addTimeProfiles(plotObject, data, metaData, dataMapping)}{Add time profiles to plot.}
}
}

\keyword{datasets}
