% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeprofile-plotconfiguration.R
\name{TimeProfilePlotConfiguration}
\alias{TimeProfilePlotConfiguration}
\title{TimeProfilePlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for time profile plots
}
\section{Super class}{
\code{\link[tlf:PlotConfiguration]{tlf::PlotConfiguration}} -> \code{TimeProfilePlotConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TimeProfilePlotConfiguration$new()}}
\item \href{#method-clone}{\code{TimeProfilePlotConfiguration$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TimeProfilePlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfilePlotConfiguration$new(
  lines = NULL,
  ribbons = NULL,
  points = NULL,
  errorbars = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lines}}{\code{ThemeAestheticSelections} defining properties of lines}

\item{\code{ribbons}}{\code{ThemeAestheticSelections} defining properties of ribbons}

\item{\code{points}}{\code{ThemeAestheticSelections} defining properties of points}

\item{\code{errorbars}}{\code{ThemeAestheticSelections} defining properties of error bars}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TimeProfilePlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeProfilePlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
