% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-input.R
\docType{data}
\name{tlfStatFunctions}
\alias{tlfStatFunctions}
\title{tlfStatFunctions}
\format{An object of class \code{list} of length 27.}
\usage{
tlfStatFunctions
}
\description{
Bank of predifined functions ready to use by Aggregation methods. Bank defined as Enum.
To access the function from its name, use match.fun: e.g. testFun <- match.fun("mean-1.96sd")
}
\keyword{datasets}
