% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enums.R
\docType{data}
\name{tlfStatFunctions}
\alias{tlfStatFunctions}
\title{tlfStatFunctions}
\format{
An object of class \code{list} of length 31.
}
\usage{
tlfStatFunctions
}
\description{
Bank of predefined functions ready to use by Aggregation methods. Bank defined as Enum.
To access the function from its name, use match.fun: e.g. testFun <- match.fun("mean-1.96sd")
}
\seealso{
Other enum helpers: 
\code{\link{AestheticFields}},
\code{\link{AestheticProperties}},
\code{\link{AestheticSelectionKeys}},
\code{\link{Alignments}},
\code{\link{AtomPlots}},
\code{\link{ColorMaps}},
\code{\link{ColorPalettes}},
\code{\link{DataMappings}},
\code{\link{DefaultDataMappingValues}},
\code{\link{Directions}},
\code{\link{ExportFormats}},
\code{\link{ExportUnits}},
\code{\link{FontFaces}},
\code{\link{HorizontalJustification}},
\code{\link{LabelOrientations}},
\code{\link{LegendPositions}},
\code{\link{LegendTypes}},
\code{\link{Linetypes}},
\code{\link{MoleculePlots}},
\code{\link{PlotAnnotationTextSize}},
\code{\link{PlotConfigurations}},
\code{\link{Scaling}},
\code{\link{Shapes}},
\code{\link{TagPositions}},
\code{\link{TickLabelTransforms}},
\code{\link{VerticalJustification}}
}
\concept{enum helpers}
\keyword{datasets}
