% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado-plotconfiguration.R
\name{TornadoPlotConfiguration}
\alias{TornadoPlotConfiguration}
\title{TornadoPlotConfiguration}
\description{
R6 class defining the configuration of a \code{ggplot} object for tornado plots
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{AxisConfiguration}},
\code{\link{BackgroundConfiguration}},
\code{\link{BackgroundElement}},
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{CumulativeTimeProfilePlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{ExportConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{LabelConfiguration}},
\code{\link{LegendConfiguration}},
\code{\link{LineElement}},
\code{\link{ObsVsPredPlotConfiguration}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{PlotGridConfiguration}},
\code{\link{QQPlotConfiguration}},
\code{\link{ResVsPredPlotConfiguration}},
\code{\link{ResVsTimePlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{XAxisConfiguration}},
\code{\link{YAxisConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{tlf::PlotConfiguration} -> \code{TornadoPlotConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bar}}{logical setting if tornado is uses a bar plot instead of regular points}

\item{\code{colorPalette}}{color palette property from \code{ggplot2}}

\item{\code{dodge}}{space between the bars/points}

\item{\code{defaultYScale}}{Default yAxis scale value when creating a \code{TornadoPlotConfiguration} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TornadoPlotConfiguration-new}{\code{TornadoPlotConfiguration$new()}}
\item \href{#method-TornadoPlotConfiguration-clone}{\code{TornadoPlotConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TornadoPlotConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-TornadoPlotConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TornadoPlotConfiguration} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TornadoPlotConfiguration$new(bar = TRUE, colorPalette = NULL, dodge = 0.5, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bar}}{logical setting if tornado is uses a bar plot instead of regular points}

\item{\code{colorPalette}}{color palette property from \code{ggplot2}}

\item{\code{dodge}}{space between the bars/points}

\item{\code{...}}{parameters inherited from \code{PlotConfiguration}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TornadoPlotConfiguration} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TornadoPlotConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TornadoPlotConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TornadoPlotConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
