% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\name{XAxisConfiguration}
\alias{XAxisConfiguration}
\title{XAxisConfiguration}
\description{
R6 class defining the configuration of X-axis
}
\section{Super class}{
\code{tlf::AxisConfiguration} -> \code{XAxisConfiguration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XAxisConfiguration-updatePlot}{\code{XAxisConfiguration$updatePlot()}}
\item \href{#method-XAxisConfiguration-clone}{\code{XAxisConfiguration$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="ggplotExpansion"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-ggplotExpansion'><code>tlf::AxisConfiguration$ggplotExpansion()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="ggplotScale"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-ggplotScale'><code>tlf::AxisConfiguration$ggplotScale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="initialize"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-initialize'><code>tlf::AxisConfiguration$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="prettyMinorTicks"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-prettyMinorTicks'><code>tlf::AxisConfiguration$prettyMinorTicks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="prettyTickLabels"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-prettyTickLabels'><code>tlf::AxisConfiguration$prettyTickLabels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="prettyTicks"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-prettyTicks'><code>tlf::AxisConfiguration$prettyTicks()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XAxisConfiguration-updatePlot"></a>}}
\if{latex}{\out{\hypertarget{method-XAxisConfiguration-updatePlot}{}}}
\subsection{Method \code{updatePlot()}}{
Update axis configuration on a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XAxisConfiguration$updatePlot(plotObject, ylim = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{\code{ggplot} object}

\item{\code{ylim}}{limits of \code{y} axis to prevent \code{coord_cartesian} to overwrite its properties}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object with updated axis properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XAxisConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XAxisConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XAxisConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
