% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-xy.R
\docType{class}
\name{XYDataMapping}
\alias{XYDataMapping}
\title{XYDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
XYDataMapping
}
\description{
Abstract class for X Y Mapping
}
\section{Fields}{

\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{y}}{Name of y variable to map}

\item{\code{data}}{Dataset after mapping}
}}

\section{Methods}{

\describe{
\item{new(x, y=NULL)}{Initialize XYDataMapping.}
\item{checkMapData(data, metaData = NULL)}{Check data mapping is correct. Create output data.frame with map data only.}
}
}

\keyword{datasets}
