% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-xy.R
\name{XYDataMapping}
\alias{XYDataMapping}
\title{XYDataMapping}
\description{
R6 class for mapping \code{x} and \code{y} variable to \code{data}
}
\seealso{
Other DataMapping classes: 
\code{\link{BoxWhiskerDataMapping}},
\code{\link{CumulativeTimeProfileDataMapping}},
\code{\link{DDIRatioDataMapping}},
\code{\link{GroupMapping}},
\code{\link{Grouping}},
\code{\link{HistogramDataMapping}},
\code{\link{ObsVsPredDataMapping}},
\code{\link{ObservedDataMapping}},
\code{\link{PKRatioDataMapping}},
\code{\link{QQDataMapping}},
\code{\link{RangeDataMapping}},
\code{\link{ResVsPredDataMapping}},
\code{\link{ResVsTimeDataMapping}},
\code{\link{TimeProfileDataMapping}},
\code{\link{TornadoDataMapping}},
\code{\link{XYGDataMapping}}
}
\concept{DataMapping classes}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{y}}{Name of y variable to map}

\item{\code{data}}{data.frame used for mapping}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XYDataMapping-new}{\code{XYDataMapping$new()}}
\item \href{#method-XYDataMapping-checkMapData}{\code{XYDataMapping$checkMapData()}}
\item \href{#method-XYDataMapping-clone}{\code{XYDataMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XYDataMapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-XYDataMapping-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{XYDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XYDataMapping$new(x, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{y}}{Name of y variable to map}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{XYDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XYDataMapping-checkMapData"></a>}}
\if{latex}{\out{\hypertarget{method-XYDataMapping-checkMapData}{}}}
\subsection{Method \code{checkMapData()}}{
Check that \code{data} variables include map variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XYDataMapping$checkMapData(data, metaData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}

\item{\code{metaData}}{list containing information on \code{data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with map and \code{defaultAes} variables.
Dummy variable \code{defaultAes} is necessary to allow further modification of plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XYDataMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XYDataMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XYDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
