% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-xy.R
\name{XYDataMapping}
\alias{XYDataMapping}
\title{XYDataMapping}
\description{
R6 class for mapping \code{x} and \code{y} variable to \code{data}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{y}}{Name of y variable to map}

\item{\code{data}}{data.frame used for mapping}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{XYDataMapping$new()}}
\item \href{#method-checkMapData}{\code{XYDataMapping$checkMapData()}}
\item \href{#method-clone}{\code{XYDataMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{XYDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XYDataMapping$new(x, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{y}}{Name of y variable to map}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{XYDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkMapData"></a>}}
\subsection{Method \code{checkMapData()}}{
Check that \code{data} variables include map variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XYDataMapping$checkMapData(data, metaData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}

\item{\code{metaData}}{list containing information on \code{data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with map and \code{defaultAes} variables.
Dummy variable \code{defaultAes} is necessary to allow further modification of plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XYDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
