% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-xygroup.R
\docType{class}
\name{XYGDataMapping}
\alias{XYGDataMapping}
\title{XYGDataMapping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
XYGDataMapping
}
\description{
Abstract class for X Y Group Mapping
}
\keyword{datasets}
