% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamapping-xygroup.R
\name{XYGDataMapping}
\alias{XYGDataMapping}
\title{XYGDataMapping}
\description{
R6 class for mapping \code{x}, \code{y} and \code{GroupMapping} variables to \code{data}
}
\seealso{
Other DataMapping classes: 
\code{\link{BoxWhiskerDataMapping}},
\code{\link{CumulativeTimeProfileDataMapping}},
\code{\link{DDIRatioDataMapping}},
\code{\link{GroupMapping}},
\code{\link{Grouping}},
\code{\link{HistogramDataMapping}},
\code{\link{ObsVsPredDataMapping}},
\code{\link{ObservedDataMapping}},
\code{\link{PKRatioDataMapping}},
\code{\link{QQDataMapping}},
\code{\link{RangeDataMapping}},
\code{\link{ResVsPredDataMapping}},
\code{\link{ResVsTimeDataMapping}},
\code{\link{TimeProfileDataMapping}},
\code{\link{TornadoDataMapping}},
\code{\link{XYDataMapping}}
}
\concept{DataMapping classes}
\section{Super class}{
\code{\link[tlf:XYDataMapping]{tlf::XYDataMapping}} -> \code{XYGDataMapping}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{groupMapping}}{R6 class \code{GroupMapping} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XYGDataMapping-new}{\code{XYGDataMapping$new()}}
\item \href{#method-XYGDataMapping-checkMapData}{\code{XYGDataMapping$checkMapData()}}
\item \href{#method-XYGDataMapping-clone}{\code{XYGDataMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XYGDataMapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-XYGDataMapping-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{XYGDataMapping} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XYGDataMapping$new(
  x = NULL,
  y = NULL,
  groupMapping = NULL,
  color = NULL,
  fill = NULL,
  linetype = NULL,
  shape = NULL,
  size = NULL,
  group = NULL,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Name of x variable to map}

\item{\code{y}}{Name of y variable to map}

\item{\code{groupMapping}}{R6 class \code{GroupMapping} object}

\item{\code{color}}{R6 class \code{Grouping} object or its input}

\item{\code{fill}}{R6 class \code{Grouping} object or its input}

\item{\code{linetype}}{R6 class \code{Grouping} object or its input}

\item{\code{shape}}{R6 class \code{Grouping} object or its input}

\item{\code{size}}{R6 class \code{Grouping} object or its input}

\item{\code{group}}{R6 class \code{Grouping} object or its input}

\item{\code{data}}{data.frame to map used by \code{.smartMapping}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{XYGDataMapping} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XYGDataMapping-checkMapData"></a>}}
\if{latex}{\out{\hypertarget{method-XYGDataMapping-checkMapData}{}}}
\subsection{Method \code{checkMapData()}}{
Check that \code{data} variables include map variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XYGDataMapping$checkMapData(data, metaData = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame to check}

\item{\code{metaData}}{list containing information on \code{data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with map and \code{defaultAes} variables.
Dummy variable \code{defaultAes} is necessary to allow further modification of plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XYGDataMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XYGDataMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XYGDataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
