% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration.R
\docType{class}
\name{YAxisConfiguration}
\alias{YAxisConfiguration}
\title{YAxisConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
YAxisConfiguration
}
\description{
Generic Y axis Configuration
}
\keyword{datasets}
