% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\name{YAxisConfiguration}
\alias{YAxisConfiguration}
\title{YAxisConfiguration}
\description{
R6 class defining the configuration of Y-axis
}
\seealso{
Other PlotConfiguration classes: 
\code{\link{AxisConfiguration}},
\code{\link{BackgroundConfiguration}},
\code{\link{BackgroundElement}},
\code{\link{BoxWhiskerPlotConfiguration}},
\code{\link{CumulativeTimeProfilePlotConfiguration}},
\code{\link{DDIRatioPlotConfiguration}},
\code{\link{ExportConfiguration}},
\code{\link{HistogramPlotConfiguration}},
\code{\link{LabelConfiguration}},
\code{\link{LegendConfiguration}},
\code{\link{LineElement}},
\code{\link{ObsVsPredPlotConfiguration}},
\code{\link{PKRatioPlotConfiguration}},
\code{\link{PlotConfiguration}},
\code{\link{PlotGridConfiguration}},
\code{\link{QQPlotConfiguration}},
\code{\link{ResVsPredPlotConfiguration}},
\code{\link{ResVsTimePlotConfiguration}},
\code{\link{TimeProfilePlotConfiguration}},
\code{\link{TornadoPlotConfiguration}},
\code{\link{XAxisConfiguration}}
}
\concept{PlotConfiguration classes}
\section{Super class}{
\code{\link[tlf:AxisConfiguration]{tlf::AxisConfiguration}} -> \code{YAxisConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{position}}{character position of the Y-axis}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-YAxisConfiguration-updatePlot}{\code{YAxisConfiguration$updatePlot()}}
\item \href{#method-YAxisConfiguration-clone}{\code{YAxisConfiguration$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="ggplotExpansion"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-ggplotExpansion'><code>tlf::AxisConfiguration$ggplotExpansion()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="ggplotScale"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-ggplotScale'><code>tlf::AxisConfiguration$ggplotScale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="initialize"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-initialize'><code>tlf::AxisConfiguration$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="prettyMinorTicks"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-prettyMinorTicks'><code>tlf::AxisConfiguration$prettyMinorTicks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="prettyTickLabels"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-prettyTickLabels'><code>tlf::AxisConfiguration$prettyTickLabels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tlf" data-topic="AxisConfiguration" data-id="prettyTicks"><a href='../../tlf/html/AxisConfiguration.html#method-AxisConfiguration-prettyTicks'><code>tlf::AxisConfiguration$prettyTicks()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-YAxisConfiguration-updatePlot"></a>}}
\if{latex}{\out{\hypertarget{method-YAxisConfiguration-updatePlot}{}}}
\subsection{Method \code{updatePlot()}}{
Update axis configuration on a \code{ggplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YAxisConfiguration$updatePlot(plotObject, xlim = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{\code{ggplot} object}

\item{\code{xlim}}{limits of \code{x} axis to prevent \code{coord_cartesian} to overwrite its properties}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ggplot} object with updated axis properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-YAxisConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-YAxisConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YAxisConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
