% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotconfiguration-axis.R
\docType{class}
\name{YAxisConfiguration}
\alias{YAxisConfiguration}
\title{YAxisConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
YAxisConfiguration
}
\description{
Class for Y-axis Configuration
}
\section{Fields}{

\describe{
\item{\code{limits}}{Axis limits}

\item{\code{scale}}{Plot scale (lin, log...)}

\item{\code{ticks}}{Values where ticks should be applied}

\item{\code{ticklabels}}{Names of associated to ticks}
}}

\section{Methods}{

\describe{
\item{new(limits = NULL, scale = Scaling$lin, ticks = "default", ticklabels = "default")}{Initialize AxisConfiguration.}
\item{setPlotAxis(plotObject)}{Apply properties to plot X-axis.}
}
}

\keyword{datasets}
